% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_slide_opt_edf.R
\name{across_ish_names_info}
\alias{across_ish_names_info}
\title{Calculate input and output column names for an \code{{epiprocess}} \code{\link[dplyr:across]{dplyr::across}}-like operations}
\usage{
across_ish_names_info(
  .x,
  time_type,
  col_names_quo,
  .f_namer,
  .window_size,
  .align,
  .prefix,
  .suffix,
  .new_col_names
)
}
\arguments{
\item{.x}{data.frame to perform input column tidyselection on}

\item{time_type}{as in \code{\link{new_epi_df}}}

\item{col_names_quo}{enquosed input column tidyselect expression}

\item{.f_namer}{function taking an input column object and outputting a name
for a corresponding output column; see \code{\link{upstream_slide_f_info}}}

\item{.window_size}{as in \code{\link{epi_slide_opt}}}

\item{.align}{as in \code{\link{epi_slide_opt}}}

\item{.prefix}{as in \code{\link{epi_slide_opt}}}

\item{.suffix}{as in \code{\link{epi_slide_opt}}}

\item{.new_col_names}{as in \code{\link{epi_slide_opt}}}
}
\value{
named list with two elements: \code{input_col_names}, chr, subset of
\code{names(.x)}; and \code{output_colnames}, chr, same length as \code{input_col_names}
}
\description{
Calculate input and output column names for an \code{{epiprocess}} \code{\link[dplyr:across]{dplyr::across}}-like operations
}
\keyword{internal}
