% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_patterns.R
\name{add_shift}
\alias{add_shift}
\title{Used for data shifting in terms of reference date}
\usage{
add_shift(df, n_day, refd_col)
}
\arguments{
\item{df}{Data Frame of aggregated counts within a single location
reported for each reference date and issue date.}

\item{n_day}{number of days to be shifted}

\item{refd_col}{column name for the column of reference date}
}
\description{
Used for data shifting in terms of reference date
}
