% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_df.R
\name{aggregate_epi_df}
\alias{aggregate_epi_df}
\title{Aggregate an \code{epi_df} object}
\usage{
aggregate_epi_df(.x, value_col = "value", group_cols = "time_value")
}
\arguments{
\item{.x}{an \code{epi_df}}

\item{value_col}{character name of the column to aggregate}

\item{group_cols}{character vector of column names to group by}
}
\value{
an \code{epi_df} object
}
\description{
Aggregates an \code{epi_df} object by the specified group columns, summing the
\code{value} column, and returning an \code{epi_df}. If aggregating over \code{geo_value},
the resulting \code{epi_df} will have \code{geo_value} set to \code{"total"}.
}
