% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{archive_cases_dv}
\alias{archive_cases_dv}
\title{Daily doctor visits and cases from California and Florida in archive format}
\format{
An \code{epi_archive} data format. The data table DT has 160 rows and 5 columns:
\describe{
\item{geo_value}{the geographic value associated with each row of measurements.}
\item{time_value}{the time value associated with each row of measurements.}
\item{version}{ the time value specifying the version for each row of measurements. }
\item{percent_cli}{percentage of doctor’s visits with CLI (COVID-like illness) computed from medical insurance claims}
\item{case_rate}{7-day average signal of number of new confirmed deaths due to COVID-19 per 100,000 population, daily}
}
}
\source{
These data sources are provided under the terms of the
\href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution license:}
\itemize{
\item \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/doctor-visits.html}{Doctor Visits}
}

COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University.
\url{https://github.com/CSSEGISandData/COVID-19}
}
\usage{
archive_cases_dv
}
\description{
This data source is based on information about outpatient visits,
provided to us by health system partners, and also contains confirmed
COVID-19 cases based on reports made available by the Center for
Systems Science and Engineering at Johns Hopkins University.
This example data ranges from June 1 to June 15, 2020.
}
\keyword{datasets}
