% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{data}
\name{archive_cases_dv_subset_dt}
\alias{archive_cases_dv_subset_dt}
\title{Subset of daily doctor visits and cases in archive format}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 129638 rows and 5 columns.
}
\source{
This object contains a modified part of the \href{https://github.com/CSSEGISandData/COVID-19}{COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University} as \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{republished in the COVIDcast Epidata API}. This data set is licensed under the terms of the
\href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0 International license}
by Johns Hopkins University on behalf of its Center for Systems Science in Engineering.
Copyright Johns Hopkins University 2020.

Modifications:
\itemize{
\item \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/doctor-visits.html}{From the COVIDcast Doctor Visits API}: The signal \code{percent_cli} is taken directly from the API without changes.
\item \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{From the COVIDcast Epidata API}: \code{case_rate_7d_av} signal was computed by Delphi from the original JHU-CSSE data by calculating moving averages of the preceding 7 days, so the signal for June 7 is the average of the underlying data for June 1 through 7, inclusive.
\item Furthermore, the data is a subset of the full dataset, the signal names slightly altered, and formatted into a tibble.
}
}
\usage{
archive_cases_dv_subset_dt
}
\description{
This data source is based on information about outpatient visits,
provided to us by health system partners, and also contains confirmed
COVID-19 cases based on reports made available by the Center for
Systems Science and Engineering at Johns Hopkins University.
This example data ranges from June 1, 2020 to Dec 1, 2021, and
is also limited to California, Florida, Texas, and New York.
}
\examples{
 # Since this is a re-exported dataset, it cannot be loaded using
 # the `data()` function. `data()` looks for a file of the same name
 # in the `data/` directory, which doesn't exist in this package.
 # works
 epiprocess::archive_cases_dv_subset_dt

 # works
 library(epiprocess)
 archive_cases_dv_subset_dt

 # fails
 data(archive_cases_dv_subset_dt, package = "epiprocess")
}
\keyword{datasets}
