% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{arrange.epi_df}
\alias{arrange.epi_df}
\alias{filter.epi_df}
\alias{group_by.epi_df}
\alias{group_modify.epi_df}
\alias{mutate.epi_df}
\alias{relocate.epi_df}
\alias{rename.epi_df}
\alias{slice.epi_df}
\alias{ungroup.epi_df}
\alias{unnest.epi_df}
\title{dplyr verbs}
\usage{
\method{arrange}{epi_df}(x, ...)

\method{filter}{epi_df}(x, ...)

\method{group_by}{epi_df}(x, ...)

\method{group_modify}{epi_df}(x, ...)

\method{mutate}{epi_df}(x, ...)

\method{relocate}{epi_df}(x, ...)

\method{rename}{epi_df}(x, ...)

\method{slice}{epi_df}(x, ...)

\method{ungroup}{epi_df}(x, ...)

\method{unnest}{epi_df}(x, ...)
}
\description{
dplyr verbs for \code{epi_df} objexts, preserving class and attributes.
}
