% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_signal.R
\name{as.epi_signal}
\alias{as.epi_signal}
\alias{as.epi_signal.epi_signal}
\alias{as.epi_signal.tibble}
\alias{as.epi_signal.data.frame}
\title{Convert data frame to \code{epi_signal} format}
\usage{
as.epi_signal(x, ...)

\method{as.epi_signal}{epi_signal}(x, ...)

\method{as.epi_signal}{tibble}(
  x,
  name,
  geo_type,
  time_type,
  signal_type,
  signal_unit,
  issue,
  metadata = list(),
  ...
)

\method{as.epi_signal}{data.frame}(
  x,
  name,
  geo_type,
  time_type,
  signal_type,
  signal_unit,
  issue,
  metadata = list(),
  ...
)
}
\arguments{
\item{x}{Object to be converted. See Methods section below for details on
formatting of each input type.}

\item{...}{Additional arguments passed to methods.}

\item{name}{The name to use for this signal.}

\item{geo_type}{The geographic resolution. If missing, then it will be
guessed from the geo values present.}

\item{time_type}{The temporal resolution. If missing, then it will be guessed
from the time values present.}

\item{signal_type}{The type of the signal value.}

\item{signal_unit}{The units of the signal value.}

\item{issue}{Issue date to use for this data, if not present in \code{x}. If no
issue date is present in \code{x} and \code{issue} is missing, then today's date will
be used.}

\item{metadata}{List or tibble of additional metadata to attach to the
\code{epi_signal} object. All objects will have \code{geo_type}, \code{time_type},
\code{signal_type} (optional), and \code{signal_unit} (optional) entries included in
their metadata, derived from the above arguments; any entries in the passed
argument will be preserved in the metadata as well.}
}
\value{
An \code{epi_signal} object.
}
\description{
Converts a data frame or tibble into a format consistent with the
\code{epi_signal} class (ensuring that it has a certain minimal set of columns,
and that it has certain metadata stored in its attributes).
}
\details{
An \code{epi_signal} object is simply a tibble, with (at least) the
following columns (with data types written in tibble notation):
\itemize{
\item \code{value} <dbl>: the value of the signal
\item \code{geo_value} <int> or <str>: the associated geographic value
\item \code{time_value} <date>: the associated time value
\item \code{issue} <date>: the time value at which the given signal value was issued
}

An \code{epi_signal} object also has a tibble \code{metadata} stored in its attributes,
with (at least) the following columns:
\itemize{
\item \code{name} <str>: the name of the signal
\item \code{geo_type} <str>: the geographic resolution
\item \code{time_type} <str>: the temporal resolution
\item \code{signal_type} <str>: the type of the signal value (optional)
\item \code{signal_unit} <str>: the units associated with the signal value (optional)
}

More information on geo types, time types, and signal types, is given
below.
}
\section{Methods (by class)}{
\itemize{
\item \code{epi_signal}: Simply returns the \code{epi_signal} object unchanged.

\item \code{tibble}: The input tibble \code{x} must contain the columns
\code{value}, \code{geo_value}, and \code{time_value}. If an \code{issue} column is present in
\code{x}, it will be used as the issue date for each observation; if not, the
\code{issue} argument will be used. Other columns will be preserved as-is.

\item \code{data.frame}: The input data frame \code{x} must contain the columns
\code{value}, \code{geo_value}, and \code{time_value}. If an \code{issue} column is present in
\code{x}, it will be used as the issue date for each observation; if not, the
\code{issue} argument will be used. Other columns will be preserved as-is.
}}

\section{Geo types}{

The allowable geo types, and their coding (allowable range for geo values),
are as follows.
\itemize{
\item \code{county}: U.S. counties; coded by 5-digit FIPS codes.
\item \code{hrr}: U.S. hospital referral regions (designed to represent regional
health care markets); there are 306 HRRs in the U.S; coded by number
(nonconsecutive, between 1 and 457).
\item \code{state}: U.S. states; coded by 2-digit postal abbreviation (lowercase);
note that Puerto Rico is "pr" and Washington D.C. is "dc".
\item \code{hhs}: U.S. HHS regions; coded by number (between 1 and 10).
\item \code{nation}: country; coded by ISO 3166-1 alpha-2 country codes (lowercase).
}
}

\section{Time types}{

The allowable time types are as follows. In each case, their coding
(allowable range for time values) is an object of class \code{Date}, or string
in the format "YYYY-MM-DD".
\itemize{
\item "day": each observation covers one day.
\item "week": each observation covers one week, and the time value is assumed to
be the start date of the epiweek (MMWR week) that the data represents.
}
}

\section{Signal types}{

todo
}

