% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tibble.R
\name{as.epi_tibble}
\alias{as.epi_tibble}
\alias{as.epi_tibble.epi_tibble}
\alias{as.epi_tibble.tibble}
\alias{as.epi_tibble.data.frame}
\title{Convert data to \code{epi_tibble} format}
\usage{
as.epi_tibble(x, ...)

\method{as.epi_tibble}{epi_tibble}(x, ...)

\method{as.epi_tibble}{tibble}(x, geo_type, time_type, issue, additional_metadata = list(), ...)

\method{as.epi_tibble}{data.frame}(x, geo_type, time_type, issue, additional_metadata = list(), ...)
}
\arguments{
\item{x}{The object to be converted. See the methods section below for
details on formatting of each input type.}

\item{...}{Additional arguments passed to methods.}

\item{geo_type}{The type for the geo values. If missing, then the function
will attempt to infer it from the geo values present; if this fails, then
it will be set to "custom".}

\item{time_type}{The type for the time values. If missing, then the function
will attempt to infer it from the time values present; if this fails, then
it will be set to "custom".}

\item{issue}{Issue to use for this data. If missing, then the function will
attempt to infer it from the passed object \code{x}; if this fails, then the
current day-time will be used.}

\item{additional_metadata}{List of additional metadata to attach to the
\code{epi_tibble} object. All objects will have \code{time_type}, \code{geo_type}, and
\code{issue} fields; named entries from the passed list or will be included as
well.}
}
\value{
An \code{epi_tibble} object.
}
\description{
Converts a data frame or tibble into a format that is consistent with the
\code{epi_tibble} class, ensuring that it has a certain minimal set of columns,
and that it has certain minimal metadata.
}
\details{
An \code{epi_tibble} is a tibble with (at least) the following columns:
\itemize{
\item \code{geo_value}: the geographic value associated with each measurement.
\item \code{time_value}: the time value associated with each measurement.
}

Other columns can be considered as measured variables, which we also broadly
refer to as signal variables. An \code{epi_tibble} object also has metadata with
(at least) the following fields:
\itemize{
\item \code{geo_type}: the type for the geo values.
\item \code{time_type}: the type for the time values.
\item \code{issue}: the time value at which the given data set was issued.
}

The first two fields above, \code{geo_type} and \code{time_type}, can usually be
inferred from the \code{geo_value} and \code{time_value} columns, respectively. The
last field above, \code{issue}, is the most unique to the \code{epi_tibble} format.
In a typical case, this represents the maximum of the issues of individual
signal values measured in the data set; hence we would also say that the
data set is comprised of all signal values observed "as of" the given issue
in the metadata.

Metadata for an \code{epi_tibble} object \code{x} can be accessed (and altered) via
\code{attributes(x)$metadata}. More information on geo types, time types, and
issues is given below.
}
\section{Methods (by class)}{
\itemize{
\item \code{epi_tibble}: Simply returns the \code{epi_tibble} object unchanged.

\item \code{tibble}: The input tibble \code{x} must contain the columns
\code{geo_value} and \code{time_value}. All other columns will be preserved as is,
and treated as measured variables. If \code{issue} is missing, then the function
will look for \code{issue} as a column of \code{x}, or as a field in its metadata
(stored in its attributes), to infer the issue; if this fails, then the
current day-time will be used.

\item \code{data.frame}: The input data frame \code{x} must contain the columns
\code{geo_value} and \code{time_value}. All other columns will be preserved as is,
and treated as measured variables. If \code{issue} is missing, then the function
will look for \code{issue} as a column of \code{x}, or as a field in its metadata
(stored in its attributes), to infer the issue; if this fails, then the
current day-time will be used.
}}

\section{Geo types}{

The following geo types are supported in an \code{epi_tibble}. Their geo coding
(specification of geo values for each geo type) is also described below.
\itemize{
\item \code{"county"}: each observation corresponds to a U.S. county; coded by 5-digit
FIPS code.
\item \code{"hrr"}: each observation corresponds to a U.S. hospital referral region
(designed to represent regional healthcare markets); there are 306 HRRs in
the U.S; coded by number (nonconsecutive, between 1 and 457).
\item \code{"state"}: each observation corresponds to a U.S. state; coded by 2-digit
postal abbreviation (lowercase);
note that Puerto Rico is "pr" and Washington D.C. is "dc".
\item \code{"hhs"}: each observation corresponds to a U.S. HHS region; coded by number
(consecutive, between 1 and 10).
\item \code{"nation"}: each observation corresponds to a country; coded by ISO 31661-
alpha-2 country codes (lowercase).
}

The above geo types come with aggregation utilities in the package, \emph{todo:
refer to relevant functionality, vignette, and so on}. An unrecognizable
geo type is labeled as "custom".
}

\section{Time types}{

The following time types are supported in an \code{epi_tibble}. Their time coding
(specification of time values for each time type) is also described below.
\itemize{
\item \code{"day-time"}: each observation corresponds to a time on a given day (measured
to the second); coded as a \code{POSIXct} object, as in \code{as.POSIXct("2020-06-09 18:45:40")}.
\item \code{"day"}: each observation corresponds to a day; coded as a \code{Date} object,
as in \code{as.Date("2020-06-09")}.
\item \code{"week"}: each observation corresponds to a week; the alignment can be
arbitrary (as to whether a week starts on a Monday, Tuesday, etc.; the
U.S. CDC definition of an epidemiological week starts on a Sunday); coded
as a \code{Date} object, representing the start date of week.
}

An unrecognisable time type is labeled as "custom".
}

\section{Issues}{

todo
}

