% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_df.R
\name{as_inheriting}
\alias{as_inheriting}
\title{Reinterpretation of \code{x} as inheriting \code{subclass} if it doesn't already}
\usage{
as_inheriting(x, subclass, potential_subsubclasses = character(0L))
}
\arguments{
\item{x}{an S3 object}

\item{subclass}{string; S3 class name we want to ensure that the result will
\code{\link[base:inherit]{base::inherit}}. We call this a "subclass" because we generally expect it
to be in the beginning or middle of a \code{class} vector, followed by some
other S3 class(es). If \code{x} already inherits this subclass, we return \code{x}
unchanged.}

\item{potential_subsubclasses}{Optional; character vector: if \code{class(x)}
doesn't already include \code{subclass}, it should be added in the first
position after all of the \code{potential_subsubclasses} that appear in
\code{class(x)}. This lets us control the order of \code{subclass} relative to
"base"-like classes such as \code{"tbl_df"} and other wrapper classes like
\code{"grouped_df"}. Will be ignored if \code{subclass} is already in \code{class(x)}.}
}
\value{
object that inherits \code{subclass}
}
\description{
Reinterpretation of \code{x} as inheriting \code{subclass} if it doesn't already
}
