% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukey-col-listbacked.R
\name{as_non_ukey_col_listbacked}
\alias{as_non_ukey_col_listbacked}
\title{If \code{x} \code{is_ukey_col_listbacked}, make it not be}
\usage{
as_non_ukey_col_listbacked(x)
}
\description{
If \code{x} \code{is_ukey_col_listbacked}, make it not be
}
