% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_new.R
\name{as_of.epi_archive2}
\alias{as_of.epi_archive2}
\title{As of epi_archive}
\usage{
\method{as_of}{epi_archive2}(epi_archive, max_version, min_time_value = -Inf, all_versions = FALSE)
}
\arguments{
\item{epi_archive}{An \code{epi_archive} object}

\item{max_version}{Version specifying the max version to permit in the
snapshot. That is, the snapshot will comprise the unique rows of the
current archive data that represent the most up-to-date signal values, as
of the specified \code{max_version} (and whose \code{time_value}s are at least
\code{min_time_value}).}

\item{min_time_value}{Time value specifying the min \code{time_value} to permit in
the snapshot. Default is \code{-Inf}, which effectively means that there is no
minimum considered.}

\item{all_versions}{Boolean; If \code{all_versions = TRUE}, then the output will be in
\code{epi_archive} format, and contain rows in the specified \code{time_value} range
having \code{version <= max_version}. The resulting object will cover a
potentially narrower \code{version} and \code{time_value} range than \code{x}, depending
on user-provided arguments. Otherwise, there will be one row in the output
for the \code{max_version} of each \code{time_value}. Default is \code{FALSE}.}
}
\description{
Generates a snapshot in \code{epi_df} format as of a given version.
See the documentation for the wrapper function \code{\link[=epix_as_of]{epix_as_of()}} for
details. The parameter descriptions below are copied from there
}
