% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_df.R
\name{as_tibble.epi_df}
\alias{as_tibble.epi_df}
\title{Convert to tibble}
\usage{
\method{as_tibble}{epi_df}(x, ...)
}
\arguments{
\item{x}{an \code{epi_df}}

\item{...}{arguments to forward to \code{NextMethod()}}
}
\description{
Converts an \code{epi_df} object into a tibble, dropping metadata and any
grouping.
}
