% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukey-col-listbacked.R
\name{as_ukey_col_listbacked}
\alias{as_ukey_col_listbacked}
\title{Make \code{x} satisfy \code{is_ukey_col_listbacked}, if it isn't already}
\usage{
as_ukey_col_listbacked(x, ...)
}
\description{
Make \code{x} satisfy \code{is_ukey_col_listbacked}, if it isn't already
}
