% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_sufficient_f_args}
\alias{assert_sufficient_f_args}
\title{Assert that a sliding computation function takes enough args}
\usage{
assert_sufficient_f_args(.f, ..., .ref_time_value_label)
}
\arguments{
\item{...}{Dots that will be forwarded to \code{f} from the dots of \code{epi_slide} or
\code{epix_slide}.}

\item{.ref_time_value_label}{String; how to describe/label the \code{ref_time_value} in
error messages; e.g., "reference time value" or "version".}

\item{f}{Function; specifies a computation to slide over an \code{epi_df} or
\code{epi_archive} in \code{epi_slide} or \code{epix_slide}.}
}
\description{
Assert that a sliding computation function takes enough args
}
\keyword{internal}
