% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting-helpers.R
\name{assess_available_predictor_shift}
\alias{assess_available_predictor_shift}
\title{Assess whether an available predictor looks suitable for training}
\usage{
assess_available_predictor_shift(
  candidate_shift_description,
  archive,
  training_tbl,
  testing_tbl,
  included_shift_descriptions,
  predictor_search_records,
  min_predictor_shift_spacing,
  min_n_training_versions,
  min_n_training_rows_to_n_predictor_shifts_ratio
)
}
\description{
Assess whether an available predictor looks suitable for training
}
\examples{

library(zeallot)

archive <- as_epi_archive(dplyr::bind_rows(
  tibble(
    geo_value = 1,
    time_value = 1:6,
    version = 2:7,
    a = 111:116,
    b = 211:216
  ),
  tibble(
    geo_value = 1,
    time_value = 1:5,
    version = 3:7,
    a = 121:125,
    b = 221:225
  )
))
shift_descriptions <- tibble::tribble(
  ~predictor, ~relative_time,
  "a", -1,
  "a", -2,
  "a", -3,
  "b", -2,
  "a", -5,
)
training_tbl <- epix_target_evaluation_data(archive, "a", -1, 0)
testing_tbl <- tibble(geo_value = 1, anchor_version = archive$versions_end)
included_shift_descriptions <- shift_descriptions[0,]
predictor_search_records <- tibble()
for (i in seq_len(nrow(shift_descriptions))) {
  c(training_tbl, testing_tbl, included_shift_descriptions, predictor_search_records, last_predictor_search_record) \%<-\%
    assess_available_predictor_shift(
      shift_descriptions[i,],
      archive,
      training_tbl,
      testing_tbl,
      included_shift_descriptions,
      predictor_search_records,
      min_predictor_shift_spacing = 2L,
      min_n_training_versions = 2L,
      min_n_training_rows_to_n_predictor_shifts_ratio = 0.5 # absurd value, so that example data can be small
    )
  print(last_predictor_search_record$included)
}
print(training_tbl)
print(testing_tbl)
print(included_shift_descriptions)
print(predictor_search_records, width = 200)

}
\keyword{internal}
