% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting-parts.R
\name{attach_semistable_target_training_segment}
\alias{attach_semistable_target_training_segment}
\title{attach_semistable_target_training_segment}
\usage{
attach_semistable_target_training_segment(
  time_until_semistable = as.difftime(60, units = "days"),
  anchor_versions = NULL,
  out_name = "{.col}_{.amt}{.dir}_evaluation",
  nomatch = NA,
  drop_na = TRUE
)
}
\description{
attach_semistable_target_training_segment
}
\examples{

attach_semistable_target_training_segment()(list(
  targets = tibble(varname = "percent_cli", version_relative_time = 0:1),
  features = tibble(),
  training = NULL,
  testing = NULL,
  inferences = NULL,
  archive = archive_cases_dv_subset
))$training
}
