% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.epi_df}
\alias{autoplot.epi_df}
\title{Automatically plot an epi_df}
\usage{
\method{autoplot}{epi_df}(
  object,
  ...,
  .color_by = c("all_keys", "geo_value", "other_keys", ".response", "all", "none"),
  .facet_by = c(".response", "other_keys", "all_keys", "geo_value", "all", "none"),
  .base_color = "#3A448F",
  .max_facets = Inf
)
}
\arguments{
\item{object}{An \code{epi_df}}

\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}

\item{.color_by}{Which variables should determine the color(s) used to plot
lines. Options include:
\itemize{
\item \code{all_keys} - the default uses the interaction of any key variables
including the \code{geo_value}
\item \code{geo_value} - \code{geo_value} only
\item \code{other_keys} - any available keys that are not \code{geo_value}
\item \code{.response} - the numeric variables (same as the y-axis)
\item \code{all} - uses the interaction of all keys and numeric variables
\item \code{none} - no coloring aesthetic is applied
}}

\item{.facet_by}{Similar to \code{.color_by} except that the default is to display
each numeric variable on a separate facet}

\item{.base_color}{Lines will be shown with this color. For example, with a
single numeric variable and faceting by \code{geo_value}, all locations would
share the same color line.}

\item{.max_facets}{Cut down of the number of facets displayed. Especially
useful for testing when there are many \code{geo_value}'s or keys.}
}
\description{
Automatically plot an epi_df
}
\examples{
autoplot(jhu_csse_daily_subset, cases, death_rate_7d_av)
autoplot(jhu_csse_daily_subset, case_rate_7d_av, .facet_by = "geo_value")
autoplot(jhu_csse_daily_subset, case_rate_7d_av, .color_by = "none", 
  .facet_by = "geo_value")
}
