% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_edf.R, R/methods-epi_df.R
\name{group_by.epi_df}
\alias{group_by.epi_df}
\alias{grouped_edf}
\alias{group_data.grouped_edf}
\alias{ungroup.grouped_edf}
\alias{basic_s3_for_epi_df}
\alias{tbl_sum.epi_df}
\alias{summary.epi_df}
\alias{unnest.epi_df}
\title{Basic S3 methods for an \code{epi_df} object}
\usage{
\method{group_by}{epi_df}(.data, ...)

\method{group_data}{grouped_edf}(.data)

\method{ungroup}{grouped_edf}(x, ...)

\method{tbl_sum}{epi_df}(x, ...)

\method{summary}{epi_df}(object, ...)

\method{unnest}{epi_df}(data, ...)
}
\arguments{
\item{...}{Additional arguments, for compatibility with \code{summary()}.
Currently unused.}

\item{object}{The \code{epi_df} object.}

\item{data}{The \code{epi_df} object.}
}
\description{
Prints a variety of summary statistics about the \code{epi_df} object, such as
the time range included and geographic coverage.
}
