% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca_daily_cases}
\alias{ca_daily_cases}
\title{California daily COVID-19 new cases.}
\format{
A tibble with 62 rows and 3 variables:
\describe{
\item{geo_value}{location such as the state name}
\item{time_value}{date}
\item{cases}{number of new confirmed COVID-19 cases, daily}
...
}
}
\source{
\url{https://cmu-delphi.github.io/delphi-epidata/}
}
\usage{
ca_daily_cases
}
\description{
A dataset containing the daily COVID-19 new cases
from March 1, 2020 to May 1, 2020.
}
\keyword{datasets}
