% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca_fl_death_rate_and_cases}
\alias{ca_fl_death_rate_and_cases}
\title{California and Florida Daily COVID-19 incidence rate and death Rates.}
\format{
A tibble with 124 rows and 4 variables:
\describe{
\item{geo_value}{location such as the state name}
\item{time_value}{date}
\item{case_rate}{number of new confirmed COVID-19 cases per 100,000 population, daily}
\item{death_rate}{number of new confirmed deaths due to COVID-19 per 100,000 population, daily}
...
}
}
\source{
\url{https://cmu-delphi.github.io/delphi-epidata/}
}
\usage{
ca_fl_death_rate_and_cases
}
\description{
A dataset containing the daily COVID-19 incidence rates and death rates
from March 1, 2020 to May 1, 2020.
}
\keyword{datasets}
