% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditions.R
\name{check_global_handlers_stack_support}
\alias{check_global_handlers_stack_support}
\title{Raise an error if  isn't supported}
\usage{
check_global_handlers_stack_support()
}
\description{
Raise an error if  isn't supported
}
\seealso{
Other functions working with global calling handler stack: 
\code{\link{global_handlers_list}()},
\code{\link{global_handlers_push}()},
\code{\link{global_handlers_remove}()}
}
\concept{functions working with global calling handler stack}
