% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditions.R
\name{check_named_handlers_dots}
\alias{check_named_handlers_dots}
\title{Raise an error if dots don't look like named condition handlers}
\usage{
check_named_handlers_dots(...)
}
\arguments{
\item{...}{named calling handlers: each a function that takes a condition
object as an argument and performs some handling (such as printing or
recording a stack trace), passed by name, where the name associates the
handler with a class of conditions it should be called on (e.g.,
\code{"condition"} as a catch-all, \code{"error"}, \code{"warning"}, \code{"message"}, or more
specific classes provided by some packages)}
}
\description{
Raise an error if dots don't look like named condition handlers
}
