% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_ukey_unique}
\alias{check_ukey_unique}
\title{Check that a unique key is indeed unique in a tibble (TRUE/str)}
\usage{
check_ukey_unique(x, ukey_names, end_cli_message = character())
}
\arguments{
\item{x}{a tibble, with no particular row or column order (if you have a
guaranteed row order based on the ukey you can probably do something more
efficient)}

\item{ukey_names}{character vector; subset of column names of \code{x} denoting a
unique key.}

\item{end_cli_message}{optional character vector, a cli message format
string/vector; information/advice to tack onto any error messages.}
}
\value{
\code{TRUE} if no ukey is duplicated (i.e., \code{x[ukey_names]} has no
duplicated rows); string with an error message if there are errors.
}
\description{
A \code{checkmate}-style check function.
}
\keyword{internal}
