% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{clone}
\alias{clone}
\alias{clone.epi_archive}
\title{Clone an \code{epi_archive} object.}
\usage{
clone(x)

\method{clone}{epi_archive}(x)
}
\arguments{
\item{x}{An \code{epi_archive} object.}
}
\description{
Clone an \code{epi_archive} object.
}
