% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{compactify_tibble}
\alias{compactify_tibble}
\title{given a tibble as would be found in an epi_archive, remove duplicate entries.}
\usage{
compactify_tibble(df, keys, tolerance = .Machine$double.eps^0.5)
}
\description{
works by shifting all rows except the version, then comparing values to see
if they've changed. We need to arrange in descending order, but note that
we don't need to group, since at least one column other than version has
changed, and so is kept.
}
\keyword{internal}
