% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditions.R
\name{condition_noncli_message}
\alias{condition_noncli_message}
\title{\code{rlang} non-\code{cli}-format condition \code{message} arg from potentially-long text and subfields}
\usage{
condition_noncli_message(
  init,
  base_message,
  display_subfields,
  more_subfields,
  more_intro,
  more_var_name,
  no_more_intro = more_intro
)
}
\description{
\code{rlang} non-\code{cli}-format condition \code{message} arg from potentially-long text and subfields
}
