% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_converters.R
\name{convert_epiweek_to_date}
\alias{convert_epiweek_to_date}
\title{get a canonical date to represent a given epiweek}
\usage{
convert_epiweek_to_date(epiyear, epiweek, day_of_week = 1)
}
\description{
get a canonical date to represent a given epiweek
}
\keyword{internal}
