% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_converters.R
\name{convert_epiweek_to_season}
\alias{convert_epiweek_to_season}
\title{get a season e.g. "2020/21" for a given year-week pair}
\usage{
convert_epiweek_to_season(epiyear, epiweek, season_start_week = 40)
}
\description{
get a season e.g. "2020/21" for a given year-week pair
}
\keyword{internal}
