% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_converters.R
\name{convert_epiweek_to_season_week}
\alias{convert_epiweek_to_season_week}
\title{get the week in a season}
\usage{
convert_epiweek_to_season_week(epiyear, epiweek, season_start = 40)
}
\description{
get the week in a season
}
\keyword{internal}
