% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_converters.R
\name{convert_to_period_upsample}
\alias{convert_to_period_upsample}
\title{fill in values between, and divide any numeric values equally between them}
\usage{
convert_to_period_upsample(
  to_complete,
  groups,
  columns_to_complete,
  aggregate_columns,
  source_period = 7,
  target_period = 1
)
}
\arguments{
\item{to_complete}{epi_archive}

\item{groups}{to be grouped by. Should include both time_value and version, and any epi_keys}

\item{columns_to_complete}{any columns that need their values extended}

\item{aggregate_columns}{any columns which have numerical data that is a sum
across days, and thus needs to be divided into equal parts distributed
accross days}
}
\description{
fill in values between, and divide any numeric values equally between them
}
