% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{data}
\name{covid_incidence_outliers}
\alias{covid_incidence_outliers}
\title{Subset of JHU daily cases from California and Florida}
\format{
An object of class \code{epi_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 730 rows and 3 columns.
}
\source{
This object contains a modified part of the \href{https://github.com/CSSEGISandData/COVID-19}{COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University} as \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{republished in the COVIDcast Epidata API}. This data set is licensed under the terms of the
\href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0 International license}
by the Johns Hopkins University on behalf of its Center for Systems Science in Engineering.
Copyright Johns Hopkins University 2020.

Modifications:
\itemize{
\item \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{From the COVIDcast Epidata API}:
These signals are taken directly from the JHU CSSE \href{https://github.com/CSSEGISandData/COVID-19}{COVID-19 GitHub repository} without changes.
\item Furthermore, the data has been limited to a very small number of rows, the signal names slightly altered, and formatted into a tibble.
}

This object contains a modified part of the \href{https://github.com/CSSEGISandData/COVID-19}{COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University} as \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{republished in the COVIDcast Epidata API}. This data set is licensed under the terms of the
\href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0 International license}
by the Johns Hopkins University on behalf of its Center for Systems Science in Engineering.
Copyright Johns Hopkins University 2020.
}
\usage{
covid_incidence_outliers
}
\description{
This data source of confirmed COVID-19 cases
is based on reports made available by the Center for
Systems Science and Engineering at Johns Hopkins University.
This example data is a snapshot as of Oct 28, 2021 and captures the cases
from June 1, 2020 to May 31, 2021
and is limited to California and Florida.

Data set for 2 states containing COVID-19 Incident Cases as reported
by JHU-CSSE and Downloaded from the CMU Delphi COVIDcast Epidata API.
This data set covers the period from June 2020 to May 2021, and is
used in the \link{epiprocess} Vignette on Outliers.
}
\examples{
 # Since this is a re-exported dataset, it cannot be loaded using
 # the `data()` function. `data()` looks for a file of the same name
 # in the `data/` directory, which doesn't exist in this package.
 # works
 epiprocess::covid_incidence_outliers

 # works
 library(epiprocess)
 covid_incidence_outliers

 # fails
 data(covid_incidence_outliers, package = "epiprocess")
}
\keyword{datasets}
