% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_converters.R
\name{daily_to_weekly}
\alias{daily_to_weekly}
\title{convert an archive from daily data to weekly data, summing where appropriate}
\usage{
daily_to_weekly(
  epi_arch,
  agg_columns,
  agg_method = c("sum", "mean"),
  day_of_week = 4L,
  day_of_week_end = 6L
)
}
\arguments{
\item{day_of_week}{integer, day of the week, starting from Sunday, select the
date to represent the week in the time_value column, based on it's
corresponding day of the week. The default value represents the week using
Wednesday.}

\item{day_of_week_end}{integer, day of the week starting on Sunday.
Represents the last day, so the week consists of data summed to this day.
The default value \code{6} means that the week is summed from Sunday through
Saturday.}
}
\description{
convert an archive from daily data to weekly data, summing where appropriate
}
\details{
this function is slow, so make sure you are calling it correctly, and
consider testing it on a small portion of your archive first
}
