% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_df.R
\name{decay_epi_df}
\alias{decay_epi_df}
\title{Drop any \code{epi_df} metadata and class on a data frame}
\usage{
decay_epi_df(x)
}
\arguments{
\item{x}{an \code{epi_df} or other data frame}
}
\value{
\code{x} with any metadata dropped and the \code{"epi_df"} class, if previously
present, dropped
}
\description{
Useful in implementing \code{?dplyr_extending} when manipulations cause invariants
of \code{epi_df}s to be violated and we need to return some other class. Note that
this will maintain any grouping (keeping the \code{grouped_df} class and
associated attributes, if present).
}
\keyword{internal}
