% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukey-col-listbacked.R
\name{decay_ukey_col_listbacked}
\alias{decay_ukey_col_listbacked}
\title{Turn an object from being \code{is_ukey_col_listbacked} to not}
\usage{
decay_ukey_col_listbacked(x)
}
\description{
Turn an object from being \code{is_ukey_col_listbacked} to not
}
