% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deletion_column_names}
\alias{deletion_column_names}
\title{potential version_deleted columns}
\usage{
deletion_column_names()
}
\description{
the full list of potential substitutions for the \code{version_deleted} column name:
version_deleted, delete_date, removal_date, deleted_date, deletion_date, removed_date, superseded_date, invalidated_date, Version_Deleted, Delete_Date, Removal_Date, Deleted_Date, Deletion_Date, Removed_Date, Superseded_Date, Invalidated_Date
}
\keyword{internal}
