% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect outliers in variables in an \code{epi_tibble} object}
\usage{
detect_outliers(
  x,
  var,
  methods = tibble(method = "rolling_median", method_args = list(list()), method_abbr =
    "rolling_median"),
  combine_method = c("median", "mean", "none"),
  new_col_name = "outliers"
)
}
\arguments{
\item{x}{The \code{epi_tibble} object under consideration.}

\item{var}{The variable in \code{x} in which to look for outliers.}

\item{methods}{A tibble specifying methods to use for outlier
detection. Contains the following columns:
\itemize{
\item \code{method} <str>: The name of the detection method or a function for
outlier detection.
\item \code{method_args} <lst>: A named list of arguments that will be passed to the
detection method.
\item \code{method_abbr} <chr>: An abbreviation to use in naming output columns with
results from this method
}}

\item{combine_method}{A string, either "median", "mean", or "none",
specifying how to combine results from different outlier detection methods
for the thresholds determining whether a particular observation is
classified as an outlier, as well as a replacement value for any outliers.
If "none", no summarized results are calculated. Note that if the number of
\code{detection_methods} is odd, "median" is equivalent to a majority vote for
purposes of determining whether a given observation is an outlier.}

\item{new_col_name}{String indicating the name of the new column that will
contain the results of outlier detection. Default is "outliers"; note that
setting \code{new_col_name} equal to an existing column name will overwrite this
column.}
}
\value{
The input \code{epi_tibble} \code{x} augmented with outlier detection
thresholds and replacement values from all detection methods.
}
\description{
Applies one or more outlier detection algorithms to variables in an
\code{epi_tibble}, and optionally aggregates the results to create consensus
results.
}
