% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{detect_outliers_one_grp}
\alias{detect_outliers_one_grp}
\title{Function to call all specified outlier detection methods for one group,
assemble results as columns of a tibble.}
\usage{
detect_outliers_one_grp(
  .data_group,
  var,
  methods,
  combine_method,
  new_col_name,
  grouping_vars,
  ...
)
}
\arguments{
\item{.data_group}{epi_tibble for one group. should represent one combination
of key factors determining a reporting unit, e.g., one geo_type or one
combination of geo_type and age group}

\item{var}{The variable in \code{x} in which to look for outliers.}

\item{methods}{A tibble specifying methods to use for outlier
detection. Contains the following columns:
\itemize{
\item \code{method} <str>: The name of the detection method or a function for
outlier detection.
\item \code{method_args} <lst>: A named list of arguments that will be passed to the
detection method.
\item \code{method_abbr} <chr>: An abbreviation to use in naming output columns with
results from this method
}}

\item{combine_method}{A string, either "median", "mean", or "none",
specifying how to combine results from different outlier detection methods
for the thresholds determining whether a particular observation is
classified as an outlier, as well as a replacement value for any outliers.
If "none", no summarized results are calculated. Note that if the number of
\code{detection_methods} is odd, "median" is equivalent to a majority vote for
purposes of determining whether a given observation is an outlier.}

\item{new_col_name}{String indicating the name of the new column that will
contain the results of outlier detection. Default is "outliers"; note that
setting \code{new_col_name} equal to an existing column name will overwrite this
column.}

\item{grouping_vars}{variables that defined groups in the original input to
detect_outliers. These will be dropped from the return value for this
function}
}
\description{
This function is not intended to be called by package users.
}
