% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{detect_outliers_rm}
\alias{detect_outliers_rm}
\title{Detect outliers based on a rolling median}
\usage{
detect_outliers_rm(
  x,
  var,
  n = 21,
  log_transform = FALSE,
  detect_negatives = FALSE,
  detection_multiplier = 2,
  min_radius = 0,
  replacement_multiplier = 0
)
}
\arguments{
\item{x}{The \code{epi_tibble} object under consideration.}

\item{var}{The variable in \code{x} on which to run outlier detection.}

\item{n}{Number of time steps to use in the rolling window. Default is 21.}

\item{log_transform}{Should a log transform be applied before running outlier
detection? Default is \code{FALSE}. If \code{TRUE}, and zeros are present, then the
log transform will be padded by 1.}

\item{detect_negatives}{Should negative values automatically count as
outliers? Default is \code{FALSE}.}

\item{detection_multiplier}{Value determining how far the outlier detection
thresholds are from the rolling median, which are calculated as (rolling
median) +/- (detection multiplier) * (rolling IQR). Default is 2.}

\item{min_radius}{Minimum distance between rolling median and threshold, on
transformed scale. Default is 0.}

\item{replacement_multiplier}{Value determining how far the replacement
values are from the rolling median. The replacement is the original value
if it is within the detection thresholds, or otherwise it is rounded to the
nearest (rolling median) +/- (replacement multiplier) * (rolling IQR).
Default is 0.}
}
\value{
A tibble with the same number of rows as \code{x}, and columns \code{lower},
\code{upper}, and \code{replacement}.
}
\description{
Detects outliers based on a distance from the rolling median specified in
terms of multiples of the rolling interquartile range (IQR).
}
