% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{detect_outliers_rolling_median}
\alias{detect_outliers_rolling_median}
\title{Detect outliers based on a distance from the rolling median specified in
terms of multiples of the rolling IQR.}
\usage{
detect_outliers_rolling_median(
  x,
  var,
  n = 21,
  transform = c("none", "log"),
  detection_multiplier = 2,
  min_radius = 0,
  detect_negatives = FALSE,
  replacement_multiplier = 0
)
}
\arguments{
\item{x}{The \code{epi_tibble} object under consideration.}

\item{var}{The variable in \code{x} in which to look for outliers.}

\item{n}{Number of time steps to use in the window. Default is 15.}

\item{transform}{string specifying transformation to use: "none" or "log"}

\item{detection_multiplier}{specification of how far the outlier detection
threshold is from the rolling median. The thresholds are calculated as
rolling median +/- detection_multiplier * rolling IQR}

\item{min_radius}{minimum distance between rolling median and threshold,
on transformed scale}

\item{detect_negatives}{boolean; if TRUE, negative values count as outliers}

\item{replacement_multiplier}{specification of how far the replacement value
is from the rolling median. The replacement is the original value if it is
within the detection thresholds, or otherwise
rolling median +/- replacement_multiplier * rolling IQR}
}
\value{
a tibble with columns \code{lower}, \code{upper}, and \code{replacement}
}
\description{
Detect outliers based on a distance from the rolling median specified in
terms of multiples of the rolling IQR.
}
