% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{detect_outliers_stl}
\alias{detect_outliers_stl}
\title{Detect outliers based on residuals from an STL decomposition of a signal.}
\usage{
detect_outliers_stl(
  x,
  var,
  n_trend = 21,
  n_seasonal = 21,
  n_threshold = 21,
  seasonal_period = NULL,
  include_seasonality = TRUE,
  transform = c("none", "log"),
  detection_multiplier = 2,
  min_radius = 0,
  detect_negatives = FALSE,
  replacement_multiplier = 0
)
}
\arguments{
\item{x}{The \code{epi_tibble} object under consideration.}

\item{var}{The variable in \code{x} in which to look for outliers.}

\item{n_trend}{Number of time steps to use in the window for trend.
Default is 21.}

\item{n_seasonal}{Number of time steps to use in the window for seasonality.
Default is 21.}

\item{n_threshold}{Number of time steps to use in window for IQR used to
set outlier thresholds.}

\item{seasonal_period}{integer period of seasonality, e.g. 7 for weekly
seasonality with daily data. The default is \code{NULL}, meaning that no
seasonal term is included in the STL decomposition.}

\item{transform}{string specifying transformation to use: "none" or "log"}

\item{detection_multiplier}{specification of how far the outlier detection
threshold is from the rolling median. The thresholds are calculated as
rolling median +/- detection_multiplier * rolling IQR}

\item{min_radius}{minimum distance between rolling median and threshold,
on transformed scale}

\item{detect_negatives}{boolean; if TRUE, negative values count as outliers}

\item{replacement_multiplier}{specification of how far the replacement value
is from the rolling median. The replacement is the original value if it is
within the detection thresholds, or otherwise
rolling median +/- replacement_multiplier * rolling IQR}
}
\value{
a tibble with columns \code{lower}, \code{upper}, and \code{replacement}
}
\description{
Detect outliers based on residuals from an STL decomposition of a signal.
}
