% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{difftime_approx_ceiling_time_delta}
\alias{difftime_approx_ceiling_time_delta}
\title{Closest time_delta that's approximately greater than given difftime}
\usage{
difftime_approx_ceiling_time_delta(difftime, time_type)
}
\arguments{
\item{difftime}{a difftime object}

\item{time_type}{as in \code{\link{validate_slide_window_arg}}}
}
\value{
An object representing an integerish number (or vector of numbers) of
time steps between consecutive time_values of type \code{time_type}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
