% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_analysis.R
\name{difftime_summary}
\alias{difftime_summary}
\title{summary doesn't work on difftimes}
\usage{
difftime_summary(diff_time_val)
}
\description{
summary doesn't work on difftimes
}
\keyword{internal}
