% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_df.R
\name{dplyr_reconstruct.epi_df}
\alias{dplyr_reconstruct.epi_df}
\title{dplyr_reconstruct}
\usage{
\method{dplyr_reconstruct}{epi_df}(data, template)
}
\arguments{
\item{data}{tibble or \code{epi_df} (\code{dplyr} feeds in former, but we may
directly feed in latter from our other methods)}

\item{template}{\code{epi_df} template to use to restore}
}
\value{
\code{epi_df} or degrade into \code{tbl_df}
}
\description{
dplyr_reconstruct
}
\keyword{internal}
