% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{epi_cor}
\alias{epi_cor}
\title{Compute correlations between variables in an \code{epi_df} object}
\usage{
epi_cor(
  x,
  var1,
  var2,
  dt1 = 0,
  dt2 = 0,
  shift_by = geo_value,
  cor_by = geo_value,
  use = "na.or.complete",
  method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{x}{The \code{epi_df} object under consideration.}

\item{var1, var2}{The variables in \code{x} to correlate.}

\item{dt1, dt2}{Time shifts to consider for the two variables, respectively,
before computing correlations. Negative shifts translate into in a lag
value and positive shifts into a lead value; for example, if \code{dt = -1},
then the new value on June 2 is the original value on June 1; if \code{dt = 1},
then the new value on June 2 is the original value on June 3; if \code{dt = 0},
then the values are left as is. Default is 0 for both \code{dt1} and \code{dt2}.}

\item{shift_by}{The variables(s) to group by, for the time shifts. The
default is \code{geo_value}. However, we could also use, for example, \code{shift_by = c(geo_value, age_group)}, assuming \code{x} has a column \code{age_group}, to
perform time shifts per geo value and age group. To omit a grouping
entirely, use \code{cor_by = NULL}. Note that the grouping here is always undone
\emph{before} the correlation computations.}

\item{cor_by}{The variable(s) to group by, for the correlation
computations. If \code{geo_value}, the default, then correlations are computed
for each geo value, over all time; if \code{time_value}, then correlations are
computed for each time, over all geo values. A grouping can also be any
specified using number of columns of \code{x}; for example, we can use \code{cor_by = c(geo_value, age_group)}, assuming \code{x} has a column \code{age_group}, in order
to compute correlations for each pair of geo value and age group. To omit a
grouping entirely, use \code{cor_by = NULL}. Note that the grouping here is
always done \emph{after} the time shifts.}

\item{use, method}{Arguments to pass to \code{cor()}, with "na.or.complete" the
default for \code{use} (different than \code{cor()}) and "pearson" the default for
\code{method} (same as \code{cor()}).}
}
\value{
An tibble with the grouping columns first (\code{geo_value}, \code{time_value},
or possibly others), and then a column \code{cor}, which gives the correlation.
}
\description{
Computes correlations between variables in an \code{epi_df} object, allowing for
grouping by geo value, time value, or any other variables. See the
\href{https://cmu-delphi.github.io/epiprocess/articles/correlation.html}{correlation vignette}
for examples.
}
\examples{
 
# linear association of case and death rates on any given day
epi_cor(x = jhu_csse_daily, 
        var1 = case_rate_7d_av, 
        var2 = death_rate_7d_av, 
        cor_by = "time_value")

# correlation of death rates and lagged case rates
epi_cor(x = jhu_csse_daily, 
        var1 = case_rate_7d_av, 
        var2 = death_rate_7d_av, 
        cor_by = time_value, 
        dt1 = -10)

# correlation grouped by location 
epi_cor(x = jhu_csse_daily, 
        var1 = case_rate_7d_av, 
        var2 = death_rate_7d_av, 
        cor_by = geo_value)

# correlation grouped by location and incorporates lagged cases rates
epi_cor(x = jhu_csse_daily, 
        var1 = case_rate_7d_av, 
        var2 = death_rate_7d_av, 
        cor_by = geo_value, 
        dt1 = -10)
}
