% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_slide_opt_archive.R
\name{epi_slide_opt_archive_one_epikey}
\alias{epi_slide_opt_archive_one_epikey}
\title{Core operation of \code{epi_slide_opt.epi_archive} for a single epikey's history}
\usage{
epi_slide_opt_archive_one_epikey(
  grp_updates,
  in_colnames,
  f_dots_baked,
  f_from_package,
  before,
  after,
  time_type,
  out_colnames
)
}
\arguments{
\item{grp_updates}{tibble with a \code{version} column and measurement columns for
a single epikey, without the epikey labeling columns (e.g., from
\code{group_modify}). Interpretation is analogous to an \code{epi_archive} \code{DT}, but
a specific row order is not required.}

\item{in_colnames}{chr; names of columns to which to apply \code{f_dots_baked}}

\item{f_dots_baked}{supported sliding function from \code{{data.table}} or
\code{{slider}}, potentially with some arguments baked in with
\code{\link[purrr:partial]{purrr::partial}}}

\item{f_from_package}{string; name of package from which \code{f_dots_baked}
(pre-\code{partial}) originates}

\item{before}{integerish >=0 or Inf; number of time steps before each
ref_time_value to include in the sliding window computation; Inf to include
all times beginning with the min \code{time_value}}

\item{after}{integerish >=0; number of time steps after each ref_time_value
to include in the sliding window computation}

\item{time_type}{as in \code{new_epi_archive}}

\item{out_colnames}{chr, same length as \code{in_colnames}; column names to use
for results}
}
\value{
tibble with a \code{version} column, pre-existing measurement columns, and
new measurement columns; (compactified) diff data to put into an
\code{epi_archive}. May not match column ordering; may not ensure any row
ordering.
}
\description{
Core operation of \code{epi_slide_opt.epi_archive} for a single epikey's history
}
\examples{

library(dplyr)
grp_updates <- bind_rows(
  tibble(version = 20, time_value = 1:20, value = 1:20),
  tibble(version = 22, time_value = 4:5, value = 5:4),
  tibble(version = 23, time_value = 8, value = 9),
  tibble(version = 24, time_value = 11, value = NA),
  tibble(version = 25, time_value = -10, value = -10),
  tibble(version = 26, time_value = 50, value = 50)
) \%>\%
  mutate(across(c(version, time_value), ~ as.Date("2020-01-01") - 1 + .x))

f <- purrr::partial(data.table::frollmean, algo = "exact")

grp_updates \%>\%
  epiprocess:::epi_slide_opt_archive_one_epikey(
    "value", f, "data.table", 2L, 0L, "day", "slide_value"
  )

}
\keyword{internal}
