% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_slide_opt_edf.R
\name{epi_slide_opt_edf_one_epikey}
\alias{epi_slide_opt_edf_one_epikey}
\title{Run a specialized slide computation on a single \code{epi_df} epikey, with temporary completion}
\usage{
epi_slide_opt_edf_one_epikey(
  inp_tbl,
  in_colnames,
  f_dots_baked,
  f_from_package,
  before_n_steps,
  after_n_steps,
  unit_step,
  time_type,
  out_colnames,
  out_filter_time_range,
  out_filter_time_set
)
}
\arguments{
\item{inp_tbl}{tibble; should have a \code{time_value} column and columns named in
\code{in_colnames}; should not contain any columns named in \code{out_colnames}}

\item{in_colnames}{chr; names of columns to which to apply \code{f_dots_baked}}

\item{f_dots_baked}{supported sliding function from \code{{data.table}} or
\code{{slider}}, potentially with some arguments baked in with
\code{\link[purrr:partial]{purrr::partial}}}

\item{f_from_package}{string; name of package from which \code{f_dots_baked}
(pre-\code{partial}) originates}

\item{before_n_steps}{integerish \verb{>=0} or \code{Inf}; number of time steps before
each \code{ref_time_value} to include in the sliding window computation; \code{Inf}
to include all times beginning with the min \code{time_value}}

\item{after_n_steps}{integerish \verb{>=0}; number of time steps after each
\code{ref_time_value} to include in the sliding window computation}

\item{time_type}{as in \code{new_epi_archive}}

\item{out_colnames}{chr, same length as \code{in_colnames}; column names to use
for results}

\item{out_filter_time_range, out_filter_time_set}{\code{time_value} filter;
\code{time_values} in the output should match the result of applying this filter
to \code{inp_tbl$time_value}. Exactly one of the two must be provided
(non-\code{NULL}) and the other must be \code{NULL}. \code{out_filter_time_range}, if
provided, should be a length-2 vector/list containing the minimum and
maximum \code{time_value} to allow in the output. \code{out_filter_time_set}, if
provided, should be a vector of \code{time_values} to intersect with the input
\code{time_value}s.}
}
\value{
tibble; like \code{inp_tbl} with addition of \code{out_colnames} holding the
slide computation results, with times filtered down as specified
}
\description{
Run a specialized slide computation on a single \code{epi_df} epikey, with temporary completion
}
\examples{

library(dplyr)
tbl <- tibble(
  time_value = c(11:12, 15:18) + 0,
  value = c(c(1, 2), c(4, 8, 16, 32))
)

tbl \%>\%
  epi_slide_opt_edf_one_epikey(
    "value",
    frollmean, "data.table",
    1L, 0L, 1L, "integer",
    "slide_value",
    c(11L, 16L), NULL
  )

tbl \%>\%
  epi_slide_opt_edf_one_epikey(
    "value",
    frollmean, "data.table",
    0L, 1L, 1L, "integer",
    "slide_value",
    NULL, c(11, 15, 16, 17, 18)
  )

tbl \%>\%
  epi_slide_opt_edf_one_epikey(
    "value",
    frollmean, "data.table",
    Inf, 0L, 1L, "integer",
    "slide_value",
    NULL, c(12, 17)
  )

}
\keyword{internal}
