% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{data}
\name{cases_deaths_subset}
\alias{cases_deaths_subset}
\alias{covid_incidence_county_subset}
\alias{covid_incidence_outliers}
\alias{archive_cases_dv_subset}
\alias{covid_case_death_rates_extended}
\title{Selected example data sets from \code{epidatasets}}
\format{
An object of class \code{epi_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4026 rows and 6 columns.

An object of class \code{epi_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16212 rows and 5 columns.

An object of class \code{epi_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 730 rows and 3 columns.

An object of class \code{epi_archive} of length 6.

An object of class \code{epi_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 37576 rows and 4 columns.
}
\usage{
cases_deaths_subset

covid_incidence_county_subset

covid_incidence_outliers

archive_cases_dv_subset

covid_case_death_rates_extended
}
\description{
Data sets re-exported from \code{epidatasets}; please see
documentation for each of these objects in \code{epidatasets}.

A brief description of the format of each of the objects above are described
in matching order below.
}
\keyword{internal}
