% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitools.R
\docType{package}
\name{epipred}
\alias{epipred}
\title{epipred: Tools for Basic Signal Processing in Epidemiology}
\description{
The \code{epitools} package provides contains tools for basic signal processing
tasks in epidemiology, which includes computing percentage changes in
signals, estimating derivatives, and computing correlations.
}
