% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiprocess-package.R
\docType{package}
\name{epiprocess-package}
\alias{epiprocess}
\alias{epiprocess-package}
\title{epiprocess: Tools for basic signal processing in epidemiology}
\description{
This package introduces common data structures for working with epidemiological data reported by location and time and offers associated utilities to perform basic signal processing tasks. The package is designed to be used in conjunction with `epipredict` for building and evaluating epidemiological models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cmu-delphi/epiprocess}
  \item \url{https://cmu-delphi.github.io/epiprocess/}
}

}
\author{
\strong{Maintainer}: Logan Brooks \email{lcbrooks+github@andrew.cmu.edu}

Authors:
\itemize{
  \item Logan Brooks \email{lcbrooks+github@andrew.cmu.edu}
  \item Nat DeFries
  \item Daniel McDonald
  \item Evan Ray
  \item Dmitry Shemetov
  \item Ryan Tibshirani
}

Other contributors:
\itemize{
  \item Jacob Bien [contributor]
  \item Rafael Catoia [contributor]
  \item Rachel Lobay [contributor]
  \item Ken Mawer [contributor]
  \item Chloe You [contributor]
  \item Quang Nguyen [contributor]
  \item David Weber \email{davidweb@andrew.cmu.edu} [contributor]
  \item Lionel Henry (Author of included rlang fragments) [contributor]
  \item Hadley Wickham (Author of included rlang fragments) [contributor]
  \item Posit (Copyright holder of included rlang fragments) [copyright holder]
  \item Johns Hopkins University Center for Systems Science and Engineering (Owner of COVID-19 cases and deaths data from the COVID-19 Data Repository) [data contributor]
  \item Johns Hopkins University (Copyright holder of COVID-19 cases and deaths data from the COVID-19 Data Repository) [copyright holder]
  \item Carnegie Mellon University Delphi Group (Owner of claims-based CLI data from the Delphi Epidata API) [data contributor]
}

}
\keyword{internal}
