% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiprocess.R
\docType{package}
\name{epiprocess}
\alias{epiprocess}
\title{epiprocess: Tools for basic signal processing in epidemiology}
\description{
This package introduces a common data structure for epidemiological data sets
measured over space and time, and offers associated utilities to perform
basic signal processing tasks.
}
