% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitools.R
\docType{package}
\name{epitools}
\alias{epitools}
\title{epitools: Tools for basic signal processing in epidemiology}
\description{
The \code{epitools} package provides contains tools for basic signal processing
tasks in epidemiology, which includes computing percentage changes in
signals, estimating derivatives, and computing correlations.
}
