% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_converters.R
\name{epiweeks_in_year}
\alias{epiweeks_in_year}
\title{get a total count of the epiweeks in a given year}
\usage{
epiweeks_in_year(year)
}
\description{
get a total count of the epiweeks in a given year
}
\keyword{internal}
