% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting-helpers2.R
\name{epix_confkeys}
\alias{epix_confkeys}
\title{For each \code{ekt_var_subset} key, the \code{version}s it had with a "real" diff for any \code{val_var_subset}}
\usage{
epix_confkeys(
  x,
  val_var_subset,
  ekt_var_subset = key_colnames(x, exclude = c("time_value", "version"))
)
}
\arguments{
\item{x}{an \code{epi_archive}}

\item{val_var_subset}{technically optional; var_subset of
\code{val_colnames(x)}, but likely just a single var; defaults to
entire set}

\item{ekt_var_subset}{optional; var_subset of
\code{key_colnames(x, exclude = "version")}; defaults to
\code{key_colnames(x, exclude = c("time_value" ,"version"))}}
}
\value{
tibble with names matching \code{ekt_var_subset}; note that if you
want downstream code to adapt to the choice of \code{ekt_var_subset}, you
will also need to pass \code{ekt_var_subset} to later processing and may
need to specify \code{relationship = "many-to-{one,many}"} if joining
to other data.
}
\description{
Here, "real" means either (a) changing from unobserved to a non-NA
initial estimate, or (b) changing from one estimate to a different
estimate.  It does not count moving from unobserved to an NA
initial estimate, as these are often not actually present in the
underlying data and are simply present due to storing multiple val
vars in the same archive.
}
\details{
The choice of \code{ekt_var_subset} is a tradeoff between (a) recognizing the
maturity of estimates that remained the same between versions for
reasons other than an outage, plus enabling models to treat more
estimates as equally mature, vs. (b) recognizing the immaturity of
estimates that remained the same between versions due to more
complex types of outages.
\itemize{
\item If some epikeys (geodemographic groups) can have pauses in
reporting not shared by others, then you should probably include
\code{c("geo_value", x$other_keys)} in \code{ekt_var_subset}.
\item If some \code{time_value}s can have pauses in reporting not shared by
others, then you should probably include \code{"time_value"} in
\code{ekt_var_subset}.  Depending on how downstream models treat NAs, you
may be able to get away with not adding it if these pauses in
reporting simply mean that no new \code{time_value}s are being added
to the data set, but all previously-observed \code{time_value}s are
receiving revised estimates.
}
}
\keyword{internal}
