% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting-helpers2.R
\name{epix_diff_ekvs}
\alias{epix_diff_ekvs}
\title{epikey-versions that added a non-NA measurement or changed a measurement for \code{vars}}
\usage{
epix_diff_ekvs(x, vars = val_colnames(x))
}
\arguments{
\item{x}{(unvalidated) an \code{epi_archive}}

\item{vars}{(unvalidated) subset of \code{val_colnames(x)}}
}
\value{
tibble with names \code{key_colnames(x, exclude = "time_value")}
}
\description{
For each epikey, this tries to get the set of versions with "real"
updates for the \code{vars} columns.  We can't just directly use
\code{unique} because (i) more-available signals than \code{vars} may create
spurious results, and (ii) \code{x} might not be compactified.
}
