% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epix_doctor_visits}
\alias{epix_doctor_visits}
\title{California and Florida daily percentage of doctor’s visits with COVID-like illness.}
\format{
A tibble with 718 rows and 4 variables:
\describe{
\item{geo_value}{the geographic value associated with each row of measurements.}
\item{time_value}{the time value associated with each row of measurements.}
\item{version}{the time value specifying the version for each row of measurements. For example, if in a given row the version is January 15, 2022 and time_value is January 14, 2022, then this row contains the measurements of the data for January 14, 2022 that were available one day later.}
\item{percent_cli}{percentage of doctor’s visits with CLI (COVID-like illness) computed from medical insurance claims}
...
}
}
\source{
\url{https://cmu-delphi.github.io/delphi-epidata/api/covidcast.html}
}
\usage{
epix_doctor_visits
}
\description{
An epi_archive dataset containing the daily COVID-19 percentage of doctor’s visits with CLI (COVID-like illness)
computed from medical insurance claims from June 1, 2020 to July 1, 2020.
}
\keyword{datasets}
