% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epix_epi_slide_opt.R
\name{epix_epi_slide_opt_one_epikey}
\alias{epix_epi_slide_opt_one_epikey}
\title{Core operation of \code{epix_epi_slide_opt} for a single epikey's history}
\usage{
epix_epi_slide_opt_one_epikey(
  updates,
  in_colnames,
  f_dots_baked,
  f_from_package,
  before,
  after,
  time_type,
  out_colnames
)
}
\arguments{
\item{updates}{tibble with two columns: \code{version} and \code{subtbl}; \code{subtbl} is a
list of tibbles, each with a \code{time_value} column and measurement columns.
The epikey should not appear.}

\item{in_colnames}{chr; names of columns to which to apply \code{f_dots_baked}}

\item{f_dots_baked}{supported sliding function from \code{{data.table}} or
\code{{slider}}, potentially with some arguments baked in with
\code{\link[purrr:partial]{purrr::partial}}}

\item{f_from_package}{string; name of package from which \code{f_dots_baked}
(pre-\code{partial}) originates}

\item{before}{integerish >=0 or Inf; number of time steps before each
ref_time_value to include in the sliding window computation; Inf to include
all times beginning with the min \code{time_value}}

\item{after}{integerish >=0; number of time steps after each ref_time_value
to include in the sliding window computation}

\item{time_type}{as in \code{new_epi_archive}}

\item{out_colnames}{chr, same length as \code{in_colnames}; column names to use
for results}
}
\value{
list of tibbles with same names as \code{subtbl}s plus: \code{c(out_colnames, "version")}; (compactified) diff data to put into an \code{epi_archive}
}
\description{
Core operation of \code{epix_epi_slide_opt} for a single epikey's history
}
\examples{

library(dplyr)
updates <- bind_rows(
  tibble(
    version = 40, time_value = 1:10, value = 1:10
  ),
  tibble(
    version = 12, time_value = 2:3, value = 3:2
  ),
  tibble(
    version = 13, time_value = 6, value = 7,
  ),
  tibble(
    version = 13, time_value = 7, value = NA,
  )
) \%>\%
  mutate(across(c(version, time_value), ~ as.Date("2020-01-01") - 1 + .x)) \%>\%
  tidyr::nest(.by = version, .key = "subtbl")

updates \%>\%
  epix_epi_slide_opt_one_epikey("value", data.table::frollmean, "data.table", 1L, 0L, "day", "slide_value")

}
\keyword{internal}
