% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_archive_new.R
\name{epix_merge2}
\alias{epix_merge2}
\title{Merge two \code{epi_archive} objects}
\usage{
epix_merge2(
  x,
  y,
  sync = c("forbid", "na", "locf", "truncate"),
  compactify = TRUE
)
}
\arguments{
\item{x, y}{Two \code{epi_archive} objects to join together.}

\item{sync}{Optional; \code{"forbid"}, \code{"na"}, \code{"locf"}, or \code{"truncate"}; in the
case that \code{x$versions_end} doesn't match \code{y$versions_end}, what do we do?:
\code{"forbid"}: emit an error; "na": use \code{max(x$versions_end, y$versions_end)}
as the result's \code{versions_end}, but ensure that, if we request a snapshot
as of a version after \code{min(x$versions_end, y$versions_end)}, the
observation columns from the less up-to-date archive will be all NAs (i.e.,
imagine there was an update immediately after its \code{versions_end} which
revised all observations to be \code{NA}); \code{"locf"}: use \code{max(x$versions_end, y$versions_end)} as the result's \code{versions_end}, allowing the last version
of each observation to be carried forward to extrapolate unavailable
versions for the less up-to-date input archive (i.e., imagining that in the
less up-to-date archive's data set remained unchanged between its actual
\code{versions_end} and the other archive's \code{versions_end}); or \code{"truncate"}:
use \code{min(x$versions_end, y$versions_end)} as the result's \code{versions_end},
and discard any rows containing update rows for later versions.}

\item{compactify}{Optional; \code{TRUE}, \code{FALSE}, or \code{NULL}; should the result be
compactified? See \code{\link{as_epi_archive}} for an explanation of what this means.
Default here is \code{TRUE}.}
}
\value{
the resulting \code{epi_archive}
}
\description{
Merges two \code{epi_archive}s that share a common \code{geo_value}, \code{time_value}, and
set of key columns. When they also share a common \code{versions_end},
using \verb{$as_of} on the result should be the same as using \verb{$as_of} on \code{x} and
\code{y} individually, then performing a full join of the \code{DT}s on the non-version
key columns (potentially consolidating multiple warnings about clobberable
versions). If the \code{versions_end} values differ, the
\code{sync} parameter controls what is done.
}
\details{
This function, \code{\link{epix_merge}}, does not mutate its inputs and will not alias
either archive's \code{DT}, but may alias other fields; \code{x$merge} will overwrite
\code{x} with the result of the merge, reseating its \code{DT} and several other fields
(making them point to different objects), but avoiding mutation of the
contents of the old \code{DT} (only relevant if you have another reference to the
old \code{DT} in another object).

In all cases, \code{additional_metadata} will be an empty list, and
\code{clobberable_versions_start} will be set to the earliest version that could
be clobbered in either input archive.
}
\examples{
# create two example epi_archive datasets
x <- archive_cases_dv_subset$DT \%>\%
  dplyr::select(geo_value, time_value, version, case_rate_7d_av) \%>\%
  as_epi_archive(compactify = TRUE)
y <- archive_cases_dv_subset$DT \%>\%
  dplyr::select(geo_value, time_value, version, percent_cli) \%>\%
  as_epi_archive(compactify = TRUE)
# merge results stored in a third object:
xy <- epix_merge(x, y)
# vs. mutating x to hold the merge result:
x$merge(y)

}
