% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{epix_pivot_wider}
\alias{epix_pivot_wider}
\title{Pivot an archive from long to wide}
\usage{
epix_pivot_wider(x, names_from, values_from)
}
\description{
Pivot an archive from long to wide
}
\keyword{internal}
