% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting-helpers.R
\name{epix_select_predictor_shifts}
\alias{epix_select_predictor_shifts}
\title{Version-aware {fore,now,back}casting: select predictor time shifts to use, prepare training&test data}
\usage{
epix_select_predictor_shifts(
  training_tbl,
  archive,
  predictors,
  search_predictor_shifts_within = as.difftime(60, units = "days"),
  predictor_search_offset = 0L,
  min_n_predictor_shifts = 0L,
  max_n_predictor_shifts = 3L,
  min_predictor_shift_spacing = difftime_approx_ceiling_time_delta(as.difftime(7, units =
    "days"), archive$time_type),
  min_n_training_versions = 12L,
  min_n_training_rows_to_n_predictor_shifts_ratio = 30
)
}
\description{
Version-aware {fore,now,back}casting: select predictor time shifts to use, prepare training&test data
}
\examples{

fcd_horizon <- 0L
archive <- archive_cases_dv_subset \%>\%
  epix_as_of(.$versions_end - 80, all_versions = TRUE)
training_tbl <- archive \%>\%
  epix_target_evaluation_data("case_rate_7d_av", fcd_horizon) \%>\%
  na.omit()
# TODO single epikey?
model_data <- epix_select_predictor_shifts(training_tbl, archive, c("case_rate_7d_av", "percent_cli"))
fit <- quantreg::rq(case_rate_7d_av_0dlag_evaluation ~ ., data = model_data$training_tbl)
pred <- predict(fit, 0.5, model_data$testing_tbl)
print(model_data$testing_tbl)
print(pred)
eval_data <- archive_cases_dv_subset \%>\%
  epix_target_evaluation_data("case_rate_7d_av", fcd_horizon, anchor_versions = .$versions_end - 80)
print(eval_data)

}
