% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_epi_archive.R
\name{epix_truncate_versions_after.grouped_epi_archive}
\alias{epix_truncate_versions_after.grouped_epi_archive}
\title{Truncate versions after a given version, grouped}
\usage{
\method{epix_truncate_versions_after}{grouped_epi_archive}(grouped_epi_archive, max_version)
}
\description{
Truncate versions after a given version, grouped
}
