% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_archive.R, R/grouped_epi_archive.R
\name{epix_truncate_versions_after}
\alias{epix_truncate_versions_after}
\alias{epix_truncate_versions_after.epi_archive}
\alias{epix_truncate_versions_after.grouped_epi_archive}
\title{Filter an \code{epi_archive} object to keep only older versions}
\usage{
epix_truncate_versions_after(x, max_version)

\method{epix_truncate_versions_after}{epi_archive}(x, max_version)

\method{epix_truncate_versions_after}{grouped_epi_archive}(x, max_version)
}
\arguments{
\item{x}{An \code{epi_archive} object.}

\item{max_version}{The latest version to include in the archive.}
}
\value{
An \code{epi_archive} object
}
\description{
Generates a filtered \code{epi_archive} from an \code{epi_archive} object, keeping
only rows with \code{version} falling on or before a specified date.
}
