% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\name{exclusive}
\alias{exclusive}
\title{Mark a threshold as an exclusive bound using a wrapper class}
\usage{
exclusive(x)
}
\description{
Mark a threshold as an exclusive bound using a wrapper class
}
\keyword{internal}
