% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_patterns.R
\name{fill_missing_updates}
\alias{fill_missing_updates}
\title{Perform LOCF to fill missing values in issues}
\usage{
fill_missing_updates(df, value_col, refd_col, lag_col, issued_col)
}
\arguments{
\item{df}{Data Frame of aggregated counts within a single location
reported for each reference date and issue date.}

\item{value_col}{column name for the column of counts}

\item{refd_col}{column name for the column of reference date}

\item{lag_col}{column name for the column of lag}

\item{issued_col}{name of the issue (report) date column}
}
\description{
Perform LOCF to fill \code{NA}s if a group is missing from an issue but was
previously available. If there is no update on issue date \eqn{D} but
previous reports exist for issue date \eqn{D_p} < \eqn{D}, all the dates
between \eqn{[D_p, D]} are filled with the value reported on date \eqn{D_p}.
If there is no update for any previous issue date, fill in with 0.
}
