% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_patterns.R
\name{fill_rows}
\alias{fill_rows}
\title{Create empty obs for every sequentially missing date-lag combo}
\usage{
fill_rows(df, refd_col, lag_col, min_refd, max_refd, ref_lag)
}
\arguments{
\item{df}{Data Frame of aggregated counts within a single location
reported for each reference date and issue date.}

\item{refd_col}{column name for the column of reference date}

\item{lag_col}{column name for the column of lag}

\item{min_refd}{the earliest reference date considered in the data}

\item{max_refd}{the latest reference date considered in the data}

\item{ref_lag}{the maximum lag value through which to complete}
}
\value{
df_new Data Frame with filled rows for missing lags
}
\description{
Make sure all reference date have enough rows for updates. Re-index
dataframe to insert a row for each missing date between \code{min_refd}
and \code{max_refd} for every requested lag from 1 to \code{ref_lag}, filling with \code{NA}.
}
