% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{fill_through_version.epi_archive}
\alias{fill_through_version.epi_archive}
\title{Fill through version}
\usage{
\method{fill_through_version}{epi_archive}(epi_archive, fill_versions_end, how = c("na", "locf"))
}
\arguments{
\item{epi_archive}{an \code{epi_archive} object}

\item{fill_versions_end}{as in \code{\link{epix_fill_through_version}}}

\item{how}{as in \code{\link{epix_fill_through_version}}}
}
\description{
Fill in unobserved history using requested scheme by mutating
the given object and potentially reseating its fields. See
\code{\link{epix_fill_through_version}}, which doesn't mutate the input archive but
might alias its fields.
}
