% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditions.R
\name{fn_max_n_args}
\alias{fn_max_n_args}
\title{Get the max number of args a function can accept (as a double)}
\usage{
fn_max_n_args(f)
}
\arguments{
\item{f}{the function to inspect}
}
\value{
a double; if \code{f} takes in dots arguments (\code{...}), then \code{Inf}; else,
the number of arguments that \code{f} accepts, converted from an integer to a
double
}
\description{
Get the max number of args a function can accept (as a double)
}
