% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{force_meta}
\alias{force_meta}
\title{Manually override \code{geo_type} and/or \code{time_type} metadata}
\usage{
force_meta(x, geo_type = NULL, time_type = NULL)
}
\arguments{
\item{x}{an \code{epi_df}}

\item{geo_type}{optional; new \code{geo_type} string}

\item{time_type}{optional; new \code{time_type} string}
}
\value{
\code{x} with updated metadata
}
\description{
Manually override \code{geo_type} and/or \code{time_type} metadata
}
\keyword{internal}
