% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forcing_raises_error}
\alias{forcing_raises_error}
\title{Does attempting to (quietly) force \code{x} raise an error?}
\usage{
forcing_raises_error(x)
}
\arguments{
\item{x}{argument to \code{\link{force}}}
}
\value{
Boolean
}
\description{
Does not silence any direct output to the message stream / stderr()
not from messages, warnings, or errors.
}
\keyword{internal}
