% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_chr_deparse}
\alias{format_chr_deparse}
\title{Format a character vector as a string via deparsing it}
\usage{
format_chr_deparse(x)
}
\arguments{
\item{x}{\code{chr}; e.g., output of \code{class(object)} for some \code{object}}
}
\value{
string
}
\description{
Format a character vector as a string via deparsing it
}
\keyword{internal}
