% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_chr_with_quotes}
\alias{format_chr_with_quotes}
\title{Format a character vector as a string via deparsing/quoting each}
\usage{
format_chr_with_quotes(x, empty = "*none*")
}
\arguments{
\item{x}{\code{chr}; e.g., \code{colnames} of some data frame}

\item{empty}{string; what should be output if \code{x} is of length 0?}
}
\value{
string
}
\description{
Format a character vector as a string via deparsing/quoting each
}
\keyword{internal}
