% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_class_vec}
\alias{format_class_vec}
\title{Format a class vector as a string via deparsing it}
\usage{
format_class_vec(class_vec)
}
\arguments{
\item{class_vec}{\code{chr}; output of \code{class(object)} for some \code{object}}
}
\value{
string
}
\description{
Format a class vector as a string via deparsing it
}
\keyword{internal}
