% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_tibble_row}
\alias{format_tibble_row}
\title{Format a tibble row as chr}
\usage{
format_tibble_row(x, empty = "*none*")
}
\arguments{
\item{x}{a tibble with a single row}
}
\value{
\code{chr} with one entry per column, of form "\if{html}{\out{<colname>}} = \if{html}{\out{<deparsed col value>}}"
}
\description{
Format a tibble row as chr
}
\keyword{internal}
