% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{format_time_delta}
\alias{format_time_delta}
\title{Format a length-1 time delta to a character to assist messaging}
\usage{
format_time_delta(x, time_type, require_integer = FALSE)
}
\description{
This is meant to address the following:
\itemize{
\item \code{glue::glue("{as.difftime(1, units = 'days')}")} is "1"
\item \code{glue::glue("{format(as.difftime(1, units = 'days'))}")} is "1 days"
\item time deltas for yearmonths and integers don't have units attached at all
}
}
\keyword{internal}
