% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_varname}
\alias{format_varname}
\title{"Format" column/variable name for cli interpolation}
\usage{
format_varname(x)
}
\arguments{
\item{x}{string; e.g., a colname}
}
\value{
string
}
\description{
Designed to give good output if interpolated with cli. Main purpose is to add
backticks around variable names when necessary.
}
\keyword{internal}
