% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_varnames}
\alias{format_varnames}
\title{"Format" a character vector of column/variable names for cli interpolation}
\usage{
format_varnames(x, empty = "*none*")
}
\arguments{
\item{x}{\code{chr}; e.g., \code{colnames} of some data frame}

\item{empty}{string; what should be output if \code{x} is of length 0?}
}
\value{
\code{chr}
}
\description{
Designed to give good output if interpolated with cli. Main purpose is to add
backticks around variable names when necessary, and something other than an
empty string if length 0.
}
\keyword{internal}
