% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{full_date_seq}
\alias{full_date_seq}
\title{Make a complete date sequence between min(x$time_value) and max
(x$time_value). Produce lists of dates before min(x$time_value) and after
max(x$time_value) for padding initial and final windows to size \code{n}.}
\usage{
full_date_seq(x, before, after, time_type)
}
\description{
\code{before} and \code{after} args are assumed to have been validated by the calling
function (using \code{validate_slide_window_arg}).
}
\keyword{internal}
