% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gcd_num}
\alias{gcd_num}
\title{Find the greatest common divisor of all entries in a numeric vector}
\usage{
gcd_num(dividends, ..., rrtol = 0.000001, pqlim = 1000000, irtol = 0.000001)
}
\arguments{
\item{dividends}{\code{is.numeric}, \code{length} > 0; the dividends for which to find
the greatest common divisor.}

\item{...}{should be empty; forces the following parameters to be passed by
name}

\item{rrtol}{Optional, length 1, \code{is.numeric}, non-negative; the remainder
relative tolerance: consider a remainder from a division operation to be
zero if it is \code{abs(remainder/divisor) <= rrtol}. Could also be described as
a tolerance on the fractional part of the proper quotient. Default is 1e-6.}

\item{pqlim}{Optional, length 1, \code{is.numeric}, non-negative; the proper
quotient limit: consider a divisor to be zero if \code{abs(dividend/divisor) >= pqlim}.}

\item{irtol}{Optional, length 1, \code{is.numeric}, non-negative; the iterand
relative tolerance: consider \code{a} and \code{b} to have no gcd if the absolute
value of an iterand (and consequently also any result that might be
subsequently produced, as absolute values of iterands are decreasing) is
\verb{<= irtol * a} or \verb{<= irtol * b}. Also can be seen as the reciprocal of a
limit on the number \code{k} needed to achieve \code{k * gcd_result == max(abs(a),abs(b))}.}
}
\value{
Same \code{\link[vctrs:vec_ptype]{vctrs::vec_ptype}} as \code{dividends}, \code{length} 1: the gcd. (Or an
error.)
}
\description{
Find the greatest common divisor of all entries in a numeric vector
}
\keyword{internal}
