% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geo_column_names}
\alias{geo_column_names}
\title{potential geo_value columns}
\usage{
geo_column_names()
}
\description{
the full list of potential substitutions for the \code{geo_value} column name:
geo_value, geo_values, geo_id, geos, location, jurisdiction, fips, zip, county, hrr, msa, state, province, nation, states, provinces, counties, geo_Value, region, regions, country, countries, state_code, fips_code, zip_code, location_id, location_code, Geo_Value, Geo_Values, Geo_Id, Geos, Location, Jurisdiction, Fips, Zip, County, Hrr, Msa, State, Province, Nation, States, Provinces, Counties, Geo_Value, Region, Regions, Country, Countries, State_Code, Fips_Code, Zip_Code, Location_Id, Location_Code
}
\keyword{internal}
