% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_patterns.R
\name{get_7dav}
\alias{get_7dav}
\title{Calculate 7 day moving average for each issue date
The 7dav for date \eqn{D} reported on issue date \eqn{D_i} is the average from \eqn{D-7} to \eqn{D-1}}
\usage{
get_7dav(pivot_df, refd_col)
}
\arguments{
\item{pivot_df}{Data Frame where the columns are issue dates and the rows are
reference dates}

\item{refd_col}{column name for the column of reference date}
}
\description{
Calculate 7 day moving average for each issue date
The 7dav for date \eqn{D} reported on issue date \eqn{D_i} is the average from \eqn{D-7} to \eqn{D-1}
}
