% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_patterns.R
\name{get_lags}
\alias{get_lags}
\title{Calculate an implied lag column as the difference between reference
date and report date. Meant to be used when \code{lag_col} is not specified.
Don't add field to dataframe directly.}
\usage{
get_lags(df, lag_col, refd_col, issued_col)
}
\arguments{
\item{df}{Data Frame of aggregated counts within a single location
reported for each reference date and issue date.}

\item{lag_col}{column name for the column of lag}

\item{refd_col}{column name for the column of reference date}

\item{issued_col}{name of the issue (report) date column}
}
\description{
Calculate an implied lag column as the difference between reference
date and report date. Meant to be used when \code{lag_col} is not specified.
Don't add field to dataframe directly.
}
