% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_analysis.R
\name{get_last_run}
\alias{get_last_run}
\title{return the first value in values_from from the last string of trues in bool_vec}
\usage{
get_last_run(bool_vec, values_from)
}
\description{
the point of this operation is to get the value in values_from which occurs
at the same index as the start of the last run of true values in bool_vec.
for example, in c(1,1,0,1,1), we want the 4th entry, since there's a 0
breaking the run
}
\keyword{internal}
