% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditions.R
\name{global_handlers_list}
\alias{global_handlers_list}
\title{List current named global calling handlers}
\usage{
global_handlers_list()
}
\description{
This is just a version check followed by \code{\link[base:conditions]{base::globalCallingHandlers}}.
It's here just to make it easier for a user to find if they are first
introduced to global handlers via other functions in this package.
}
\seealso{
Other functions working with global calling handler stack: 
\code{\link{check_global_handlers_stack_support}()},
\code{\link{global_handlers_push}()},
\code{\link{global_handlers_remove}()}
}
\concept{functions working with global calling handler stack}
