% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_epi_archive.R, R/methods-epi_archive.R
\name{group_by_drop_default.grouped_epi_archive}
\alias{group_by_drop_default.grouped_epi_archive}
\alias{group_by.grouped_epi_archive}
\alias{groups.grouped_epi_archive}
\alias{ungroup.grouped_epi_archive}
\alias{group_by.epi_archive}
\title{\code{group_by} and related methods for \code{epi_archive}, \code{grouped_epi_archive}}
\usage{
\method{group_by_drop_default}{grouped_epi_archive}(.tbl)

\method{group_by}{grouped_epi_archive}(.data, ..., .add = FALSE, .drop = dplyr::group_by_drop_default(.data))

\method{groups}{grouped_epi_archive}(x)

\method{ungroup}{grouped_epi_archive}(x, ...)

\method{group_by}{epi_archive}(.data, ..., .add = FALSE, .drop = dplyr::group_by_drop_default(.data))
}
\description{
\code{group_by} and related methods for \code{epi_archive}, \code{grouped_epi_archive}
}
\details{
To match \code{dplyr}, \code{group_by} allows "data masking" (also referred to as
"tidy evaluation") expressions \code{...}, not just column names, in a way similar
to \code{mutate}. Note that replacing or removing key columns with these
expressions is disabled.

Mutation and aliasing: \code{group_by} tries to use a shallow copy of the \code{DT},
introducing column-level aliasing between its input and its result. This
doesn't follow the general model for most \code{data.table} operations, which
seems to be that, given an nonaliased (i.e., unique) pointer to a
\code{data.table} object, its pointers to its columns should also be nonaliased.
If you mutate any of the columns of either the input or result, first ensure
that it is fine if columns of the other are also mutated, but do not rely on
such behavior to occur. Additionally, never perform mutation on the key
columns at all (except for strictly increasing transformations), as this will
invalidate sortedness assumptions about the rows.

\code{group_by_drop_default} on (ungrouped) \code{epi_archive}s is expected to dispatch
to \code{group_by_drop_default.default} (but there is a dedicated method for
\code{grouped_epi_archive}s).
}
