% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_df.R
\name{group_by.epi_df}
\alias{group_by.epi_df}
\alias{ungroup.epi_df}
\title{Group or ungroup \code{epi_df} object}
\usage{
\method{group_by}{epi_df}(x, ...)

\method{ungroup}{epi_df}(x, ...)
}
\description{
Groups or ungroups an \code{epi_df}, preserving class and attributes.
}
