% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tibble.R
\name{group_by.epi_tibble}
\alias{group_by.epi_tibble}
\title{Group \code{epi_tibble} object}
\usage{
\method{group_by}{epi_tibble}(x, ...)
}
\description{
Groups an \code{epi_tibble} object by one or more variables, preserving class and
attributes.
}
