% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_df.R
\name{group_epi_df}
\alias{group_epi_df}
\title{Group an \code{epi_df} object by default keys}
\usage{
group_epi_df(x, exclude = character())
}
\arguments{
\item{x}{an \code{epi_df}}

\item{exclude}{character vector of column names to exclude from grouping}
}
\value{
a grouped \code{epi_df}
}
\description{
Group an \code{epi_df} object by default keys
}
