% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate.R
\name{growth_rate}
\alias{growth_rate}
\title{Estimate growth rate}
\usage{
growth_rate(
  x = seq_along(y),
  y,
  x0 = x,
  method = c("pct_change", "lin_reg", "smooth_spline", "trend_filter"),
  log_scale = FALSE,
  n = 14,
  align = c("center", "right", "left"),
  before,
  complete = FALSE,
  time_step,
  na_rm = FALSE,
  ...
)
}
\arguments{
\item{x}{Design points corresponding to the signal values \code{y}. Default is
\code{seq_along(y)} (that is, equally-spaced points from 1 to the length of
\code{y}).}

\item{y}{Signal values whose growth rates are of interest.}

\item{x0}{Points at which we should estimate the growth rate. Must be a
subset of \code{x} (no extrapolation allowed). Default is \code{x}.}

\item{method}{Either "pct_change", "lin_reg", "smooth_spline", or
"trend_filter", indicating the method to use for the growth rate
calculation. The first two are local methods: they are run in a sliding
fashion over the sequence (in order to estimate derivatives and hence
growth rates); the latter two are global methods: they are run once over
the entire sequence. See details for more explanation.}

\item{log_scale}{Should growth rates be estimated using the parametrization
on the log scale? See details for an explanation. Default is \code{FALSE}.}

\item{n}{Number of time steps to use in the sliding window, when \code{method} is
"pct_change" or "lin_reg".

a value on November 5, we apply the given function or formula to
data in between November 1 and 5. Default is 14.}

\item{na_rm}{Should missing values be removed before the computation? Default
is \code{FALSE}.}

\item{...}{Additional arguments to pass to the method used to estimate the
derivative. See details below.}
}
\value{
Vector of growth rate estimates at the specified points \code{x0}.
}
\description{
Estimates the growth rate of a signal at given points along the underlying
sequence. Several methodologies are available; see the \href{https://cmu-delphi.github.io/epiprocess/articles/growth_rate.html}{growth rate vignette}
for examples.
}
