% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_time_column_name}
\alias{guess_time_column_name}
\title{rename potential time_value columns}
\usage{
guess_time_column_name(x, substitutions = NULL)
}
\description{
rename potential time_value columns
}
\keyword{internal}
