% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditions.R
\name{handle_via_muffle_warning}
\alias{handle_via_muffle_warning}
\title{Condition handler muffling warnings, with a consistent identity}
\usage{
handle_via_muffle_warning(wrn)
}
\arguments{
\item{cnd}{a condition object (e.g., an error, warning, or message)}
}
\value{
output of \code{\link[base:conditions]{base::invokeRestart}} (likely/always \code{NULL})
}
\description{
This is a very simple function, but simply repeating its definition
everywhere of interest prevents
\code{\link[base:conditions]{base::globalCallingHandlers}}
and straightforward handler-related functions from recognizing these repeated
redefinitions as the same things (due to differing bytecode, srcrefs, closure
environments, and/or other things). This function provides a consistent
identity for this type of handler, and attempts to maintain this identity
even across package reloads (by setting its closure environment to a
something that should always be available and never reloaded; to match
\code{rlang:::hnd_entrace}, this is \code{\link[base:environment]{base::baseenv()}}; this means that things
not available from \code{base} require use of \verb{<pkg>::} to access).
}
\section{Side effects}{
 muffles a warning (invokes the \code{"muffleWarning"}
restart)
}

\seealso{
\code{\link[base:conditions]{base::withCallingHandlers}} to tie this handler to a condition
subclass for a block of code, or
\code{\link[base:conditions]{base::globalCallingHandlers}},
\code{\link{global_handlers_push}}, \code{\link{global_handlers_remove}},
\code{\link{global_handlers_list}} to enable/disable/list these associations globally

Other condition/calling handler functions: 
\code{\link{handle_via_dump_frames}()},
\code{\link{handle_via_entrace}()},
\code{\link{handle_via_muffle_message}()},
\code{\link{handle_via_recover}()}
}
\concept{condition/calling handler functions}
