% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_column_names}
\alias{id_column_names}
\title{potential id columns}
\usage{
id_column_names()
}
\description{
the full list of potential substitutions for the \code{id} column name:
id, case_id, event_id, uid, uuid, row_id, record_id, case_no, record_no, patient_id, subject_id, Id, Case_Id, Event_Id, Uid, Uuid, Row_Id, Record_Id, Case_No, Record_No, Patient_Id, Subject_Id
}
\keyword{internal}
