% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incidence_num_outlier_example}
\alias{incidence_num_outlier_example}
\title{Subset of JHU daily cases from California and Florida}
\format{
A tibble with 730 rows and 3 variables:
\describe{
\item{geo_value}{the geographic value associated with each row of measurements.}
\item{time_value}{the time value associated with each row of measurements.}
\item{cases}{Number of new confirmed COVID-19 cases, daily}
}
}
\source{
This object contains a modified part of the \href{https://github.com/CSSEGISandData/COVID-19}{COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University} as \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{republished in the COVIDcast Epidata API}. This data set is licensed under the terms of the
\href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0 International license}
by the Johns Hopkins University on behalf of its Center for Systems Science in Engineering.
Copyright Johns Hopkins University 2020.

Modifications:
\itemize{
\item \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{From the COVIDcast Epidata API}:
These signals are taken directly from the JHU CSSE \href{https://github.com/CSSEGISandData/COVID-19}{COVID-19 GitHub repository} without changes.
\item Furthermore, the data has been limited to a very small number of rows, the signal names slightly altered, and formatted into a tibble.
}
}
\usage{
incidence_num_outlier_example
}
\description{
This data source of confirmed COVID-19 cases
is based on reports made available by the Center for
Systems Science and Engineering at Johns Hopkins University.
This example data is a snapshot as of Oct 28, 2021 and captures the cases from June 1, 2020 to May 31, 2021
and is limited to California and Florida.
}
\keyword{datasets}
