% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\name{inclusive}
\alias{inclusive}
\title{Mark a threshold as an inclusive bound using a wrapper class}
\usage{
inclusive(x)
}
\description{
Mark a threshold as an inclusive bound using a wrapper class
}
\keyword{internal}
