% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tibble.R
\name{inner_join.epi_tibble}
\alias{inner_join.epi_tibble}
\alias{left_join.epi_tibble}
\alias{right_join.epi_tibble}
\alias{full_join.epi_tibble}
\title{Join two \code{epi_tibble} objects}
\usage{
\method{inner_join}{epi_tibble}(x, ...)

\method{left_join}{epi_tibble}(x, ...)

\method{right_join}{epi_tibble}(x, ...)

\method{full_join}{epi_tibble}(x, ...)
}
\description{
Joins two \code{epi_tibble} object by one or more variables, preserving class and
attributes.
}
