% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_deletion_column_names}
\alias{is_deletion_column_names}
\title{potential is_deletion columns}
\usage{
is_deletion_column_names()
}
\description{
the full list of potential substitutions for the \code{is_deletion} column name:
is_deletion, is_removal, is_deleted, deletion_flag, is_void, is_invalid, Is_Deletion, Is_Removal, Is_Deleted, Deletion_Flag, Is_Void, Is_Invalid
}
\keyword{internal}
