% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{is_epi_archive}
\alias{is_epi_archive}
\title{Test for \code{epi_archive} format}
\usage{
is_epi_archive(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from \code{epi_archive}.
}
\description{
Test for \code{epi_archive} format
}
