% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_df.R
\name{is_epi_df}
\alias{is_epi_df}
\title{Test for \code{epi_df} format}
\usage{
is_epi_df(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from \code{epi_df}.
}
\description{
Test for \code{epi_df} format
}
