% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{is_locf}
\alias{is_locf}
\title{Checks to see if a value in a vector is LOCF}
\usage{
is_locf(vec, abs_tol, is_key)
}
\description{
LOCF meaning last observation carried forward (to later
versions). Lags the vector by 1, then compares with itself. If \code{is_key} is
\code{TRUE}, only values that are exactly the same between the lagged and
original are considered LOCF. If \code{is_key} is \code{FALSE} and \code{vec} is a vector
of numbers (\code{\link[base:is.numeric]{base::is.numeric}}), then approximate equality will be used,
checking whether the absolute difference between each pair of entries is
\verb{<= abs_tol}; if \code{vec} is something else, then exact equality is used
instead.
}
\details{
We include epikey-time columns in LOCF comparisons as part of an optimization
to avoid slower grouped operations while still ensuring that the first
observation for each time series will not be marked as LOCF. We test these
key columns for exact equality to prevent chopping off consecutive
time_values during flat periods when \code{abs_tol} is high.

We use exact equality for non-\code{is.numeric} double/integer columns such as
dates, datetimes, difftimes, \code{tsibble::yearmonth}s, etc., as these may be
used as part of re-indexing or grouping procedures, and we don't want to
change the number of groups for those operations when we remove LOCF data
during compactification.
}
\keyword{internal}
