% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{data}
\name{jhu_confirmed_cumulative_num}
\alias{jhu_confirmed_cumulative_num}
\title{Subset of COVID-19 cumulative case counts from 4 states}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2808 rows and 14 columns.
}
\source{
This object contains a modified part of the \href{https://github.com/CSSEGISandData/COVID-19}{COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University} as \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{republished in the COVIDcast Epidata API}. This data set is licensed under the terms of the
\href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0 International license}
by the Johns Hopkins University on behalf of its Center for Systems Science in Engineering.
Copyright Johns Hopkins University 2020.
}
\usage{
jhu_confirmed_cumulative_num
}
\description{
Data set for 4 states containing COVID-19 Cumulative Cases as reported by
JHU-CSSE and downloaded from the CMU Delphi COVIDcast Epidata API.
This example data is a snapshot as of March 20, 2024, and
ranges from March 1, 2020 to January 31, 2022. It is limited
to California, Florida, New York, and Texas.

It is used in the {epiprocess} "Getting Started" vignette.
}
\section{Data dictionary}{


The data has columns:
\describe{
\item{geo_value}{the geographic value associated with each row of measurements.}
\item{signal}{name of metric, derived from upstream data.}
\item{source}{name of upstream data source.}
\item{geo_type}{spatial resolution of the signal.}
\item{time_type}{temporal resolution of the signal.}
\item{time_value}{the time value associated with each row of measurements.}
\item{issue}{time unit (e.g., date) when the signal data were published.}
\item{lag}{time delta (e.g. days) between when the underlying events happened and when the data were published.}
\item{missing_value}{an integer code that is zero when the value field is present and non-zero when the data is missing (see \href{https://cmu-delphi.github.io/delphi-epidata/api/missing_codes.html}{missing codes}).}
\item{missing_stderr}{an integer code that is zero when the stderr field is present and non-zero when the data is missing (see \href{https://cmu-delphi.github.io/delphi-epidata/api/missing_codes.html}{missing codes}).}
\item{missing_sample_size}{an integer code that is zero when the sample_size field is present and non-zero when the data is missing (see \href{https://cmu-delphi.github.io/delphi-epidata/api/missing_codes.html}{missing codes}).}
\item{value}{cumulative number of confirmed COVID-19 cases, derived from the underlying data source.}
\item{stderr}{approximate standard error of the statistic with respect to its sampling distribution, NA when not applicable.}
\item{sample_size}{number of “data points” used in computing the statistic, NA when not applicable.}
}

}

\examples{
# Since this is a re-exported dataset, it cannot be loaded using
# the `data()` function. `data()` looks for a file of the same name
# in the `data/` directory, which doesn't exist in this package.
# works
epiprocess::jhu_confirmed_cumulative_num

# works
library(epiprocess)
jhu_confirmed_cumulative_num

# fails
data(jhu_confirmed_cumulative_num, package = "epiprocess")
}
\keyword{datasets}
