% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jhu_csse_county_level_subset}
\alias{jhu_csse_county_level_subset}
\title{Subset of JHU daily cases from counties in Massachusetts and Vermont}
\format{
A tibble with 16,212 rows and 5 variables:
\describe{
\item{geo_value}{the geographic value associated with each row of measurements.}
\item{time_value}{the time value associated with each row of measurements.}
\item{cases}{Number of new confirmed COVID-19 cases, daily}
\item{county_name}{the name of the county}
\item{state_name}{the full name of the state}
}
}
\source{
This object contains a modified part of the \href{https://github.com/CSSEGISandData/COVID-19}{COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University} as \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{republished in the COVIDcast Epidata API}. This data set is licensed under the terms of the
\href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0 International license}
by the Johns Hopkins University on behalf of its Center for Systems Science in Engineering.
Copyright Johns Hopkins University 2020.

Modifications:
\itemize{
\item \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{From the COVIDcast Epidata API}: These signals are taken directly from the JHU CSSE \href{https://github.com/CSSEGISandData/COVID-19}{COVID-19 GitHub repository} without changes. The 7-day average signals are computed by Delphi by calculating moving averages of the preceding 7 days, so the signal for June 7 is the average of the underlying data for June 1 through 7, inclusive.
\item Furthermore, the data has been limited to a very small number of rows, the signal names slightly altered, and formatted into a tibble.
}
}
\usage{
jhu_csse_county_level_subset
}
\description{
This data source of confirmed COVID-19 cases and deaths
is based on reports made available by the Center for
Systems Science and Engineering at Johns Hopkins University.
This example data ranges from Mar 1, 2020 to Dec 31, 2021,
and is limited to Massachusetts and Vermont.
}
\keyword{datasets}
