% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_colnames.R
\name{key_colnames}
\alias{key_colnames}
\title{Grab any keys associated to an epi_df}
\usage{
key_colnames(x, ...)
}
\arguments{
\item{x}{a data.frame, tibble, or epi_df}

\item{...}{additional arguments passed on to methods}
}
\value{
If an \code{epi_df}, this returns all "keys". Otherwise \code{NULL}
}
\description{
Grab any keys associated to an epi_df
}
\keyword{internal}
