% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_analysis.R
\name{lag_within_x_latest}
\alias{lag_within_x_latest}
\title{pull the value from lags when values starts indefinitely being within prop of its latest value.}
\usage{
lag_within_x_latest(lags, values, prop = 0.2)
}
\arguments{
\item{lags}{vector of lags; should be sorted}

\item{values}{this should be a vector (e.g., a column) with length matching that of \code{lags}}

\item{prop}{optional length-1 double; proportion}
}
\description{
pull the value from lags when values starts indefinitely being within prop of its latest value.
}
\keyword{internal}
