% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_analysis.R
\name{max_no_na}
\alias{max_no_na}
\title{the default behavior returns a warning on empty lists, which we do not want,
and there is no super clean way of preventing this}
\usage{
max_no_na(x)
}
\description{
the default behavior returns a warning on empty lists, which we do not want,
and there is no super clean way of preventing this
}
\keyword{internal}
