% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R, R/archive_new.R
\name{max_version_with_row_in}
\alias{max_version_with_row_in}
\title{\code{max(x$version)}, with error if \code{x} has 0 rows}
\usage{
max_version_with_row_in(x)

max_version_with_row_in(x)
}
\arguments{
\item{x}{\code{x} argument of \code{\link{as_epi_archive}}}
}
\value{
\code{max(x$version)} if it has any rows; raises error if it has 0 rows or
an \code{NA} version value

\code{max(x$version)} if it has any rows; raises error if it has 0 rows or
an \code{NA} version value
}
\description{
Exported to make defaults more easily copyable.

Exported to make defaults more easily copyable.
}
