% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_new.R
\name{merge_epi_archive2}
\alias{merge_epi_archive2}
\title{Merge epi archive}
\usage{
merge_epi_archive2(
  x,
  y,
  sync = c("forbid", "na", "locf", "truncate"),
  compactify = TRUE
)
}
\arguments{
\item{x}{as in \code{\link{epix_merge}}}

\item{y}{as in \code{\link{epix_merge}}}

\item{sync}{as in \code{\link{epix_merge}}}

\item{compactify}{as in \code{\link{epix_merge}}}
}
\description{
Merges another \code{epi_archive} with the current one, mutating the
current one by reseating its \code{DT} and several other fields, but avoiding
mutation of the old \code{DT}; returns the current archive
\link[base:invisible]{invisibly}. See \code{\link{epix_merge}} for a full description
of the non-R6-method version, which does not mutate either archive, and
does not alias either archive's \code{DT}.a
}
