% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Min}
\alias{Min}
\alias{Max}
\alias{Sum}
\alias{Mean}
\alias{Median}
\title{Computations with \code{NA} values removed}
\usage{
Min(x)

Max(x)

Sum(x)

Mean(x)

Median(x)
}
\description{
These are just wrapper functions for common computational utilities with \code{NA}
values removed by default.
}
