% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_new}
\alias{mutate_new}
\title{Like \code{mutate}, but refusing to overwrite or delete existing columns}
\usage{
mutate_new(.data, ...)
}
\arguments{
\item{.data}{data frame}
}
\description{
Like \code{mutate}, but refusing to overwrite or delete existing columns
}
\keyword{internal}
