% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\name{new_bound0}
\alias{new_bound0}
\title{Make a new epiprocess_bound wrapper object, with no input validation}
\usage{
new_bound0(threshold, inclusive)
}
\description{
Make a new epiprocess_bound wrapper object, with no input validation
}
\keyword{internal}
