% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_new.R
\name{new_epi_archive2}
\alias{new_epi_archive2}
\title{New epi archive}
\usage{
new_epi_archive2(
  x,
  geo_type = NULL,
  time_type = NULL,
  other_keys = NULL,
  additional_metadata = NULL,
  compactify = NULL,
  clobberable_versions_start = NA,
  versions_end = NULL
)
}
\arguments{
\item{x}{A data.frame, data.table, or tibble, with columns \code{geo_value},
\code{time_value}, \code{version}, and then any additional number of columns.}

\item{geo_type}{Type for the geo values. If missing, then the function will
attempt to infer it from the geo values present; if this fails, then it
will be set to "custom".}

\item{time_type}{Type for the time values. If missing, then the function will
attempt to infer it from the time values present; if this fails, then it
will be set to "custom".}

\item{other_keys}{Character vector specifying the names of variables in \code{x}
that should be considered key variables (in the language of \code{data.table})
apart from "geo_value", "time_value", and "version".}

\item{additional_metadata}{List of additional metadata to attach to the
\code{epi_archive} object. The metadata will have \code{geo_type} and \code{time_type}
fields; named entries from the passed list or will be included as well.}

\item{compactify}{Optional; Boolean or \code{NULL}: should we remove rows that are
considered redundant for the purposes of \code{epi_archive}'s built-in methods
such as \code{as_of}? As these methods use the last version of each observation
carried forward (LOCF) to interpolate between the version data provided,
rows that don't change these LOCF results can potentially be omitted to
save space while maintaining the same behavior (with the help of the
\code{clobberable_versions_start} and \code{versions_end} fields in some edge cases).
\code{TRUE} will remove these rows, \code{FALSE} will not, and missing or \code{NULL} will
remove these rows and issue a warning. Generally, this can be set to
\code{TRUE}, but if you directly inspect or edit the fields of the \code{epi_archive}
such as its \code{DT}, or rely on redundant updates to achieve a certain
behavior of the \code{ref_time_values} default in \code{epix_slide}, you will have to
determine whether \code{compactify=TRUE} will produce the desired results. If
compactification here is removing a large proportion of the rows, this may
indicate a potential for space, time, or bandwidth savings upstream the
data pipeline, e.g., by avoiding fetching, storing, or processing these
rows of \code{x}.}

\item{clobberable_versions_start}{Optional; as in \code{\link{as_epi_archive}}}

\item{versions_end}{Optional; as in \code{\link{as_epi_archive}}}
}
\value{
An \code{epi_archive} object.
}
\description{
Creates a new \code{epi_archive} object.
}
\details{
Refer to the documentation for \code{\link[=as_epi_archive]{as_epi_archive()}} for more information
and examples of parameter names.
}
