% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_df.R
\name{new_epi_df}
\alias{new_epi_df}
\title{Creates an \code{epi_df} object}
\usage{
new_epi_df(
  x = tibble::tibble(),
  geo_type,
  time_type,
  as_of,
  additional_metadata = list(),
  ...
)
}
\arguments{
\item{x}{A data.frame, \link[tibble:tibble]{tibble::tibble}, or \link[tsibble:tsibble]{tsibble::tsibble} to be converted}

\item{geo_type}{Type for the geo values. If missing, then the function will
attempt to infer it from the geo values present; if this fails, then it
will be set to "custom".}

\item{time_type}{Type for the time values. If missing, then the function will
attempt to infer it from the time values present; if this fails, then it
will be set to "custom".}

\item{as_of}{Time value representing the time at which the given data were
available. For example, if \code{as_of} is January 31, 2022, then the \code{epi_df}
object that is created would represent the most up-to-date version of the
data available as of January 31, 2022. If the \code{as_of} argument is missing,
then the current day-time will be used.}

\item{additional_metadata}{List of additional metadata to attach to the
\code{epi_df} object. The metadata will have \code{geo_type}, \code{time_type}, and
\code{as_of} fields; named entries from the passed list will be included as
well. If your tibble has additional keys, be sure to specify them as a
character vector in the \code{other_keys} component of \code{additional_metadata}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An \code{epi_df} object.
}
\description{
Creates a new \code{epi_df} object. By default, builds an empty tibble with the
correct metadata for an \code{epi_df} object (ie. \code{geo_type}, \code{time_type}, and \code{as_of}).
Refer to the below info. about the arguments for more details.
}
