% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{new_struct}
\alias{new_struct}
\title{List that requires \code{$}, \verb{$<-}, etc., to provide exact matches to names}
\usage{
new_struct(fields)
}
\description{
Does not validate or maintain invariant that field names are unique.
}
\keyword{internal}
