% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R, R/archive_new.R
\name{next_after}
\alias{next_after}
\title{Get the next possible value greater than \code{x} of the same type}
\usage{
next_after(x)

next_after(x)
}
\arguments{
\item{x}{the starting "value"(s)}
}
\value{
same class, typeof, and length as \code{x}

same class, typeof, and length as \code{x}
}
\description{
Get the next possible value greater than \code{x} of the same type

Get the next possible value greater than \code{x} of the same type
}
