% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_analysis.R
\name{num_percent}
\alias{num_percent}
\title{simple util for printing a fraction and it's percent}
\usage{
num_percent(a, b)
}
\description{
simple util for printing a fraction and it's percent
}
\keyword{internal}
