% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblish.R
\name{obj_is_tblish}
\alias{obj_is_tblish}
\title{Is \code{obj} a (subclass of) "data.frame", "tbl", or anything similar?}
\usage{
obj_is_tblish(x)
}
\description{
Is \code{obj} a (subclass of) "data.frame", "tbl", or anything similar?
}
\keyword{internal}
