% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting-parts.R
\name{partial_chr_mapping_standardize}
\alias{partial_chr_mapping_standardize}
\title{Standardize a partial names->values mapping}
\usage{
partial_chr_mapping_standardize(
  partial_mapping,
  default_full_mapping,
  chr_keys,
  partial_mapping_arg = rlang::caller_arg(partial_mapping),
  call = rlang::caller_env()
)
}
\description{
Standardize a partial names->values mapping
}
\examples{

partial_chr_mapping_standardize(c(target = 1), 0, c("target", "aux1", "aux2"))

partial_chr_mapping_standardize(c(target = 1), c(target = 0, aux1 = 0, aux2 = 0), c("target", "aux1", "aux2"))

}
\keyword{internal}
