% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_lines}
\alias{paste_lines}
\title{Paste \code{chr} entries (lines) together with \code{"\\n"} separators, trailing \code{"\\n"}}
\usage{
paste_lines(lines)
}
\arguments{
\item{lines}{\code{chr}}
}
\value{
string
}
\description{
Paste \code{chr} entries (lines) together with \code{"\\n"} separators, trailing \code{"\\n"}
}
\keyword{internal}
