% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap for an epi_df}
\usage{
plot_heatmap(x, fill = NULL, ..., .max_keys = 60)
}
\arguments{
\item{x}{An \code{epi_df} object.}

\item{fill}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> The column to use for the fill aesthetic.
If NULL, the first numeric non-key column is used.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}.}

\item{.max_keys}{Maximum number of key combinations to display. If the data
contains more key combinations than this limit, a random sample of size
\code{.max_keys} is displayed, and a warning is issued. Set to \code{Inf} to
display all keys. Does not apply if \code{interactive = TRUE}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot a heatmap for an epi_df
}
