% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{plot_outliers}
\alias{plot_outliers}
\title{Plot outlier detection bands and/or points identified as outliers}
\usage{
plot_outliers(
  x,
  var,
  outliers_col = outliers,
  include_bands = FALSE,
  include_points = TRUE,
  combined_only = TRUE,
  facet_vars = vars(geo_value),
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_scales = "fixed",
  plot = TRUE
)
}
\arguments{
\item{x}{An \code{epi_tibble} object including results for outlier detection as
returned by \code{detect_outliers}}

\item{var}{The variable in \code{x} in which to look for outliers.}

\item{include_bands}{Boolean; if \code{TRUE}, outlier detection thresholds are
shown}

\item{include_points}{Boolean; if \code{TRUE}, outliers are highlighted with
points}

\item{combined_only}{Boolean; if \code{TRUE}, only the combined results across
all detection methods are displayed}

\item{facet_vars}{character vector of variables to facet by.
Default is \code{"geo_type"}. If \code{NULL}, no facetting is done.}

\item{facet_nrow, }{facet_ncol Integer number of rows and columns for
facetting.}

\item{plot}{Boolean; if \code{TRUE}, the plot is printed.}

\item{outliers_col_name}{The variable in \code{x} with outlier detection results.
This is the value used as the \code{new_col_name} argument to \code{detect_outliers}.}
}
\value{
invisibly, a \code{ggplot} object
}
\description{
Plot outlier detection bands and/or points identified as outliers
}
