% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_new.R
\name{print.epi_archive2}
\alias{print.epi_archive2}
\title{Print information about an \code{epi_archive} object}
\usage{
\method{print}{epi_archive2}(epi_archive, class = TRUE, methods = TRUE)
}
\arguments{
\item{class}{Boolean; whether to print the class label header}

\item{methods}{Boolean; whether to print all available methods of
the archive}
}
\description{
Print information about an \code{epi_archive} object
}
