% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_signal.R
\name{print.epi_signal}
\alias{print.epi_signal}
\title{Print \code{epi_signal} object}
\usage{
\method{print}{epi_signal}(x, ...)
}
\arguments{
\item{x}{The \code{epi_signal} object.}

\item{...}{Additional arguments passed to \code{print.tibble()} to print the
data.}
}
\value{
The \code{epi_signal} object, unchanged.
}
\description{
Prints a brief summary of the signal, then prints the underlying data frame
(tibble), for an \code{epi_signal} object.
}
