% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tibble.R
\name{print.epi_tibble}
\alias{print.epi_tibble}
\title{Print \code{epi_tibble} object}
\usage{
\method{print}{epi_tibble}(x, ...)
}
\arguments{
\item{x}{The \code{epi_tibble} object.}

\item{...}{Additional arguments passed to \code{print.tibble()} to print the
data.}
}
\value{
The \code{epi_tibble} object, unchanged.
}
\description{
Prints a brief summary of the \code{epi_tibble} object, then prints the underlying
tibble.
}
