% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet}
\alias{quiet}
\title{Force \code{x} with regular console output silenced; return invisibly}
\usage{
quiet(x)
}
\arguments{
\item{x}{argument to \code{\link{force}}}
}
\value{
result, invisibly
}
\description{
Does not silence the message stream (which is normally \code{stderr()}).
}
\keyword{internal}
