% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{select_substr_prefix_columns}
\alias{select_substr_prefix_columns}
\title{Select columns named with a given prefix and remove the prefix}
\usage{
select_substr_prefix_columns(x, prefix, ...)
}
\arguments{
\item{x}{an object with a structure like a data frame}

\item{prefix}{the column name prefix to select on and to remove}
}
\value{
the same type of structure with the selected renamed columns
}
\description{
Select columns named with a given prefix and remove the prefix
}
\examples{

select_substr_prefix_columns(data.table::data.table(prefix.a = 1:5, prefix.b=11:15, nonprefixed=letters[1:5]), "prefix.")

}
