% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{set_time_type0}
\alias{set_time_type0}
\title{Set, without validation, the \code{time_type} of an \verb{epi_*} object}
\usage{
set_time_type0(x, value)
}
\arguments{
\item{x}{an \verb{epi_*} object}
}
\description{
Set, without validation, the \code{time_type} of an \verb{epi_*} object
}
\keyword{internal}
