% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_archive.R
\name{set_versions_end}
\alias{set_versions_end}
\title{Set the \code{versions_end} attribute of an \code{epi_archive} object}
\usage{
set_versions_end(x, versions_end)
}
\arguments{
\item{x}{An \code{epi_archive} object}

\item{versions_end}{The new \code{versions_end} value}
}
\value{
An \code{epi_archive} object with the updated \code{versions_end} attribute
}
\description{
An escape hatch for epix_as_of, which does not allow version >
\verb{$versions_end}.
}
