% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_epi_archive.R
\name{slide.grouped_epi_archive}
\alias{slide.grouped_epi_archive}
\title{Slide over grouped epi archive}
\usage{
\method{slide}{grouped_epi_archive}(
  grouped_epi_archive,
  f,
  ...,
  before,
  ref_time_values,
  time_step,
  new_col_name = "slide_value",
  as_list_col = FALSE,
  names_sep = "_",
  all_versions = FALSE
)
}
\description{
Slides a given function over variables in a \code{grouped_epi_archive}
object. See the documentation for the wrapper function \code{\link[=epix_slide]{epix_slide()}} for
details.
}
