% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_analysis.R
\name{spread_vec}
\alias{spread_vec}
\title{the default behavior returns a warning on empty lists, which we do not want}
\usage{
spread_vec(x)
}
\description{
the default behavior returns a warning on empty lists, which we do not want
}
\keyword{internal}
