% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Start}
\alias{Start}
\alias{Middle}
\alias{End}
\title{Start, middle, and end extrators}
\usage{
Start(x)

Middle(x, floor = TRUE)

End(x)
}
\description{
These are just convenience functions for extracting the start, middle, and
end of a sequence.
}
