% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting-parts.R
\name{str.pipeline_segment}
\alias{str.pipeline_segment}
\title{\code{str} impl for pipeline segments}
\usage{
\method{str}{pipeline_segment}(
  object,
  ...,
  nest.lev = 0,
  indent.str = paste(rep.int(" ", max(0, nest.lev + 1)), collapse = "..")
)
}
\description{
\code{str} impl for pipeline segments
}
\examples{
transform_target_segment("y", log, exp, engine_segment(parsnip::linear_reg())) \%>\% str()

}
\keyword{internal}
