% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_df.R
\name{summary.epi_df}
\alias{summary.epi_df}
\title{Summarize \code{epi_df} object}
\usage{
\method{summary}{epi_df}(object, ...)
}
\arguments{
\item{object}{an \code{epi_df}}

\item{...}{Additional arguments, for compatibility with \code{summary()}.
Currently unused.}
}
\description{
Prints a variety of summary statistics about the \code{epi_df} object, such as
the time range included and geographic coverage.
}
