% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_signal.R
\name{summary.epi_signal}
\alias{summary.epi_signal}
\title{Summarize \code{epi_signal} object}
\usage{
\method{summary}{epi_signal}(object, ...)
}
\arguments{
\item{object}{The \code{epi_signal} object.}

\item{...}{Additional arguments, for compatibility with \code{summary()}.
Currently unused.}
}
\value{
No return value; called only to print summary statistics.
}
\description{
Prints a variety of summary statistics about the underlying data, such as the
time range included and geographic coverage, for an \code{epi_signal} object.
}
