% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tibble.R
\name{summary.epi_tibble}
\alias{summary.epi_tibble}
\title{Summarize \code{epi_tibble} object}
\usage{
\method{summary}{epi_tibble}(object, ...)
}
\arguments{
\item{object}{The \code{epi_tibble} object.}

\item{...}{Additional arguments, for compatibility with \code{summary()}.
Currently unused.}
}
\value{
No return value; called only to print summary statistics.
}
\description{
Prints a variety of summary statistics about the \code{epi_tibble} object, such as
the time range included and geographic coverage.
}
