% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{tbl_diff2}
\alias{tbl_diff2}
\title{Calculate compact patch to move from one snapshot/update to another}
\usage{
tbl_diff2(
  earlier_snapshot,
  later_tbl,
  ukey_names,
  later_format = c("snapshot", "update"),
  compactify_abs_tol = 0
)
}
\arguments{
\item{earlier_snapshot}{tibble or \code{NULL}; \code{NULL} represents that there was no
data before \code{later_tbl}.}

\item{later_tbl}{tibble; must have the same column names as
\code{earlier_snapshot} if it is a tibble.}

\item{ukey_names}{character; column names that together, form a unique key
for \code{earlier_snapshot} and for \code{later_tbl}. This is unchecked; see
\code{\link{check_ukey_unique}} if you don't already have this guaranteed.}

\item{later_format}{"snapshot" or "update"; default is "snapshot". If
"snapshot", \code{later_tbl} will be interpreted as a full snapshot of the data
set including all ukeys, and any ukeys that are in \code{earlier_snapshot} but
not in \code{later_tbl} are interpreted as deletions, which are currently
(imprecisely) represented in the output patch as revisions of all
non-\code{ukey_names} columns to NA values (using \code{{vctrs}}). If "update", then
it's assumed that any deletions have already been represented this way in
\code{later_tbl} and any ukeys not in \code{later_tbl} are simply unchanged; we are
just ensuring that the update is fully compact for the given
\code{compactify_abs_tol}.}

\item{compactify_abs_tol}{compactification tolerance; see \code{apply_compactify}}
}
\value{
a tibble in compact "update" (diff) format
}
\description{
Calculate compact patch to move from one snapshot/update to another
}
\keyword{internal}
