% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{tbl_fast_anti_join}
\alias{tbl_fast_anti_join}
\title{Variation on \code{\link[dplyr:filter-joins]{dplyr::anti_join}} for speed + tolerance setting}
\usage{
tbl_fast_anti_join(x, y, ukey_names, val_names, abs_tol = 0)
}
\arguments{
\item{x}{tibble; \code{x[ukey_names]} must not have any duplicate rows}

\item{y}{tibble; \code{y[ukey_names]} must not have any duplicate rows}

\item{ukey_names}{chr; names of columns that form a unique key, for \code{x} and
for \code{y}}

\item{val_names}{chr; names of columns which should be treated as
value/measurement columns, and compared with a tolerance}

\item{abs_tol}{scalar non-negative numeric; absolute tolerance with which to
compare value columns; see \code{\link{vec_approx_equal}}}
}
\value{
rows from \code{x} that either (a) don't have a (0-tolerance) matching
ukey in \code{y}, or (b) have a matching ukey in \code{y}, but don't have
approximately equal value column values
}
\description{
Variation on \code{\link[dplyr:filter-joins]{dplyr::anti_join}} for speed + tolerance setting
}
\keyword{internal}
