% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{tbl_patch}
\alias{tbl_patch}
\title{Apply an update (e.g., from \code{tbl_diff2}) to a snapshot}
\usage{
tbl_patch(snapshot, update, ukey_names)
}
\arguments{
\item{snapshot}{tibble or \code{NULL}; entire data set as of some version, or
\code{NULL} to treat \code{update} as the initial version of the data set.}

\item{update}{tibble; ukeys + initial values for added rows, ukeys + new
values for changed rows. Deletions must be imprecisely represented as
changing all values to NAs.}

\item{ukey_names}{character; names of columns that should form a unique key
for \code{snapshot} and for \code{update}. Uniqueness is unchecked; if you don't have
this guaranteed, see \code{\link[=check_ukey_unique]{check_ukey_unique()}}.}
}
\value{
tibble; snapshot of the data set with the update applied.
}
\description{
Apply an update (e.g., from \code{tbl_diff2}) to a snapshot
}
\keyword{internal}
