% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{time_delta_standardize}
\alias{time_delta_standardize}
\title{Standardize time_deltas to a multiple of \code{\link[=unit_time_delta]{unit_time_delta()}}}
\usage{
time_delta_standardize(time_delta, time_type, format = c("friendly", "fast"))
}
\description{
Standardize time_deltas to a multiple of \code{\link[=unit_time_delta]{unit_time_delta()}}
}
\keyword{internal}
