% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_analysis.R
\name{time_delta_summary}
\alias{time_delta_summary}
\title{Like \code{summary} but working across all "time deltas", including difftimes}
\usage{
time_delta_summary(time_delta, time_type = NULL)
}
\description{
Also standardizes units of difftimes to the natural unit for the given
\code{time_type} (via conversion to and from a corresponding number of time
steps).
}
\keyword{internal}
