% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{time_delta_to_n_steps}
\alias{time_delta_to_n_steps}
\title{Convert a time delta to a integerish number of "unit" steps between time values}
\usage{
time_delta_to_n_steps(time_delta, time_type, require_integer = TRUE)
}
\arguments{
\item{time_delta}{a vector that can be added to time values of time type
\code{time_type} to arrive at other time values of that time type, or
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} such a vector with Inf/-Inf entries mixed
in, if supported by the class of \code{time_delta}, even if \code{time_type} doesn't
necessarily support Inf/-Inf entries. Basically a slide window arg but
without sign and length restrictions.}

\item{time_type}{as in \code{validate_slide_window_arg}}
}
\value{
\link[rlang:is_integerish]{bare integerish} vector (with possible
infinite values) that produces the same result as \code{time_delta} when
multiplied by the natural \code{\link{unit_time_delta}} for
that time type and added to time values of time type \code{time_type}. If the
given time type does not support infinite values, then it should produce
+Inf or -Inf for analogous entries of \code{time_delta}, and match the addition
result match the addition result for non-infinite entries.
}
\description{
Convert a time delta to a integerish number of "unit" steps between time values
}
\keyword{internal}
