% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{time_minus_time_in_n_steps}
\alias{time_minus_time_in_n_steps}
\title{Difference between two time value vectors in terms of number of time "steps"}
\usage{
time_minus_time_in_n_steps(x, y, time_type, require_integer = TRUE)
}
\arguments{
\item{x}{a time_value (vector) of time type \code{time_type}}

\item{y}{a time_value (vector) of time type \code{time_type}}

\item{time_type}{as in \code{\link[=validate_slide_window_arg]{validate_slide_window_arg()}}}
}
\value{
integerish vector such that \code{x + n_steps_to_time_delta_fast(result)}
should equal \code{y}.
}
\description{
Difference between two time value vectors in terms of number of time "steps"
}
\keyword{internal}
