% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{time_plus_slide_window_arg}
\alias{time_plus_slide_window_arg}
\alias{time_minus_slide_window_arg}
\title{Advance/retreat time_value(s) by specified amount (slide window arg)}
\usage{
time_plus_slide_window_arg(x, y, time_type, max_time_value)

time_minus_slide_window_arg(x, y, time_type, min_time_value)
}
\arguments{
\item{x}{a time_value (vector) of time type \code{time_type}}

\item{y}{a (scalar) slide window arg; should pass \code{\link[=validate_slide_window_arg]{validate_slide_window_arg()}}}

\item{time_type}{as in \code{\link[=validate_slide_window_arg]{validate_slide_window_arg()}}}

\item{max_time_value}{when \code{y == Inf}, what should be the result of adding \code{y}?}

\item{min_time_value}{when \code{y == Inf}, what should be the result of subtracting \code{y}?}
}
\value{
a time_value (vector) of time type \code{time_type}
}
\description{
Advance/retreat time_value(s) by specified amount (slide window arg)
}
\seealso{
\code{\link{time_plus_n_steps}}, if you're working with an integerish vector
number of time steps \code{y} (output from other \verb{*n_steps} functions) instead.
}
\keyword{internal}
