% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide-refactor.R
\name{time_slide_to_simple_hop}
\alias{time_slide_to_simple_hop}
\title{Convert time slide function to a simple hop function}
\usage{
time_slide_to_simple_hop(.slide_comp, ..., .before_n_steps, .after_n_steps)
}
\description{
Convert time slide function to a simple hop function
}
\examples{
time_slide_to_simple_hop(as_time_slide_computation(~ .x[1L, ]),
  .before_n_steps = 2L, .after_n_steps = 0L
)(
  tibble(time_value = 1:5, value = 1:5),
  tibble(geo_value = 1),
  3:4
)

}
\keyword{internal}
