% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{time_type}
\alias{time_type}
\title{Get \code{time_type} of an \verb{epi_*} object}
\usage{
time_type(x)
}
\arguments{
\item{x}{an \verb{epi_*} object}
}
\description{
Get \code{time_type} of an \verb{epi_*} object
}
\keyword{internal}
