% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{time_type_unit_abbr}
\alias{time_type_unit_abbr}
\title{Get an abbreviation for the "units" of \code{unit_time_delta(time_type)}}
\usage{
time_type_unit_abbr(time_type)
}
\arguments{
\item{time_type}{str}
}
\value{
str
}
\description{
For use in formatting or automatically naming things based on
\code{time_delta_to_n_steps(time_delta)} for a \code{time_delta} between times of time
type \code{time_type}.
}
\keyword{internal}
