% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\docType{data}
\name{time_type_unit_abbrs}
\alias{time_type_unit_abbrs}
\title{Helper data for \code{\link{time_type_unit_abbr}}}
\format{
An object of class \code{character} of length 4.
}
\usage{
time_type_unit_abbrs
}
\description{
Helper data for \code{\link{time_type_unit_abbr}}
}
\keyword{internal}
