% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_analysis.R
\name{time_within_x_latest}
\alias{time_within_x_latest}
\title{pull the value from lags when values starts indefinitely being within prop of it's last value.}
\usage{
time_within_x_latest(lags, values, prop = 0.2)
}
\arguments{
\item{values}{this should be a 1 column tibble. errors may occur otherwise}
}
\description{
pull the value from lags when values starts indefinitely being within prop of it's last value.
}
\keyword{internal}
