% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{TransitionDTDiffer}
\alias{TransitionDTDiffer}
\title{Object used to produce diffs between \code{data.table}s with certain columns treated similar to keys}
\description{
Object used to produce diffs between \code{data.table}s with certain columns treated similar to keys

Object used to produce diffs between \code{data.table}s with certain columns treated similar to keys
}
\examples{

library("pipeR")

DT1 = data.table::data.table(t = 1:5, v = 1:5)
DT2 = data.table::data.table(t = 2:6, v = c(2:4, 5.1, 6))

TransitionDTDiffer$new("t")$diff("ver1", DT1, "ver2", DT2)

TransitionDTDiffer$new("t")$diff("ver1", DT1, "ver2", DT2)$patch_snapshot("ver1", DT1) \%>>\% all.equal(DT2)

}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TransitionDTDiffer$new()}}
\item \href{#method-diff}{\code{TransitionDTDiffer$diff()}}
\item \href{#method-clone}{\code{TransitionDTDiffer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionDTDiffer$new(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-diff"></a>}}
\if{latex}{\out{\hypertarget{method-diff}{}}}
\subsection{Method \code{diff()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionDTDiffer$diff(from.version.id, from.DT, to.version.id, to.DT)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionDTDiffer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
