% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{truncate_versions_after.epi_archive}
\alias{truncate_versions_after.epi_archive}
\title{Truncate versions after}
\usage{
\method{truncate_versions_after}{epi_archive}(epi_archive, max_version)
}
\arguments{
\item{epi_archive}{as in \code{\link{epix_truncate_versions_after}}}

\item{max_version}{as in \code{\link{epix_truncate_versions_after}}}
}
\description{
Filter to keep only older versions, mutating the archive by
potentially reseating but not mutating some fields. \code{DT} is likely, but not
guaranteed, to be copied. Returns the mutated archive
\link[base:invisible]{invisibly}.
}
