% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_archive_new.R
\name{truncate_versions_after.grouped_epi_archive2}
\alias{truncate_versions_after.grouped_epi_archive2}
\title{Truncate versions after a given version, grouped}
\usage{
\method{truncate_versions_after}{grouped_epi_archive2}(grouped_epi_archive, max_version)
}
\arguments{
\item{max_version}{as in \code{\link{epix_truncate_versions_after}}}

\item{x}{as in \code{\link{epix_truncate_versions_after}}}
}
\description{
Filter to keep only older versions by mutating the underlying
\code{epi_archive} using \verb{$truncate_versions_after}. Returns the mutated
\code{grouped_epi_archive} \link[base:invisible]{invisibly}.
}
