% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{unit_time_delta}
\alias{unit_time_delta}
\title{Object that, added to time_values of time_type, advances by one time step/interval}
\usage{
unit_time_delta(time_type, format = c("friendly", "fast"))
}
\arguments{
\item{time_type}{string; \code{epi_df}'s or \code{epi_archive}'s \code{time_type}}

\item{format}{"friendly" or "fast"; for some time_types, there are multiple
ways to represent time_deltas. "friendly" tries to output a format that
will be more informative when printed, and produce errors in more cases
when used in unexpected ways. "fast" tries to output a time_delta that will
be faster in downstream operations.}
}
\value{
an object \code{u} such that \code{time_values + u} represents advancing by one
time step / moving to the subsequent time interval for any \code{time_values}
object of time type \code{time_type}, and such that \code{time_values + k * u} for
integerish vector \code{k} advances by \code{k} steps (with vectorization,
recycling). At time of writing, these objects also all support
multiplication by nonintegerish numeric vectors, \code{mean}, and \code{median},
which are useful for summarizing vector time_deltas, but these fractional
time_deltas are not allowed in time_delta-specific operations.
}
\description{
Object that, added to time_values of time_type, advances by one time step/interval
}
\keyword{internal}
