% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unwrap}
\alias{unwrap}
\title{Extract singular element of a length-1 unnamed list (validated)}
\usage{
unwrap(x)
}
\arguments{
\item{x}{a length-1 list}
}
\value{
x[\link{1L}], if x actually was a length-1 list; error otherwise
}
\description{
Inverse of \code{list(elt)}.
}
\keyword{internal}
