% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{upcase_snake_case}
\alias{upcase_snake_case}
\title{given a vector of characters, add the same values, but upcased, e.g.
"date" -> c("date", "Date")
"target_date" -> c("target_date", "Target_Date")}
\usage{
upcase_snake_case(x)
}
\description{
given a vector of characters, add the same values, but upcased, e.g.
"date" -> c("date", "Date")
"target_date" -> c("target_date", "Target_Date")
}
\keyword{internal}
