% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{upstream_slide_f_info}
\alias{upstream_slide_f_info}
\title{Validate & get information about an upstream slide function}
\usage{
upstream_slide_f_info(.f)
}
\arguments{
\item{.f}{function such as \code{data.table::frollmean} or \code{slider::slide_mean};
must appear in \code{\link{upstream_slide_f_possibilities}}}
}
\value{
named list with two elements: \code{from_package}, a string containing the
upstream package name ("data.table" or "slider"), and \code{namer}, a function
that takes a column to call \code{.f} on and outputs a basic name or
abbreviation for what operation \code{.f} represents on that kind of column
(e.g., "sum", "av", "count").
}
\description{
Validate & get information about an upstream slide function
}
\keyword{internal}
