% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\docType{data}
\name{upstream_slide_f_possibilities}
\alias{upstream_slide_f_possibilities}
\title{Information about upstream (\code{{data.table}}/\code{{slider}}) slide functions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 3 columns.
}
\usage{
upstream_slide_f_possibilities
}
\description{
Underlies \code{\link{upstream_slide_f_info}}.
}
\keyword{internal}
