% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide-refactor.R
\name{upstream_slide_to_simple_hop}
\alias{upstream_slide_to_simple_hop}
\title{Convert upstream specialized slide function to a simple hop function}
\usage{
upstream_slide_to_simple_hop(
  .f,
  ...,
  .in_colnames,
  .out_colnames,
  .before_n_steps,
  .after_n_steps
)
}
\description{
upstream_slide_to_simple_hop(frollmean, .in_colnames = "value", .out_colnames = "slide_value", .before_n_steps = 1L, .after_n_steps = 0L)(
tibble(time_value = 1:5, value = 1:5),
tibble(geo_value = 1),
3:4
)
upstream_slide_to_simple_hop(slide_mean, .in_colnames = "value", .out_colnames = "slide_value", .before_n_steps = 1L, .after_n_steps = 0L)(
tibble(time_value = 1:5, value = 1:5),
tibble(geo_value = 1),
3:4
)
}
\details{
upstream_slide_to_simple_hop(frollmean, .in_colnames = "value", .out_colnames = "slide_value", .before_n_steps = Inf, .after_n_steps = 0L)(
tibble(time_value = 1:5, value = 1:5),
tibble(geo_value = 1),
3:4
)
}
\keyword{internal}
