% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide-refactor.R
\name{upstream_slide_to_simple_hopper}
\alias{upstream_slide_to_simple_hopper}
\title{upstream_slide_to_simple_hopper(frollmean, .in_colnames = "value", .out_colnames = "slide_value", .before_n_steps = 1L, .after_n_steps = 0L)(
tibble(time_value = 1:5, value = 1:5),
tibble(geo_value = 1),
3:4
)}
\usage{
upstream_slide_to_simple_hopper(
  .f,
  ...,
  .in_colnames,
  .out_colnames,
  .before_n_steps,
  .after_n_steps
)
}
\description{
upstream_slide_to_simple_hopper(frollmean, .in_colnames = "value", .out_colnames = "slide_value", .before_n_steps = 1L, .after_n_steps = 0L)(
tibble(time_value = 1:5, value = 1:5),
tibble(geo_value = 1),
3:4
)
}
