% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting-helpers2.R
\name{validate_nice_version_lag}
\alias{validate_nice_version_lag}
\title{Validate that \code{version_lag} is an unambiguous, length-1, \code{version-time_value} possibility}
\usage{
validate_nice_version_lag(
  version_lag,
  x,
  version_lag_arg = rlang::caller_arg(version_lag),
  call = rlang::caller_env()
)
}
\description{
Validate that \code{version_lag} is an unambiguous, length-1, \code{version-time_value} possibility
}
\keyword{internal}
