% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting-helpers2.R
\name{validate_nice_version_lags}
\alias{validate_nice_version_lags}
\title{Validate that adding \code{version_lags} is unambiguous and moves us from time values to versions}
\usage{
validate_nice_version_lags(
  version_lags,
  x,
  len = NULL,
  version_lags_arg = rlang::caller_arg(version_lags),
  call = rlang::caller_env()
)
}
\description{
Validate that adding \code{version_lags} is unambiguous and moves us from time values to versions
}
\keyword{internal}
