% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_signal_format}
\alias{validate_signal_format}
\title{Validate signal format and signal_var}
\usage{
validate_signal_format(
  x,
  signal_format,
  signal_var,
  other_keys,
  value_var = "value"
)
}
\description{
Validate signal format and signal_var
}
\keyword{internal}
