% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{validate_slide_window_arg}
\alias{validate_slide_window_arg}
\title{Validate \code{.before} or \code{.window_size} argument}
\usage{
validate_slide_window_arg(
  arg,
  time_type,
  lower = 1,
  allow_inf = TRUE,
  arg_name = rlang::caller_arg(arg)
)
}
\description{
Validate \code{.before} or \code{.window_size} argument
}
\keyword{internal}
