% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{validate_version_bound}
\alias{validate_version_bound}
\title{Validate a version bound arg}
\usage{
validate_version_bound(
  version_bound,
  x,
  na_ok = FALSE,
  version_bound_arg = rlang::caller_arg(version_bound),
  x_arg = rlang::caller_arg(x)
)
}
\arguments{
\item{version_bound}{the version bound to validate}

\item{x}{a data frame containing a version column with which to check
compatibility}

\item{na_ok}{Boolean; is \code{NA} an acceptable "bound"? (If so, \code{NA} will
have a special context-dependent meaning.)}

\item{version_bound_arg}{optional string; what to call the version bound in
error messages}
}
\description{
Expected to be used on \code{clobberable_versions_start}, \code{versions_end}, and
similar arguments. Some additional context-specific checks may be needed.
Side effects: raises an error if version bound appears invalid.
}
\keyword{internal}
