% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_cast_patched}
\alias{vec_cast_patched}
\title{Variant of \code{\link[vctrs:vec_cast]{vctrs::vec_cast}} that allows chr <-> date, disallows is.numeric x <-> some times}
\usage{
vec_cast_patched(
  x,
  to,
  ...,
  x_arg = caller_arg(x),
  to_arg = "",
  call = caller_env()
)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()}, \code{vec_cast_default()}, and \code{vec_restore()}, these
dots are only for future extensions and should be empty.}

\item{x_arg}{Argument name for \code{x}, used in error messages to
inform the user about the locations of incompatible types
(see \code{\link[vctrs:stop_incompatible_type]{stop_incompatible_type()}}).}

\item{to_arg}{Argument name \code{to} used in error messages to
inform the user about the locations of incompatible types
(see \code{\link[vctrs:stop_incompatible_type]{stop_incompatible_type()}}).}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Doesn't implement other conversions implied by the hierarchy, e.g.,
chr <-> POSIX\{c,l\}t.
}
\keyword{internal}
