% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{vec_position_lag}
\alias{vec_position_lag}
\title{Lag entries in a vctrs-style vector by their position in the vector}
\usage{
vec_position_lag(x, n)
}
\description{
Lag entries in a vctrs-style vector by their position in the vector
}
\keyword{internal}
