% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditions.R
\name{Warn}
\alias{Warn}
\title{Augmented version of \code{\link[rlang:abort]{rlang::warn}}}
\usage{
Warn(
  message,
  class_suffix = NULL,
  display_subfields = list(),
  more_subfields = list(),
  call = caller_env(),
  show_tutorial = TRUE,
  ...
)
}
\arguments{
\item{message}{character vector; an \code{rlang} non-\code{cli}-format message; this
will be line-wrapped and augmented with information about
\code{display_subfields} and how to access any \code{more_subfields}}

\item{class_suffix}{string or NULL; if non-NULL, and we are in a package,
then \code{"<pkgname>__<class_suffix>"} will be tacked on as an S3 class of the
condition object; if we are not in a package, then the suffix (without any
prefix) will be added as an S3 class. Additionally, regardless of whether
\code{class_suffix} is \code{NULL}, if we are in a package, two additional S3 classes
will be added as well: \code{"<pkgname>__<conditiontype>"} and
\code{"<pkgname>__condition"}}

\item{display_subfields}{list of objects to display as part of the condition
message using \code{\link[utils:str]{utils::str}} and save in the condition object}

\item{more_subfields}{list of objects to save in the condition object but
only point to, not display, in the condition message}

\item{call}{environment; the calling environment from which the trace should
be produced}
}
\description{
Augmented version of \code{\link[rlang:abort]{rlang::warn}}
}
