% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditions.R
\name{wrap_noncli_message}
\alias{wrap_noncli_message}
\title{Line-wraps an \code{rlang} non-\code{cli}-format condition message}
\usage{
wrap_noncli_message(
  message,
  width = getOption("width") - 1L,
  first_part_external_prefix = "! ",
  later_parts_external_prefix = "* ",
  first_part_wrap_prefix = later_parts_wrap_prefix,
  later_parts_wrap_prefix = "  "
)
}
\arguments{
\item{message}{possibly-named character vector; an \code{rlang}-style,
non-\code{cli}-format condition message}

\item{width}{\code{width} argument for \code{\link[base:strwrap]{base::strwrap}}}

\item{first_part_external_prefix}{content that we expect to be printed by
\code{rlang} before the first message part (if \code{message} has at least one part),
such as \code{"! "}}

\item{later_parts_external_prefix}{content of the same width of the content
that we expect to be printed by \code{rlang} before every part other than the
first, e.g., fancy versions of \code{"* "}, \code{"i "}, \code{"x "}, or two spaces}

\item{first_part_wrap_prefix}{prefix to add to new lines that are introduced
by line wrapping the first part of the message (if \code{message} has at least
one part)}

\item{later_parts_wrap_prefix}{prefix to add to , e.g., two spaces to align
with the text in bullet points}
}
\value{
named character vector; an \code{rlang}-style, non-\code{cli}-format condition
message
}
\description{
\code{rlang} condition messages can be character vectors with names indicating
line prefixes such as bullet points. There is a basic format and a \code{cli}
format. The basic format does not undergo line wrapping, so this function can
be used to apply line wrapping ahead of time; it is incompatible with
\code{use_cli_format=TRUE}(or would at least require \code{wrap_prefix=""}), as \code{cli}
does its own paragraph formatting that will join across newlines and rewrap.
}
