% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_symbolics}
\alias{wrap_symbolics}
\title{Line wrap list holding \link[rlang:is_symbolic]{symbolic}, with prefix&indent}
\usage{
wrap_symbolics(
  symbolics,
  initial = "",
  common_prefix = "",
  none_str = "<none>",
  width = getOption("width", 80L)
)
}
\arguments{
\item{symbolics}{List of \link[rlang:is_symbolic]{symbolic} objects: the variable
names (potentially empty)}

\item{initial}{Optional; single string: a prefix for the initial line in the
result; e.g., "Variable names: ". Defaults to "". Any non-initial lines
will be indented with whitespace matching the (estimated) visual width of
\code{initial}.}

\item{common_prefix}{Optional; single string: a prefix for every line (will
appear before \code{initial}); e.g., "# ". Defaults to "".}

\item{none_str}{Optional; single string: what to display when given
\code{length}-0 input. Will be combined with \code{common_prefix} and \code{initial}.}

\item{width}{Optional; single integer: desired maximum formatted line width.
The formatted output may not obey this setting if \code{common_prefix} plus
\code{initial} is long or the printing width is very narrow.}
}
\value{
\code{chr}; to print, use \code{\link[base:writeLines]{base::writeLines}}.
}
\description{
Helps pretty-print these objects. Adds backticks, commas, prefixes, and
indentation. Wraps lines, but won't insert line breaks in the middle of any
name while doing so.
}
\keyword{internal}
