% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_aware_training_data.R
\name{add_7davs_and_target}
\alias{add_7davs_and_target}
\title{Add 7dav and target to the data
Target is the updates made ref_lag days after the first release}
\usage{
add_7davs_and_target(df, value_col, refd_col, lag_col)
}
\arguments{
\item{df}{Data Frame of aggregated counts within a single location
reported for each reference date and issue date.}

\item{value_col}{column name for the column of raw value}

\item{refd_col}{column name for the column of reference date}

\item{lag_col}{column name for the column of lag}
}
\description{
Add 7dav and target to the data
Target is the updates made ref_lag days after the first release
}
