% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_archive.R
\name{as_epi_archive}
\alias{as_epi_archive}
\title{Convert to \code{epi_archive} object}
\usage{
as_epi_archive(
  x,
  geo_type,
  time_type,
  max_version,
  other_keys,
  additional_metadata = list()
)
}
\description{
Converts a data frame, tibble, or data table into an \code{epi_archive} object,
allowing for fast querying of data snapshots into \code{epi_df} format as of
certain versions. See the \href{https://cmu-delphi.github.io/epiprocess/articles/archive.html}{data versioning vignette} for
examples.
}
