% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{as_epi_archive}
\alias{as_epi_archive}
\title{Convert to \code{epi_archive} format}
\usage{
as_epi_archive(
  x,
  geo_type,
  time_type,
  other_keys,
  additional_metadata = list()
)
}
\arguments{
\item{x}{A data frame, data table, or tibble, with columns \code{geo_value},
\code{time_value}, \code{version}, and then any additional number of columns.}

\item{geo_type}{Type for the geo values. If missing, then the function will
attempt to infer it from the geo values present; if this fails, then it
will be set to "custom".}

\item{time_type}{Type for the time values. If missing, then the function will
attempt to infer it from the time values present; if this fails, then it
will be set to "custom".}

\item{other_keys}{Character vector specifying the names of variables in \code{x}
that should be considered key variables (in the language of \code{data.table})
apart from "geo_value", "time_value", and "version".}

\item{additional_metadata}{List of additional metadata to attach to the
\code{epi_archive} object. The metadata will have \code{geo_type} and \code{time_type}
fields; named entries from the passed list or will be included as well.}
}
\value{
An \code{epi_archive} object.
}
\description{
Converts a data frame, data table, or tibble into an \code{epi_archive}
object. See the \href{https://cmu-delphi.github.io/epiprocess/articles/archive.html}{archive vignette} for
examples.
}
\details{
This simply a wrapper around the \code{new()} method of the \code{epi_archive}
class, so for example:\preformatted{x <- as_epi_archive(df, geo_type = "state", time_type = "day")
}

would be equivalent to:\preformatted{x <- epi_archive$new(df, geo_type = "state", time_type = "day")
}
}
\examples{
library(delphi.epidata)
library(epiprocess)
library(data.table)
library(dplyr)

# must be given a dataframe and contain a `version` column
covidcast(data_source = "jhu-csse",
signals = "confirmed_7dav_incidence_prop",
time_type = "day",
geo_type = "state",
time_value = epirange(20200601, 20200603),
geo_values = "ca,fl",
issues = epirange(20200601, 20200603)
) \%>\%
fetch_tbl() \%>\%
select(geo_value, time_value, version = issue, case_rate = value) \%>\%
as_epi_archive() 
}
