% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_df.R
\name{as_tibble.epi_df}
\alias{as_tibble.epi_df}
\title{Convert to tibble}
\usage{
\method{as_tibble}{epi_df}(x, ...)
}
\arguments{
\item{x}{an \code{epi_df}}

\item{...}{forwarded to \code{as_tibble} for \code{data.frame}s}
}
\description{
Converts an \code{epi_df} object into a tibble, dropping metadata and any
grouping.
}
\details{
Advanced: if you are working with a third-party package that uses
\code{as_tibble()} on \code{epi_df}s but you actually want them to remain \code{epi_df}s,
use \code{attr(your_epi_df, "decay_to_tibble") <- FALSE} beforehand.
}
