% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cli_cat}
\alias{cli_cat}
\title{Like \code{\link[cli:cli_abort]{cli::cli_inform}}, but nicely mixable with \code{print} and \code{cat} in Rmd & qmd}
\usage{
cli_cat(message, .envir = parent.frame())
}
\description{
See https://github.com/cmu-delphi/epipredict/issues/277 for an
example of what can happen otherwise.
}
\keyword{internal}
