% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{data}
\name{covid_incidence_outliers}
\alias{covid_incidence_outliers}
\title{Subset of JHU daily COVID-19 cases from New Jersey and Florida}
\format{
An object of class \code{epi_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 730 rows and 3 columns.
}
\source{
This object contains a modified part of the
\href{https://github.com/CSSEGISandData/COVID-19}{COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University}
as \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{republished in the COVIDcast Epidata API}.
This data set is licensed under the terms of the
\href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0 International license}
by the Johns Hopkins University on behalf of its Center for Systems
Science in Engineering. Copyright Johns Hopkins University 2020.

Modifications:
\itemize{
\item \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{From the COVIDcast Epidata API}:
These signals are taken directly from the JHU CSSE \href{https://github.com/CSSEGISandData/COVID-19}{COVID-19 GitHub repository} without changes.
\item Furthermore, the data has been limited to a very small number of rows,
formatted into an \code{epi_df}, and the signal names slightly altered.
}
}
\usage{
covid_incidence_outliers
}
\description{
This data source of confirmed COVID-19 cases is based on reports made
available by the Center for Systems Science and Engineering at Johns
Hopkins University. This example data is downloaded from the CMU Delphi
COVIDcast Epidata API. It is a snapshot as of October 28, 2021, and captures the
cases from June 1, 2020 to May 31, 2021. It is limited to New Jersey and
Florida.

This data set is used in the {epiprocess} vignette on outliers.
}
\section{Data dictionary}{


The data has columns:
\describe{
\item{geo_value}{the geographic value associated with each row of measurements.}
\item{time_value}{the time value associated with each row of measurements.}
\item{cases}{Number of new confirmed COVID-19 cases, daily}
}

}

\examples{
# Since this is a re-exported dataset, it cannot be loaded using
# the `data()` function. `data()` looks for a file of the same name
# in the `data/` directory, which doesn't exist in this package.
# works
epiprocess::covid_incidence_outliers

# works
library(epiprocess)
covid_incidence_outliers

# fails
data(covid_incidence_outliers, package = "epiprocess")
}
\keyword{datasets}
