% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect outliers in a variable in an \code{epi_tibble} object}
\usage{
detect_outliers(
  x,
  var,
  methods = tibble(method = "rm", args = list(list()), abbr = "rm"),
  combiner = c("median", "mean", "none"),
  new_col_name = "outlier_info"
)
}
\arguments{
\item{x}{The \code{epi_tibble} object under consideration.}

\item{var}{The variable in \code{x} on which to run outlier detection.}

\item{methods}{A tibble specifying the method(s) to use for outlier
detection, with one row per method, and the following columns:
\itemize{
\item \code{method} <str>: Either "rm" or "stl", or a custom function for outlier
detection; see details for further explanation.
\item \code{args} <lst>: Named list of arguments that will be passed to the
detection method.
\item \code{abbr} <chr>: Abbreviation to use in naming output columns with results
from this method.
}}

\item{combiner}{String, one of "median", "mean", or "none", specifying how to
combine results from different outlier detection methods for the thresholds
determining whether a particular observation is classified as an outlier,
as well as a replacement value for any outliers.  If "none", then no
summarized results are calculated. Note that if the number of \code{methods}
(number of rows) is odd, then "median" is equivalent to a majority vote for
purposes of determining whether a given observation is an outlier.}

\item{new_col_name}{String indicating the name of the new column that will
contain the results of outlier detection. Default is "outlier_info"; note
that setting \code{new_col_name} equal to an existing column name will overwrite
this column.}
}
\value{
An \code{epi_tibble} object given by appending a new column to \code{x}, named
according to the \code{new_col_name} argument, containing the outlier detection
thresholds and replacement values from all detection methods.
}
\description{
Applies one or more outlier detection methods to a variable in an
\code{epi_tibble} object, and optionally aggregates the results to create
consensus results.
}
\details{
Each outlier detection method, one per row of the passed \code{methods}
tibble, is a function that must take as its first two arguments \code{x} and
\code{var}, indicating a data frame and a given variable on which to run outlier
detection, and then any number of additional arguments. The function must
return a tibble with the same number of rows as \code{x}, and with columns
\code{lower}, \code{upper}, and \code{replacement}, representing lower and upper bounds
for what would be considered an outlier, and a posited replacement value,
respectively.

For convenience, the outlier detection method can be specified (in the
\code{method} column of \code{methods}) by a string "rm", shorthand for
\code{detect_outliers_rm()}, which detects outliers via a rolling median; or by
"stl", shorthand for \code{detect_outliers_stl()}, which detects outliers via an
STL decomposition.
}
