% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{epi_archive}
\alias{epi_archive}
\title{\code{epi_archive} object}
\description{
An \code{epi_archive} is an R6 class which contains a data table
along with several relevant pieces of metadata. The data table can be seen
as the full archive (version history) for some signal variables of
interest.
}
\details{
An \code{epi_archive} is an R6 class which contains a data table \code{DT}, of
class \code{data.table} from the \code{data.table} package, with (at least) the
following columns:
\itemize{
\item \code{geo_value}: the geographic value associated with each row of measurements.
\item \code{time_value}: the time value associated with each row of measurements.
\item \code{version}: the time value specifying the version for each row of
measurements. For example, if in a given row the \code{version} is January 15,
2022 and \code{time_value} is January 14, 2022, then this row contains the
measurements of the data for January 14, 2022 that were available one day
later.
}

The data table \code{DT} has key variables \code{geo_value}, \code{time_value}, \code{version},
as well as any others (these can be specified when instantiating the
\code{epi_archive} object via the \code{other_keys} argument, and/or set by operating
on \code{DT} directly). Refer to the documentation for \code{\link[=as_epi_archive]{as_epi_archive()}} for
information and examples of relevant parameter names for an \code{epi_archive} object.
Note that there can only be a single row per unique combination of
key variables, and thus the key variables are critical for figuring out how
to generate a snapshot of data from the archive, as of a given version.

In general, the last version of each observation is carried forward (LOCF) to
fill in data between recorded versions, and between the last recorded
update and the \code{versions_end}. One consequence is that the \code{DT}
doesn't have to contain a full snapshot of every version (although this
generally works), but can instead contain only the rows that are new or
changed from the previous version (see \code{compactify}, which does this
automatically). Currently, deletions must be represented as revising a row
to a special state (e.g., making the entries \code{NA} or including a special
column that flags the data as removed and performing some kind of
post-processing), and the archive is unaware of what this state is. Note
that \code{NA}s \emph{can} be introduced by \code{epi_archive} methods for other reasons,
e.g., in \code{\link{epix_fill_through_version}} and \code{\link{epix_merge}}, if requested, to
represent potential update data that we do not yet have access to; or in
\code{\link{epix_merge}} to represent the "value" of an observation before the
version in which it was first released, or if no version of that
observation appears in the archive data at all.

\strong{A word of caution:} R6 objects, unlike most other objects in R, have
reference semantics. A primary consequence of this is that objects are not
copied when modified. You can read more about this in Hadley Wickham's
\href{https://adv-r.hadley.nz/r6.html#r6-semantics}{Advanced R} book. In order
to construct a modified archive while keeping the original intact, first
make a clone using the \verb{$clone} method, then overwrite the clone's \code{DT}
field with \code{data.table::copy(clone$DT)}, and finally perform the
modifications on the clone.
}
\section{Metadata}{

The following pieces of metadata are included as fields in an \code{epi_archive}
object:
\itemize{
\item \code{geo_type}: the type for the geo values.
\item \code{time_type}: the type for the time values.
\item \code{additional_metadata}: list of additional metadata for the data archive.
}

Unlike an \code{epi_df} object, metadata for an \code{epi_archive} object \code{x} can be
accessed (and altered) directly, as in \code{x$geo_type} or \code{x$time_type},
etc. Like an \code{epi_df} object, the \code{geo_type} and \code{time_type} fields in the
metadata of an \code{epi_archive} object are not currently used by any
downstream functions in the \code{epiprocess} package, and serve only as useful
bits of information to convey about the data set at hand.
}

\section{Generating Snapshots}{

An \code{epi_archive} object can be used to generate a snapshot of the data in
\code{epi_df} format, which represents the most up-to-date values of the signal
variables, as of the specified version. This is accomplished by calling the
\code{as_of()} method for an \code{epi_archive} object \code{x}. More details on this
method are documented in the wrapper function \code{\link[=epix_as_of]{epix_as_of()}}.
}

\section{Sliding Computations}{

We can run a sliding computation over an \code{epi_archive} object, much like
\code{epi_slide()} does for an \code{epi_df} object. This is accomplished by calling
the \code{slide()} method for an \code{epi_archive} object, which works similarly to
the way \code{epi_slide()} works for an \code{epi_df} object, but with one key
difference: it is version-aware. That is, for an \code{epi_archive} object, the
sliding computation at any given reference time point t is performed on
\strong{data that would have been available as of t}. More details on \code{slide()}
are documented in the wrapper function \code{\link[=epix_slide]{epix_slide()}}.
}

\examples{
tib <- tibble::tibble(
  geo_value = rep(c("ca", "hi"), each = 5),
  time_value = rep(seq(as.Date("2020-01-01"), 
                       by = 1, length.out = 5), times = 2),
  version = rep(seq(as.Date("2020-01-02"), 
                    by = 1, length.out = 5), times = 2),
  value = rnorm(10, mean = 2, sd = 1)
)

toy_epi_archive <- tib \%>\% epi_archive$new(geo_type = "state", 
                                           time_type = "day")
toy_epi_archive 
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{DT}}{(\code{data.table})\cr
the (optionally compactified) datatable}

\item{\code{geo_type}}{(string)\cr
the resolution of the geographic label (e.g. state)}

\item{\code{time_type}}{(string)\cr
the resolution of the time column (e.g. day)}

\item{\code{additional_metadata}}{(named list)\cr
any extra fields, such as \code{other_keys}}

\item{\code{clobberable_versions_start}}{(length-1 of same type&class as \code{version} column, or \code{NA})\cr
the earliest version number that might be rewritten in the future without assigning a new version
date/number, or \code{NA} if this won't happen}

\item{\code{versions_end}}{(length-1 of same type&class as \code{version} column)\cr
the latest version observed}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-epi_archive-new}{\code{epi_archive$new()}}
\item \href{#method-epi_archive-print}{\code{epi_archive$print()}}
\item \href{#method-epi_archive-as_of}{\code{epi_archive$as_of()}}
\item \href{#method-epi_archive-fill_through_version}{\code{epi_archive$fill_through_version()}}
\item \href{#method-epi_archive-truncate_versions_after}{\code{epi_archive$truncate_versions_after()}}
\item \href{#method-epi_archive-merge}{\code{epi_archive$merge()}}
\item \href{#method-epi_archive-group_by}{\code{epi_archive$group_by()}}
\item \href{#method-epi_archive-slide}{\code{epi_archive$slide()}}
\item \href{#method-epi_archive-clone}{\code{epi_archive$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-epi_archive-new"></a>}}
\if{latex}{\out{\hypertarget{method-epi_archive-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{epi_archive} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{epi_archive$new(
  x,
  geo_type,
  time_type,
  other_keys,
  additional_metadata,
  compactify,
  clobberable_versions_start,
  versions_end
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A data frame, data table, or tibble, with columns \code{geo_value},
\code{time_value}, \code{version}, and then any additional number of columns.}

\item{\code{geo_type}}{Type for the geo values. If missing, then the function will
attempt to infer it from the geo values present; if this fails, then it
will be set to "custom".}

\item{\code{time_type}}{Type for the time values. If missing, then the function will
attempt to infer it from the time values present; if this fails, then it
will be set to "custom".}

\item{\code{other_keys}}{Character vector specifying the names of variables in \code{x}
that should be considered key variables (in the language of \code{data.table})
apart from "geo_value", "time_value", and "version".}

\item{\code{additional_metadata}}{List of additional metadata to attach to the
\code{epi_archive} object. The metadata will have \code{geo_type} and \code{time_type}
fields; named entries from the passed list or will be included as well.}

\item{\code{compactify}}{Optional; Boolean or \code{NULL}: should we remove rows that are
considered redundant for the purposes of \code{epi_archive}'s built-in methods
such as \code{as_of}? As these methods use the last version of each observation
carried forward (LOCF) to interpolate between the version data provided,
rows that don't change these LOCF results can potentially be omitted to
save space while maintaining the same behavior (with the help of the
\code{clobberable_versions_start} and \code{versions_end} fields in some edge cases).
\code{TRUE} will remove these rows, \code{FALSE} will not, and missing or \code{NULL} will
remove these rows and issue a warning. Generally, this can be set to
\code{TRUE}, but if you directly inspect or edit the fields of the \code{epi_archive}
such as its \code{DT}, or rely on redundant updates to achieve a certain
behavior of the \code{ref_time_values} default in \code{epix_slide}, you will have to
determine whether \code{compactify=TRUE} will produce the desired results. If
compactification here is removing a large proportion of the rows, this may
indicate a potential for space, time, or bandwidth savings upstream the
data pipeline, e.g., by avoiding fetching, storing, or processing these
rows of \code{x}.}

\item{\code{clobberable_versions_start}}{Optional; as in \code{\link{as_epi_archive}}}

\item{\code{versions_end}}{Optional; as in \code{\link{as_epi_archive}}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Refer to the documentation for \code{\link[=as_epi_archive]{as_epi_archive()}} for more information
and examples of parameter names.
Print information about an archive
}

\subsection{Returns}{
An \code{epi_archive} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-epi_archive-print"></a>}}
\if{latex}{\out{\hypertarget{method-epi_archive-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{epi_archive$print(class = TRUE, methods = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class}}{Boolean; whether to print the class label header}

\item{\code{methods}}{Boolean; whether to print all available methods of
the archive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-epi_archive-as_of"></a>}}
\if{latex}{\out{\hypertarget{method-epi_archive-as_of}{}}}
\subsection{Method \code{as_of()}}{
Generates a snapshot in \code{epi_df} format as of a given version.
See the documentation for the wrapper function \code{\link[=epix_as_of]{epix_as_of()}} for details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{epi_archive$as_of(max_version, min_time_value = -Inf, all_versions = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_version}}{Version specifying the max version to permit in the
snapshot. That is, the snapshot will comprise the unique rows of the
current archive data that represent the most up-to-date signal values, as
of the specified \code{max_version} (and whose \code{time_value}s are at least
\code{min_time_value}).}

\item{\code{min_time_value}}{Time value specifying the min \code{time_value} to permit in
the snapshot. Default is \code{-Inf}, which effectively means that there is no
minimum considered.}

\item{\code{all_versions}}{Boolean; If \code{all_versions = TRUE}, then the output will be in
\code{epi_archive} format, and contain rows in the specified \code{time_value} range
having \code{version <= max_version}. The resulting object will cover a
potentially narrower \code{version} and \code{time_value} range than \code{x}, depending
on user-provided arguments. Otherwise, there will be one row in the output
for the \code{max_version} of each \code{time_value}. Default is \code{FALSE}.}

\item{\code{x}}{An \code{epi_archive} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-epi_archive-fill_through_version"></a>}}
\if{latex}{\out{\hypertarget{method-epi_archive-fill_through_version}{}}}
\subsection{Method \code{fill_through_version()}}{
Fill in unobserved history using requested scheme by mutating
\code{self} and potentially reseating its fields. See
\code{\link{epix_fill_through_version}} for a full description of the non-R6-method
version, which doesn't mutate the input archive but might alias its fields.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{epi_archive$fill_through_version(fill_versions_end, how = c("na", "locf"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fill_versions_end}}{as in \code{\link{epix_fill_through_version}}}

\item{\code{how}}{as in \code{\link{epix_fill_through_version}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-epi_archive-truncate_versions_after"></a>}}
\if{latex}{\out{\hypertarget{method-epi_archive-truncate_versions_after}{}}}
\subsection{Method \code{truncate_versions_after()}}{
Filter to keep only older versions, mutating the archive by
potentially reseating but not mutating some fields. \code{DT} is likely, but not
guaranteed, to be copied. Returns the mutated archive
\link[base:invisible]{invisibly}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{epi_archive$truncate_versions_after(max_version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_version}}{as in \code{\link{epix_truncate_versions_after}}}

\item{\code{x}}{as in \code{\link{epix_truncate_versions_after}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-epi_archive-merge"></a>}}
\if{latex}{\out{\hypertarget{method-epi_archive-merge}{}}}
\subsection{Method \code{merge()}}{
Merges another \code{epi_archive} with the current one, mutating the
current one by reseating its \code{DT} and several other fields, but avoiding
mutation of the old \code{DT}; returns the current archive
\link[base:invisible]{invisibly}. See \code{\link{epix_merge}} for a full description
of the non-R6-method version, which does not mutate either archive, and
does not alias either archive's \code{DT}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{epi_archive$merge(
  y,
  sync = c("forbid", "na", "locf", "truncate"),
  compactify = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{as in \code{\link{epix_merge}}}

\item{\code{sync}}{as in \code{\link{epix_merge}}}

\item{\code{compactify}}{as in \code{\link{epix_merge}}
group an epi_archive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-epi_archive-group_by"></a>}}
\if{latex}{\out{\hypertarget{method-epi_archive-group_by}{}}}
\subsection{Method \code{group_by()}}{
group an epi_archive
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{epi_archive$group_by(
  ...,
  .add = FALSE,
  .drop = dplyr::group_by_drop_default(self)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{variables or computations to group by. Computations are always
done on the ungrouped data frame. To perform computations on the grouped
data, you need to use a separate \code{\link[=mutate]{mutate()}} step before the
\code{\link[=group_by]{group_by()}}}

\item{\code{.add}}{When \code{FALSE}, the default, \code{\link[=group_by]{group_by()}} will override existing
groups. To add to the existing groups, use \code{.add = TRUE}.}

\item{\code{.drop}}{Drop groups formed by factor levels that don't appear in the
data. The default is \code{TRUE} except when \code{.data} has been previously grouped
with \code{.drop = FALSE}. See \code{\link[=group_by_drop_default]{group_by_drop_default()}} for details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-epi_archive-slide"></a>}}
\if{latex}{\out{\hypertarget{method-epi_archive-slide}{}}}
\subsection{Method \code{slide()}}{
Slides a given function over variables in an \code{epi_archive}
object. See the documentation for the wrapper function \code{\link[=epix_slide]{epix_slide()}} for
details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{epi_archive$slide(
  f,
  ...,
  before,
  ref_time_values,
  time_step,
  new_col_name = "slide_value",
  as_list_col = FALSE,
  names_sep = "_",
  all_versions = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{Function, formula, or missing; together with \code{...} specifies the
computation to slide. To "slide" means to apply a computation over a
sliding (a.k.a. "rolling") time window for each data group. The window is
determined by the \code{before} parameter described below. One time step is
typically one day or one week; see \code{\link{epi_slide}} details for more
explanation. If a function, \code{f} must take an \code{epi_df} with the same
column names as the archive's \code{DT}, minus the \code{version} column; followed
by a one-row tibble containing the values of the grouping variables for
the associated group; followed by a reference time value, usually as a
\code{Date} object; followed by any number of named arguments. If a formula,
\code{f} can operate directly on columns accessed via \code{.x$var} or \code{.$var}, as
in \code{~ mean (.x$var)} to compute a mean of a column \code{var} for each
group-\code{ref_time_value} combination. The group key can be accessed via
\code{.y} or \code{.group_key}, and the reference time value can be accessed via
\code{.z} or \code{.ref_time_value}. If \code{f} is missing, then \code{...} will specify the
computation.}

\item{\code{...}}{Additional arguments to pass to the function or formula specified
via \code{f}. Alternatively, if \code{f} is missing, then \code{...} is interpreted as an
expression for tidy evaluation; in addition to referring to columns
directly by name, the expression has access to \code{.data} and \code{.env} pronouns
as in \code{dplyr} verbs, and can also refer to the \code{.group_key} and
\code{.ref_time_value}. See details of \code{\link{epi_slide}}.}

\item{\code{before}}{How far \code{before} each \code{ref_time_value} should the sliding
window extend? If provided, should be a single, non-NA,
\link[vctrs:vec_cast]{integer-compatible} number of time steps. This window
endpoint is inclusive. For example, if \code{before = 7}, and one time step is
one day, then to produce a value for a \code{ref_time_value} of January 8, we
apply the given function or formula to data (for each group present) with
\code{time_value}s from January 1 onward, as they were reported on January 8.
For typical disease surveillance sources, this will not include any data
with a \code{time_value} of January 8, and, depending on the amount of reporting
latency, may not include January 7 or even earlier \code{time_value}s. (If
instead the archive were to hold nowcasts instead of regular surveillance
data, then we would indeed expect data for \code{time_value} January 8. If it
were to hold forecasts, then we would expect data for \code{time_value}s after
January 8, and the sliding window would extend as far after each
\code{ref_time_value} as needed to include all such \code{time_value}s.)}

\item{\code{ref_time_values}}{Reference time values / versions for sliding
computations; each element of this vector serves both as the anchor point
for the \code{time_value} window for the computation and the \code{max_version}
\code{as_of} which we fetch data in this window. If missing, then this will set
to a regularly-spaced sequence of values set to cover the range of
\code{version}s in the \code{DT} plus the \code{versions_end}; the spacing of values will
be guessed (using the GCD of the skips between values).}

\item{\code{time_step}}{Optional function used to define the meaning of one time
step, which if specified, overrides the default choice based on the
\code{time_value} column. This function must take a positive integer and return
an object of class \code{lubridate::period}. For example, we can use \code{time_step = lubridate::hours} in order to set the time step to be one hour (this
would only be meaningful if \code{time_value} is of class \code{POSIXct}).}

\item{\code{new_col_name}}{String indicating the name of the new column that will
contain the derivative values. Default is "slide_value"; note that setting
\code{new_col_name} equal to an existing column name will overwrite this column.}

\item{\code{as_list_col}}{Should the slide results be held in a list column, or be
\link[tidyr:chop]{unchopped}/\link[tidyr:unnest]{unnested}? Default is \code{FALSE},
in which case a list object returned by \code{f} would be unnested (using
\code{\link[tidyr:unnest]{tidyr::unnest()}}), and, if the slide computations output data frames,
the names of the resulting columns are given by prepending \code{new_col_name}
to the names of the list elements.}

\item{\code{names_sep}}{String specifying the separator to use in \code{tidyr::unnest()}
when \code{as_list_col = FALSE}. Default is "_". Using \code{NULL} drops the prefix
from \code{new_col_name} entirely.}

\item{\code{all_versions}}{(Not the same as \code{all_rows} parameter of \code{epi_slide}.) If
\code{all_versions = TRUE}, then \code{f} will be passed the version history (all
\code{version <= ref_time_value}) for rows having \code{time_value} between
\code{ref_time_value - before} and \code{ref_time_value}. Otherwise, \code{f} will be
passed only the most recent \code{version} for every unique \code{time_value}.
Default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-epi_archive-clone"></a>}}
\if{latex}{\out{\hypertarget{method-epi_archive-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{epi_archive$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
