% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_archive.R
\name{epi_archive}
\alias{epi_archive}
\title{Create an \code{epi_archive} object}
\usage{
epi_archive(
  geo_value,
  time_value,
  version,
  ...,
  geo_type,
  time_type,
  max_version,
  other_keys,
  additional_metadata = list()
)
}
\arguments{
\item{geo_value}{Geographic values associated with the measurements.}

\item{time_value}{Time values associated with the measurements.}

\item{version}{Time values specifying the versions of the measurements. For
example, if in a given row the \code{version} is January 15, 2022 and
\code{time_value} is January 14, 2022, then this row contains the measurements
of the data for January 14, 2022 that were available one day later.}

\item{...}{Additional arguments of the form \code{value} or \code{name = value}, which
specify any number of additional columns for the \code{epi_archive} object.}

\item{geo_type}{Type for the geo values. If missing, then the function will
attempt to infer it from the geo values present; if this fails, then it
will be set to "custom".}

\item{time_type}{Type for the time values. If missing, then the function will
attempt to infer it from the time values present; if this fails, then it
will be set to "custom".}

\item{other_keys}{Character vector specifying the names of variables that
should be considered key variables (in the language of \code{data.table}) apart
from "geo_value", "time_value", and "version". There can only be a single
row per unique combination of key variables, and thus the key variables are
critical for figuring out how to generate snapshots of data form the
archive, as of certain versions.}

\item{additional_metadata}{List of additional metadata to attach to the
\code{epi_archive} object. The metadata will have \code{time_type} and \code{geo_type}
fields; named entries from the passed list or will be included as well.}
}
\value{
An \code{epi_archive} object.
}
\description{
Creates an \code{epi_archive} object from given \code{geo_value}, \code{time_value}, and
\code{version} variables, as well as any additional number of variables.
}
\details{
Last-observation-carried-forward (LOCF) is used to data in between
recorded versions. Currently, deletions must be represented as revising a
row to a special state (e.g., making the entries \code{NA} or including a
special column that flags the data as removed and performing some kind of
post-processing), and the archive is unaware of what this state is.

TODO some details on the structure of the R6 object and the fact that it
obeys reference semantics??
}
