% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_df.R
\name{epi_df}
\alias{epi_df}
\title{Create \code{epi_df} object}
\usage{
epi_df(
  geo_value,
  time_value,
  ...,
  geo_type,
  time_type,
  as_of,
  additional_metadata = list()
)
}
\arguments{
\item{geo_value}{Geographic values associated with the measurements.}

\item{time_value}{Time values associated with the measurements.}

\item{...}{Additional arguments of the form \code{value} or \code{name = value}, which
specify any number of additional columns for the \code{epi_df} object.}

\item{geo_type}{Type for the geo values. If missing, then the function will
attempt to infer it from the geo values present; if this fails, then it
will be set to "custom".}

\item{time_type}{Type for the time values. If missing, then the function will
attempt to infer it from the time values present; if this fails, then it
will be set to "custom".}

\item{as_of}{Time value representing the time at which the given data were
available. For example, if \code{as_of} were January 31, 2022, then the \code{epi_df}
object that is created would represent the most up-to-date version of the
data available as of January 31, 2022. If the \code{as_of} argument is missing,
then the current day-time will be used.}

\item{additional_metadata}{List of additional metadata to attach to the
\code{epi_df} object. The metadata will have \code{time_type}, \code{geo_type}, and
\code{as_of} fields; named entries from the passed list or will be included as
well.}
}
\description{
Creates an \code{epi_df} object from given \code{geo_value} and \code{time_value} variables,
and any additional number of variables.
}
\details{
An \code{epi_df} is a tibble with (at least) the following columns:
\itemize{
\item \code{geo_value}: the geographic value associated with each measurement.
\item \code{time_value}: the time value associated with each measurement.
}

Other columns can be considered as measured variables, which we also broadly
refer to as signal variables. An \code{epi_df} object also has metadata with (at
least) the following fields:
\itemize{
\item \code{geo_type}: the type for the geo values.
\item \code{time_type}: the type for the time values.
\item \code{as_of}: the time value at which the given data were available.
}

Metadata for an \code{epi_df} object \code{x} can be accessed (and altered) via
\code{attributes(x)$metadata}. The first two fields in the above list,
\code{geo_type} and \code{time_type}, can usually be inferred from the \code{geo_value}
and \code{time_value} columns, respectively. More information on their coding is
given below.

The last field in the above list, \code{as_of}, is one of the most unique aspects
of an \code{epi_df} object. In brief, we can think of an \code{epi_df} object as a
single snapshot of a data set that contains the most up-to-date values of
some signals of interest, as of the time specified in the \code{as_of} field.  A
companion object is the \code{epi_archive} object, which contains the full
version history of a given data set. Revisions are common in many types of
epidemiological data streams, and paying attention to data revisions can be
important for all sorts of downstream data analysis and modeling tasks. See
the \code{epi_archive()} help file for more details on how data versioning works
in the \code{epiprocess} package (including how to create \code{epi_df} objects, as
data snapshots, from an \code{epi_archive} object).
}
\section{Geo types}{

The following geo types are supported in an \code{epi_df}. Their geo coding
(specification of geo values for each geo type) is also described below.
\itemize{
\item \code{"county"}: each observation corresponds to a U.S. county; coded by 5-digit
FIPS code.
\item \code{"hrr"}: each observation corresponds to a U.S. hospital referral region
(designed to represent regional healthcare markets); there are 306 HRRs in
the U.S; coded by number (nonconsecutive, between 1 and 457).
\item \code{"state"}: each observation corresponds to a U.S. state; coded by 2-digit
postal abbreviation (lowercase);
note that Puerto Rico is "pr" and Washington D.C. is "dc".
\item \code{"hhs"}: each observation corresponds to a U.S. HHS region; coded by number
(consecutive, between 1 and 10).
\item \code{"nation"}: each observation corresponds to a country; coded by ISO 31661-
alpha-2 country codes (lowercase).
}

The above geo types come with aggregation utilities in the package, \emph{todo:
refer to relevant functionality, vignette, and so on}. An unrecognizable
geo type is labeled as "custom".
}

\section{Time types}{

The following time types are supported in an \code{epi_df}. Their time coding
(specification of time values for each time type) is also described below.
\itemize{
\item \code{"day-time"}: each observation corresponds to a time on a given day
(measured to the second); coded as a \code{POSIXct} object, as in
\code{as.POSIXct("2022-01-31 18:45:40")}.
\item \code{"day"}: each observation corresponds to a day; coded as a \code{Date} object,
as in \code{as.Date("2022-01-31")}.
\item \code{"week"}: each observation corresponds to a week; the alignment can be
arbitrary (as to whether a week starts on a Monday, Tuesday, etc.; the
U.S. CDC definition of an epidemiological week starts on a Sunday); coded
as a \code{Date} object, representing the start date of week.
}

An unrecognizable time type is labeled as "custom". \emph{todo: refer to vignette
for time aggregation examples}
}

