% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_df.R
\name{epi_df}
\alias{epi_df}
\title{\code{epi_df} object}
\description{
An \code{epi_df} is a tibble with certain minimal column structure
and metadata. It can be seen as a snapshot of a data set that contains the
most up-to-date values of some signal variables of interest, as of a given
time.
}
\details{
An \code{epi_df} is a tibble with (at least) the following columns:
\itemize{
\item \code{geo_value}: the geographic value associated with each row of measurements.
\item \code{time_value}: the time value associated with each row of measurements.
}

Other columns can be considered as measured variables, which we also refer to
as signal variables. An \code{epi_df} object also has metadata with (at least)
the following fields:
\itemize{
\item \code{geo_type}: the type for the geo values.
\item \code{time_type}: the type for the time values.
\item \code{as_of}: the time value at which the given data were available.
}

Metadata for an \code{epi_df} object \code{x} can be accessed (and altered) via
\code{attributes(x)$metadata}. The first two fields in the above list,
\code{geo_type} and \code{time_type}, can usually be inferred from the \code{geo_value}
and \code{time_value} columns, respectively. They are not currently used by any
downstream functions in the \code{epiprocess} package, and serve only as useful
bits of information to convey about the data set at hand. More information
on their coding is given below.

The last field in the above list, \code{as_of}, is one of the most unique aspects
of an \code{epi_df} object. In brief, we can think of an \code{epi_df} object as a
single snapshot of a data set that contains the most up-to-date values of
the signals variables, as of the time specified in the \code{as_of} field.

A companion object is the \code{epi_archive} object, which contains the full
version history of a given data set. Revisions are common in many types of
epidemiological data streams, and paying attention to data revisions can be
important for all sorts of downstream data analysis and modeling tasks. See
the documentation for \code{\link{epi_archive}} for more details on how
data versioning works in the \code{epiprocess} package (including how to
generate \code{epi_df} objects, as data snapshots, from an \code{epi_archive}
object).
}
\section{Geo Types}{

The following geo types are recognized in an \code{epi_df}.
\itemize{
\item \code{"county"}: each observation corresponds to a U.S. county; coded by 5-digit
FIPS code.
\item \code{"hrr"}: each observation corresponds to a U.S. hospital referral region
(designed to represent regional healthcare markets); there are 306 HRRs in
the U.S; coded by number (nonconsecutive, between 1 and 457).
\item \code{"state"}: each observation corresponds to a U.S. state; coded by 2-digit
postal abbreviation (lowercase);
note that Puerto Rico is "pr" and Washington D.C. is "dc".
\item \code{"hhs"}: each observation corresponds to a U.S. HHS region; coded by number
(consecutive, between 1 and 10).
\item \code{"nation"}: each observation corresponds to a country; coded by ISO 31661-
alpha-2 country codes (lowercase).
}

An unrecognizable geo type is labeled "custom".
}

\section{Time Types}{

The following time types are recognized in an \code{epi_df}.
\itemize{
\item \code{"day-time"}: each observation corresponds to a time on a given day
(measured to the second); coded as a \code{POSIXct} object, as in
\code{as.POSIXct("2022-01-31 18:45:40")}.
\item \code{"day"}: each observation corresponds to a day; coded as a \code{Date} object,
as in \code{as.Date("2022-01-31")}.
\item \code{"week"}: each observation corresponds to a week; the alignment can be
arbitrary (as to whether a week starts on a Monday, Tuesday); coded as a
\code{Date} object, representing the start date of week.
\item \code{"yearweek"}: each observation corresponds to a week; the alignment can be
arbitrary; coded as a \code{tsibble::yearweek} object, where the alignment is
stored in the \code{week_start} field of its attributes.
\item \code{"yearmonth"}: each observation corresponds to a month; coded as a
\code{tsibble::yearmonth} object.
\item \code{"yearquarter"}: each observation corresponds to a quarter; coded as a
\code{tsibble::yearquarter} object.
\item \code{"year"}: each observation corresponds to a year; coded as an integer
greater than or equal to 1582.
}

An unrecognizable time type is labeled "custom".
}

