% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{epi_slide}
\alias{epi_slide}
\title{Slide a function over variables in an \code{epi_df} object}
\usage{
epi_slide(
  x,
  f,
  ...,
  n = 7,
  ref_time_values,
  align = c("right", "center", "left"),
  before,
  time_step,
  new_col_name = "slide_value",
  as_list_col = FALSE,
  names_sep = "_",
  all_rows = FALSE
)
}
\arguments{
\item{x}{The \code{epi_df} object under consideration.}

\item{f}{Function or formula to slide over variables in \code{x}. To "slide" means
to apply a function or formula over a running window of \code{n} time steps
(where one time step is typically one day or one week; see details for more
explanation). If a function, \code{f} must take \code{x}, a data frame with the same
column names as the original object; followed by any number of named
arguments; and ending with \code{...}. If a formula, \code{f} can operate directly on
columns accessed via \code{.x$var}, as in \code{~ mean(.x$var)} to compute a mean of
a column \code{var} over a sliding window of \code{n} time steps.}

\item{...}{Additional arguments to pass to the function or formula specified
via \code{f}. Alternatively, if \code{f} is missing, then the current argument is
interpreted as an expression for tidy evaluation. See details.}

\item{n}{Number of time steps to use in the running window. For example, if
\code{n = 7}, one time step is one day, and the alignment is "right", then to
produce a value on January 7 we apply the given function or formula to data
in between January 1 and 7. Default is 7.}

\item{ref_time_values}{Time values for sliding computations, meaning, each
element of this vector serves as the reference time point for one sliding
window. If missing, then this will be set to all unique time values in the
underlying data table, by default.}

\item{align}{One of "right", "center", or "left", indicating the alignment of
the sliding window relative to the reference time point. If the alignment
is "center" and \code{n} is even, then one more time point will be used after
the reference time point than before. Default is "right".}

\item{before}{Positive integer less than \code{n}, specifying the number of time
points to use in the sliding window strictly before the reference time
point. For example, setting \code{before = n-1} would be the same as setting
\code{align = "right"}. The \code{before} argument allows for more flexible
specification of alignment than the \code{align} parameter, and if specified,
overrides \code{align}.}

\item{time_step}{Optional function used to define the meaning of one time
step, which if specified, overrides the default choice based on the
\code{time_value} column. This function must take a positive integer and return
an object of class \code{lubridate::period}. For example, we can use \code{time_step = lubridate::hours} in order to set the time step to be one hour (this
would only be meaningful if \code{time_value} is of class \code{POSIXct}).}

\item{new_col_name}{String indicating the name of the new column that will
contain the derivative values. Default is "slide_value"; note that setting
\code{new_col_name} equal to an existing column name will overwrite this column.}

\item{as_list_col}{Should the new column be stored as a list column? Default
is \code{FALSE}, in which case a list object returned by \code{f} would be unnested
(using \code{tidyr::unnest()}), and the names of the resulting columns are given
by prepending \code{new_col_name} to the names of the list elements.}

\item{names_sep}{String specifying the separator to use in \code{tidyr::unnest()}
when \code{as_list_col = FALSE}. Default is "_". Using \code{NULL} drops the prefix
from \code{new_col_name} entirely.}

\item{all_rows}{If \code{all_rows = TRUE}, then all rows of \code{x} will be kept in
the output; otherwise, there will be one row for each time value in \code{x}
that acts as a reference time value. Default is \code{FALSE}.}
}
\value{
An \code{epi_df} object given by appending a new column to \code{x}, named
according to the \code{new_col_name} argument.
}
\description{
Slides a given function over variables in an \code{epi_df} object. See the \href{https://cmu-delphi.github.io/epiprocess/articles/slide.html}{slide vignette} for
examples.
}
\details{
To "slide" means to apply a function or formula over a running
window of \code{n} time steps, where the unit (the meaning of one time step) is
implicitly defined by the way the \code{time_value} column treats addition and
subtraction; for example, if the time values are coded as \code{Date} objects,
then one time step is one day, since \code{as.Date("2022-01-01") + 1} equals
\code{as.Date("2022-01-02")}. Alternatively, the time step can be set explicitly
using the \code{time_step} argument (which if specified would override the
default choice based on \code{time_value} column). If less than \code{n} time steps
are available at any given reference time value, then \code{epi_slide()} still
attempts to perform the computation anyway (it does not require a complete
window). The issue of what to do with partial computations (those run on
incomplete windows) is therefore left up to the user, either through the
specified function or formula \code{f}, or through post-processing.

If \code{f} is missing, then an expression for tidy evaluation can be specified,
for example, as in:

\if{html}{\out{<div class="sourceCode">}}\preformatted{epi_slide(x, cases_7dav = mean(cases), n = 7)
}\if{html}{\out{</div>}}

which would be equivalent to:

\if{html}{\out{<div class="sourceCode">}}\preformatted{epi_slide(x, function(x, ...) mean(x$cases), n = 7,
          new_col_name = "cases_7dav")
}\if{html}{\out{</div>}}

Thus, to be clear, when the computation is specified via an expression for
tidy evaluation (first example, above), then the name for the new column is
inferred from the given expression and overrides any name passed explicitly
through the \code{new_col_name} argument.

When \code{f} is a named function with arguments, if a tibble with an unnamed
grouping variable is passed in as the method argument to \code{f}, include a
parameter for the grouping-variable in \verb{function()} just prior to
specifying the method to prevent that from being overridden. For example:\preformatted{# Construct an tibble with an unnamed grouping variable
edf = bind_rows(tibble(geo_value = "ak", time_value = as.Date("2020-01-01") 
          + 1:10, x1=1:10, y=1:10 + rnorm(10L))) \%>\% 
  as_epi_df()

# Now, include a row parameter for the grouping variable in the tibble, 
# which we denote as g, just prior to method = "qr"
# Note that if g was not included below, then the method = "qr" would be 
# overridden, as described above
edf \%>\%
group_by(geo_value) \%>\%
epi_slide(function(x, g, method="qr", ...) tibble(model=list(
          lm(y ~ x1, x, method=method))), n=7L) 
}
}
\examples{
 # slide a 7-day trailing average formula on cases
  jhu_csse_daily_subset \%>\%
  group_by(geo_value) \%>\%
  epi_slide(cases_7dav = mean(cases), n = 7, 
            align = "right") \%>\% 
  # rmv a nonessential var. to ensure new col is printed
  dplyr::select(-death_rate_7d_av) 
 
 # slide a left-aligned 7-day average
  jhu_csse_daily_subset \%>\%
  group_by(geo_value) \%>\%
  epi_slide(cases_7dav = mean(cases), n = 7, 
            align = "left") \%>\% 
  # rmv a nonessential var. to ensure new col is printed
  dplyr::select(-death_rate_7d_av) 
 
 # nested new columns
 jhu_csse_daily_subset \%>\% 
 group_by(geo_value) \%>\%
 epi_slide(a = data.frame(cases_2dav = mean(cases), 
                          cases_2dma = mad(cases)),
           n = 2, as_list_col = TRUE)
}
