% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_archive.R
\name{epix_as_of}
\alias{epix_as_of}
\title{Generate a snapshot from an \code{epi_archive} object}
\usage{
epix_as_of(x, max_version, min_time_value = -Inf)
}
\arguments{
\item{x}{An \code{epi_archive} object}

\item{max_version}{Time value specifying the max version to permit in the
snapshot. That is, the snapshot will comprise the unique rows of the
current archive data that represent the most up-to-date signal values, as
of the specified \code{max_version} (and whose time values are at least
\code{min_time_value}.)}

\item{min_time_value}{Time value specifying the min time value to permit in
the snapshot. Default is \code{-Inf}, which effectively means that there is no
minimum considered.}
}
\value{
An \code{epi_df} object.
}
\description{
Generates a snapshot in \code{epi_df} format from an \code{epi_archive} object, as of a
given version. See the \href{https://cmu-delphi.github.io/epiprocess/articles/archive.html}{archive vignette} for
examples.
}
\details{
This is simply a wrapper around the \code{as_of()} method of the
\code{epi_archive} class, so if \code{x} is an \code{epi_archive} object, then:\preformatted{epix_as_of(x, max_version = v)
}

is equivalent to:\preformatted{x$as_of(max_version = v)
}
}
\examples{
epix_as_of(archive_cases_dv, 
           max_version = max(archive_cases_dv$DT$version))
}
