% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_archive.R
\name{epix_as_of_latest}
\alias{epix_as_of_latest}
\title{Get the latest snapshot from an \code{epi_archive} object.}
\usage{
epix_as_of_latest(x)
}
\arguments{
\item{x}{An \code{epi_archive} object}
}
\value{
The latest snapshot from an \code{epi_archive} object
}
\description{
The latest snapshot is the snapshot of the last known version.
}
