% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_archive.R
\name{epix_fill_through_version}
\alias{epix_fill_through_version}
\title{Fill \code{epi_archive} unobserved history}
\usage{
epix_fill_through_version(x, fill_versions_end, how = c("na", "locf"))
}
\arguments{
\item{x}{An \code{epi_archive}}

\item{fill_versions_end}{Length-1, same class&type as \code{x$version}: the
version through which to fill in missing version history; this will be the
result's \verb{$versions_end} unless it already had a later
\verb{$versions_end}.}

\item{how}{Optional; \code{"na"} or \code{"locf"}: \code{"na"} will fill in any missing
required version history with \code{NA}s, by inserting (if necessary) an update
immediately after the current \verb{$versions_end} that revises all
existing measurements to be \code{NA} (this is only supported for \code{version}
classes with a \code{next_after} implementation); \code{"locf"} will fill in missing
version history with the last version of each observation carried forward
(LOCF), by leaving the update \verb{$DT} alone (other \code{epi_archive} methods are
based on LOCF).  Default is \code{"na"}.}
}
\value{
An \code{epi_archive}
}
\description{
Sometimes, due to upstream data pipeline issues, we have to work with a
version history that isn't completely up to date, but with functions that
expect archives that are completely up to date, or equally as up-to-date as
another archive. This function provides one way to approach such mismatches:
pretend that we've "observed" additional versions, filling in these versions
with NAs or extrapolated values.
}
