% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting-helpers.R
\name{epix_realtime_predictor_lag}
\alias{epix_realtime_predictor_lag}
\title{Get predictor lag train&test data, factoring in data maturity (revisioning)}
\usage{
epix_realtime_predictor_lag(
  archive,
  varname,
  time_lag,
  versions = epix_slide_versions_default(archive),
  predictor_name = "{.col}_{.dir}_{.amt}_realtime",
  drop_time_value = TRUE
)
}
\arguments{
\item{archive}{An \code{epi_archive}}

\item{varname}{String; name of the predictor column you'd like to
lag}

\item{time_lag}{Length-1 "time delta"; either an integer number of
time steps (with size determined by \code{archive$time_type}), or a
\code{difftime} object.}

\item{versions}{Optional vector of versions (a.k.a. forecast
dates/times) for which you'd like the predictor lags; defaults to
all versions with updates in \code{archive$DT} plus any versions that
look like they would be part of the normal reporting schedule but
didn't change any values.}

\item{predictor_name}{Optional string containing no curly braces,
or a \code{\link[glue:glue]{glue::glue_data}} pattern string; what/how should we name
the lagged version of the predictor in the output?}

\item{drop_time_value}{Optional Boolean; should we drop the
\code{time_value} column from the result?  The default, \code{TRUE}, is
makes merging results for multiple \code{time_lag}s cleaner (the join
key columns should be based on the geo, other keys, and version
(forecast date/time), not the time value).  You can use \code{FALSE}
to keep the \code{time_value} column instead, to check that your
\code{time_lag} is interpreted properly, or to make certain types of
plots easier.}
}
\value{
tibble with columns \code{"geo_value"}, any
\code{archive$other_keys}, \code{"time_value"} if requested,\code{"version"},
and \code{processed_predictor_name}.
}
\description{
Background: in many data sources, measurements for the most recent
time values are systematically noisy or even biased.  Bias is
particularly common when working with counts or population rates,
as the first versions of a measurement are often partial counts
that will be revised upward as data for more reporters or
individuals is received.
}
\details{
Motivation: simply using \code{\link{epix_as_of_latest}} and fitting a model
in these cases may produce erratic or biased forecasts.  We'd like
for the model to rely less on systematically noisy bleeding-edge
data, and to correct for systematically biased data.  So our
predictor training data shouldn't be the latest version, but
rather, versions that exhibit the same sorts of behavior.

For each \code{version} (a.k.a. forecast date/time), this extracts the
value of \code{varname} where \code{time_value == version - processed_time_lag} for each \code{geo_value} (or geo-age combinations,
etc. if you have \code{other_keys}), \emph{as it was reported in version
\code{version}}.  Here, \code{processed_time_lag} factors in the time step
size based on the archive's time type; see \code{time_lag} below.
}
