% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting-helpers-bu.R, R/nowcasting-helpers.R
\name{epix_target_evaluation_data}
\alias{epix_target_evaluation_data}
\title{Get target evaluation/fitting data, factoring in data maturity (revisioning)}
\usage{
epix_target_evaluation_data(
  archive,
  varname,
  relative_time,
  time_until_semistable = difftime_approx_ceiling_time_delta(as.difftime(60, units =
    "days"), archive$time_type),
  anchor_versions = epix_slide_versions_default(archive),
  out_name = "{.col}_{.amt}{.dir}_evaluation",
  nomatch = NA
)

epix_target_evaluation_data(
  archive,
  varname,
  relative_time,
  time_until_semistable = difftime_approx_ceiling_time_delta(as.difftime(60, units =
    "days"), archive$time_type),
  anchor_versions = epix_slide_versions_default(archive),
  out_name = "{.col}_{.amt}{.dir}_evaluation",
  nomatch = NA
)
}
\arguments{
\item{archive}{An \code{epi_archive}}

\item{varname}{String; name of target variable/column}

\item{relative_time}{Length-1 "time delta"; either an integer
number of time periods (with size determined by
\code{archive$time_type}), or a \code{difftime} object.  How do we move
from each "anchor version" (i.e., hypothetical forecast
date/etc.) to the desired \code{time_value}.  Typically this will be
negative; e.g., in daily data, to examine each report's
measurements for the previous day, it'd be -1.}

\item{time_until_semistable}{Length-1 time delta; replace the
target value with \code{NA} if it hasn't been at least
\code{time_until_semistable} since its \code{time_value} that its latest
available version (\code{archive$versions_end}) was recorded.}

\item{anchor_versions}{Optional vector of versions (a.k.a. forecast
dates/times) for which you'd like the target evaluation data;
defaults to all versions with updates in \code{archive$DT} plus any
versions that look like they would be part of the normal
reporting schedule but didn't change any values. Note that this
is to help line up with the appropriate predictor data from
\code{\link{epix_realtime_predictor_data}}; we'll still be using the latest
available version of each target values, regardless of
\code{anchor_versions}.}

\item{out_name}{Optional \code{\link[glue:glue]{glue::glue_data}} pattern string;
what/how should we name the target in the output?  It can just be
the output column name desired, provided it contains no curly
braces. Or you can use \code{glue} features and the following
shorthand: "\{.col\}" for \code{varname}, "\{.dir\}" for "lag" or
"lead", and "\{.amt\}" for a formatted absolute value of
\code{relative_time}.  Default is
"\{.col\}_\{.amt\}\{.dir\}_evaluation".}

\item{nomatch}{\code{NA} or \code{NULL}; what do we do when there is no
measurement available due to reporting latency or being at edges
or gaps of the time series?  \code{NA} means to include a row with an
\code{NA} as the lagged predictor value; \code{NULL} means to not output a
row in these cases.  (Forwarded to [\verb{[.data.table}].)}
}
\description{
Pairs with \code{\link{epix_realtime_predictor_data}}; see its documentation
for more details.

Pairs with \code{\link{epix_realtime_predictor_data}}; see its documentation
for more details.
}
\details{
Assumes that our goal is to predict the value of the target in its
final revision.  Takes a very simplistic approach to handling the
fact that we may not have that yet:
\enumerate{
\item For most \code{time_value}s, accept the latest available
reporting as "close enough" and use as-is.
\item For recent \code{time_value}s, deem them too unreliable to use in
fitting a model and replace the target values with \code{NA}s.
(Otherwise our model may learn to output overly wide prediction
intervals and/or biased predictions.)  The definition of
"recent" is controlled by \code{time_until_semistable}.
}

Other basic alternatives that could be paired with
\code{epix_realtime_predictor_data} include assigning weights to each
measurement based on its degree of reliability.  More complex
approaches would model each revision in the revision process as a
separate target.

Assumes that our goal is to predict the value of the target in its
final revision.  Takes a very simplistic approach to handling the
fact that we may not have that yet:
\enumerate{
\item For most \code{time_value}s, accept the latest available
reporting as "close enough" and use as-is.
\item For recent \code{time_value}s, deem them too unreliable to use in
fitting a model and replace the target values with \code{NA}s.
(Otherwise our model may learn to output overly wide prediction
intervals and/or biased predictions.)  The definition of
"recent" is controlled by \code{time_until_semistable}.
}

Other basic alternatives that could be paired with
\code{epix_realtime_predictor_data} include assigning weights to each
measurement based on its degree of reliability.  More complex
approaches would model each revision in the revision process as a
separate target.
}
