% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_archive.R
\name{epix_truncate_versions_after}
\alias{epix_truncate_versions_after}
\title{Filter an \code{epi_archive} object to keep only older versions}
\usage{
epix_truncate_versions_after(x, max_version)
}
\arguments{
\item{x}{An \code{epi_archive} object}

\item{max_version}{Time value specifying the max version to permit in the
filtered archive. That is, the output archive will comprise rows of the
current archive data having \code{version} less than or equal to the
specified \code{max_version}}
}
\value{
An \code{epi_archive} object
}
\description{
Generates a filtered \code{epi_archive} from an \code{epi_archive} object, keeping
only rows with \code{version} falling on or before a specified date.
}
