% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_patterns.R
\name{fill_missing_updates}
\alias{fill_missing_updates}
\title{Get pivot table, filling NAs. If there is no update on issue date \eqn{D} but
previous reports exist for issue date \eqn{D_p} < \eqn{D}, all the dates between
\eqn{[D_p, D]} are filled with with the reported value on date \eqn{D_p}. If there is
no update for any previous issue date, fill in with 0.}
\usage{
fill_missing_updates(df, value_col, refd_col, lag_col)
}
\arguments{
\item{df}{Data Frame of aggregated counts within a single location
reported for each reference date and issue date.}

\item{value_col}{column name for the column of counts}

\item{refd_col}{column name for the column of reference date}

\item{lag_col}{column name for the column of lag}
}
\description{
Get pivot table, filling NAs. If there is no update on issue date \eqn{D} but
previous reports exist for issue date \eqn{D_p} < \eqn{D}, all the dates between
\eqn{[D_p, D]} are filled with with the reported value on date \eqn{D_p}. If there is
no update for any previous issue date, fill in with 0.
}
