% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gcd2num}
\alias{gcd2num}
\title{Find the greatest common divisor of two numeric scalars}
\usage{
gcd2num(a, b, rrtol = 1e-06, pqlim = 1e+06, irtol = 1e-06)
}
\arguments{
\item{a}{Length 1, \code{is.numeric}; the first number}

\item{b}{Length 1, \code{is.numeric}; the second number}

\item{rrtol}{Optional, length 1, \code{is.numeric}, non-negative; the remainder
relative tolerance: consider a remainder from a division operation to be
zero if it is \code{abs(remainder/divisor) <= rrtol}. Could also be described as
a tolerance on the fractional part of the proper quotient. Default is 1e-6.}

\item{pqlim}{Optional, length 1, \code{is.numeric}, non-negative; the proper
quotient limit: consider a divisor to be zero if \code{abs(dividend/divisor) >= pqlim}.}

\item{irtol}{Optional, length 1, \code{is.numeric}, non-negative; the iterand
relative tolerance: consider \code{a} and \code{b} to have no gcd if the absolute
value of an iterand (and consequently also any result that might be
subsequently produced, as absolute values of iterands are decreasing) is
\verb{<= irtol * a} or \verb{<= irtol * b}. Also can be seen as the reciprocal of a
limit on the number \code{k} needed to achieve \code{k * gcd_result == max(abs(a),abs(b))}.}
}
\value{
Length 1, \code{is.numeric}: the gcd. (Or an error.) Expected to be a
double unless \code{b} is the GCD and an integer, in which case it is expected
be an integer.
}
\description{
Not expected to be used directly; output class isn't precise, and checks
could be moved away into \code{\link{gcd_num}}.
}
\details{
An implementation of a least absolute remainder Euclidean algorithm (See,
e.g., Moore, Thomas. "On the least absolute remainder Euclidean algorithm."
The Fibonacci Quarterly (1992).)

Notes on this implementation:
\itemize{
\item We allow positive or negative inputs, and don't require |a| > |b|.
\item \code{round} combines the job of truncating division and deciding between
positive and negative remainders.
\item We use some tolerance parameters and different checks to allow this to work
on floating-point numbers. Perhaps they could be altered or removed if we
are passed integers, but for simplicity, we always perform these checks.
}
}
\keyword{internal}
