% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tibble.R
\name{group_by.epi_tibble}
\alias{group_by.epi_tibble}
\alias{ungroup.epi_tibble}
\title{Group or ungroup \code{epi_tibble} object}
\usage{
\method{group_by}{epi_tibble}(x, ...)

\method{ungroup}{epi_tibble}(x, ...)
}
\description{
Groups or ungroups an \code{epi_tibble}, preserving class and attributes.
}
