% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_column_name}
\alias{guess_column_name}
\title{rename potential time_value columns}
\usage{
guess_column_name(x, column_name, substitutions)
}
\arguments{
\item{x}{the tibble to potentially rename}

\item{substitions}{a named vector. the potential substitions, with every name \code{time_value}}
}
\description{
potentially renames
}
\keyword{internal}
