% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{guess_period}
\alias{guess_period}
\title{Use max valid period as guess for \code{period} of \code{x}}
\usage{
guess_period(x, x_arg = rlang::caller_arg(x), fallback = NULL, ...)
}
\arguments{
\item{x}{Vector containing time-interval-like or time-point-like
data, with at least two distinct values.}

\item{x_arg}{Optional, string; name to give \code{x} in error
messages. Defaults to quoting the expression the caller fed into the
\code{x} argument.}

\item{fallback}{Optional; result to use if we don't have enough
information to reasonably guess.}

\item{...}{Should be empty, there to satisfy the S3 generic.}
}
\value{
length-1 vector; \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} class will
be the same class as \code{\link[base:diff]{base::diff()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
