% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_period}
\alias{guess_period}
\title{Use max valid period as guess for \code{period} of \code{ref_time_values}}
\usage{
guess_period(
  ref_time_values,
  ref_time_values_arg = rlang::caller_arg(ref_time_values)
)
}
\arguments{
\item{`ref_time_values`}{Vector containing time-interval-like or time-like
data, with at least two distinct values, \code{\link{diff}}-able (e.g., a
\code{time_value} or \code{version} column), and should have a sensible result from
adding \code{is.numeric} versions of its \code{diff} result (via \code{as.integer} if its
\code{typeof} is \code{"integer"}, otherwise via \code{as.numeric}).}

\item{`ref_time_values_arg`}{Optional, string; name to give \code{ref_time_values}
in error messages. Defaults to quoting the expression the caller fed into
the \code{ref_time_values} argument.}
}
\value{
\code{is.numeric}, length 1; attempts to match \code{typeof(ref_time_values)}
}
\description{
Use max valid period as guess for \code{period} of \code{ref_time_values}
}
