% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_df.R
\name{is_epi_df}
\alias{is_epi_df}
\title{Test if the object is an epi_df}
\usage{
is_epi_df(x)
}
\arguments{
\item{x}{An object.}
}
\value{
TRUE if the object inherits from the \code{epi_df} class.
}
\description{
Test if the object is an epi_df
}
