% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{is_locf}
\alias{is_locf}
\title{Checks to see if a value in a vector is LOCF}
\usage{
is_locf(vec)
}
\description{
LOCF meaning last observation carried forward. lags the vector by 1, then
compares with itself. For floats it uses float comparison, otherwise it
uses equality.
\code{NA}'s are considered equal to \code{NA}'s, while nan's are not.
}
\keyword{internal}
