% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{data}
\name{jhu_confirmed_cumulative_num}
\alias{jhu_confirmed_cumulative_num}
\title{COVID Cumulative Cases Example}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2808 rows and 15 columns.
}
\source{
This object contains a modified part of the \href{https://github.com/CSSEGISandData/COVID-19}{COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University} as \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{republished in the COVIDcast Epidata API}. This data set is licensed under the terms of the
\href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0 International license}
by the Johns Hopkins University on behalf of its Center for Systems Science in Engineering.
Copyright Johns Hopkins University 2020.
}
\usage{
jhu_confirmed_cumulative_num
}
\description{
Data set for 4 states containing COVID-19 Cumulative Cases as reported
by JHU-CSSE and downloaded from the CMU Delphi COVIDcast Epidata API.
This data set covers the period from March 2020 to January 2022, and is
used in the \link{epiprocess} Vignette on ... .
}
\examples{
 # Since this is a re-exported dataset, it cannot be loaded using
 # the `data()` function. `data()` looks for a file of the same name
 # in the `data/` directory, which doesn't exist in this package.
 # works
 epiprocess::jhu_confirmed_cumulative_num

 # works
 library(epiprocess)
 jhu_confirmed_cumulative_num

 # fails
 data(jhu_confirmed_cumulative_num, package = "epiprocess")
}
\keyword{datasets}
