% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jhu_csse_daily}
\alias{jhu_csse_daily}
\title{JHU daily cases and deaths from California and Florida}
\format{
A tibble with 30 rows and 6 variables:
\describe{
\item{geo_value}{the geographic value associated with each row of measurements.}
\item{time_value}{the time value associated with each row of measurements.}
\item{case_rate_7d_av}{7-day average signal of number of new confirmed COVID-19 cases per 100,000 population, daily}
\item{death_rate_7d_av}{7-day average signal of number of new confirmed deaths due to COVID-19 per 100,000 population, daily}
\item{cases}{Number of new confirmed COVID-19 cases, daily}
\item{cases_7d_av}{7-day average signal of number of new confirmed COVID-19 cases, daily}
}
}
\source{
COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University.
\url{https://github.com/CSSEGISandData/COVID-19}
}
\usage{
jhu_csse_daily
}
\description{
This data source of confirmed COVID-19 cases and deaths
is based on reports made available by the Center for
Systems Science and Engineering at Johns Hopkins University.
This example data ranges from June 1 to June 15, 2020.
}
\keyword{datasets}
