% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_colnames.R
\name{key_colnames}
\alias{key_colnames}
\alias{key_colnames.data.frame}
\alias{key_colnames.epi_df}
\alias{key_colnames.epi_archive}
\title{Get names of columns that form a (unique) key associated with an object}
\usage{
key_colnames(x, ..., exclude = character(0L))

\method{key_colnames}{data.frame}(x, other_keys, ..., exclude = character(0L))

\method{key_colnames}{epi_df}(x, ..., exclude = character(0L))

\method{key_colnames}{epi_archive}(x, ..., exclude = character(0L))
}
\arguments{
\item{x}{an object, such as an \code{\link{epi_df}}}

\item{...}{additional arguments passed on to methods}

\item{exclude}{an optional character vector of key column names to exclude
from the result}

\item{other_keys}{character vector; what besides \code{geo_value} and \code{time_value}
(if present) should we consider to be key columns?  Used, e.g., if we}
}
\value{
character vector
}
\description{
This is entirely based on metadata and arguments passed; there are no
explicit checks that the key actually is unique in any associated data
structures.
}
\keyword{internal}
