% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelist.R
\name{linelist_to_archive}
\alias{linelist_to_archive}
\title{Convert a line list to an \code{epi_archive} object}
\usage{
linelist_to_archive(
  x,
  ...,
  geo_value = NULL,
  time_value = NULL,
  version_recorded = NULL,
  version_deleted = NULL,
  is_deleted = NULL,
  other_keys = NULL,
  value = NULL,
  id = NULL,
  clobberable_versions_start = NA,
  versions_end = NULL
)
}
\arguments{
\item{x}{A data frame (line list).}

\item{...}{Should be empty.}

\item{geo_value, time_value, version_recorded, version_deleted, other_keys}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns in \code{x} representing:
\itemize{
\item \code{geo_value}: the geographic location of the event.
\item \code{time_value}: the time of the event.
\item \code{version_recorded}: the time at which the event became known/recorded.
\item \code{version_deleted}: (optional) the time at which the event was
removed/deleted. If \code{NULL} (default), it is assumed no events are
deleted.
\item \code{other_keys}: (optional) additional key columns (e.g. age group).
}}

\item{is_deleted}{(optional) <\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Column in \code{x} (or a logical vector/predicate) indicating if the row is a
deletion (\code{TRUE}/1) or an entry (\code{FALSE}/0). Only used for "chart-style"
linelists (where \code{version_recorded} and \code{version_deleted} are the same
column).}

\item{value}{Either \code{NULL} (default) or a string specifying the name of the
output count column. If \code{NULL} and \code{other_keys} is empty, defaults to
"count".}

\item{id}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optional column
identifying unique events/cases.}

\item{clobberable_versions_start}{Optional; \code{length}-1; either a value of the
same \code{class} as \code{x$version}, or an \code{NA} of any \code{class}: specifically,
either (a) the earliest version that could be subject to "clobbering"
(being overwritten with different update data, but using the \emph{same} version
tag as the old update data), or (b) \code{NA}, to indicate that no versions are
clobberable. There are a variety of reasons why versions could be
clobberable under routine circumstances, such as (a) today's version of
one/all of the columns being published after initially being filled with
\code{NA} or LOCF, (b) a buggy version of today's data being published but then
fixed and republished later in the day, or (c) data pipeline delays (e.g.,
publisher uploading, periodic scraping, database syncing, periodic
fetching, etc.) that make events (a) or (b) reflected later in the day (or
even on a different day) than expected; potential causes vary between
different data pipelines. The default value is \code{NA}, which doesn't consider
any versions to be clobberable. Another setting that may be appropriate for
some pipelines is \code{max_version_with_row_in(x)}.}

\item{versions_end}{optional; as in \code{\link{as_epi_archive}}, or \code{NULL}.
If the latest version(s) had no new events recorded or deleted,
you can note this using \code{versions_end}.  \code{NULL}, the default,
will assume there are no such versions.}
}
\value{
An \code{\link{epi_archive}} object.
}
\description{
Converts a line list (a data frame where each row represents a case or event)
into an \code{\link{epi_archive}} object. This function requires "recorded" and
"deleted" timestamps, and generates a time series of counts (e.g. daily
hospitalizations) as they would have appeared at different points in time.
}
\examples{
library(dplyr)

linelist <- tibble(
  event_id = 1:3,
  geo_value = c("ca", "ca", "ca"),
  time_value = as.Date(c("2022-01-01", "2022-01-01", "2022-01-02")),
  report_date = as.Date(c("2022-01-02", "2022-01-02", "2022-01-03")),
  delete_date = as.Date(c("2022-01-04", NA, NA))
)

archive <- linelist_to_archive(
  linelist,
  geo_value = geo_value,
  time_value = time_value,
  version_recorded = report_date,
  version_deleted = delete_date
)

}
