% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelist.R
\name{linelist_to_archive}
\alias{linelist_to_archive}
\title{Convert a line list to an \code{epi_archive} object}
\usage{
linelist_to_archive(
  x,
  ...,
  geo_value = NULL,
  time_value = NULL,
  version_recorded = NULL,
  version_deleted = NULL,
  is_deleted = NULL,
  other_keys = NULL,
  value = NULL,
  id = NULL
)
}
\arguments{
\item{x}{A data frame (line list).}

\item{...}{Arguments passed to \code{\link{new_epi_archive}}.}

\item{geo_value, time_value, version_recorded, version_deleted, other_keys}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns in \code{x} representing:
\itemize{
\item \code{geo_value}: the geographic location of the event.
\item \code{time_value}: the time of the event.
\item \code{version_recorded}: the time at which the event became known/recorded.
\item \code{version_deleted}: (optional) the time at which the event was
removed/deleted. If \code{NULL} (default), it is assumed no events are
deleted.
\item \code{other_keys}: (optional) additional key columns (e.g. age group).
}}

\item{is_deleted}{(optional) <\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
Column in \code{x} (or a logical vector/predicate) indicating if the row is a
deletion (\code{TRUE}/1) or an entry (\code{FALSE}/0). Only used for "chart-style"
linelists (where \code{version_recorded} and \code{version_deleted} are the same
column).}

\item{value}{Either \code{NULL} (default) or a string specifying the name of the
output count column. If \code{NULL} and \code{other_keys} is empty, defaults to
"count".}

\item{id}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optional column
identifying unique events/cases.}
}
\value{
An \code{\link{epi_archive}} object.
}
\description{
Converts a line list (a data frame where each row represents a case or event)
into an \code{\link{epi_archive}} object. This function requires "recorded" and
"deleted" timestamps, and generates a time series of counts (e.g. daily
hospitalizations) as they would have appeared at different points in time.
}
\examples{
library(dplyr)

linelist <- tibble(
  event_id = 1:3,
  geo_value = c("ca", "ca", "ca"),
  time_value = as.Date(c("2022-01-01", "2022-01-01", "2022-01-02")),
  report_date = as.Date(c("2022-01-02", "2022-01-02", "2022-01-03")),
  delete_date = as.Date(c("2022-01-04", NA, NA))
)

archive <- linelist_to_archive(
  linelist,
  geo_value = geo_value,
  time_value = time_value,
  version_recorded = report_date,
  version_deleted = delete_date
)

}
