% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{n_steps_to_time_delta}
\alias{n_steps_to_time_delta}
\title{Convert from integerish/infinite/mix to time_delta}
\usage{
n_steps_to_time_delta(n_steps, time_type, format = c("friendly", "fast"))
}
\description{
Convert from integerish/infinite/mix to time_delta
}
