% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{pct_change}
\alias{pct_change}
\title{Compute percentage change of a variable in an \code{epi_tibble} object}
\usage{
pct_change(
  x,
  var,
  n = 14,
  align = c("trailing", "centered"),
  new_col_name = "pct_change",
  time_step,
  na_rm = TRUE
)
}
\arguments{
\item{x}{The \code{epi_tibble} object.}

\item{var}{The variable in \code{x} whose percentage change values are to be
computed.}

\item{n}{Number of time steps to use in the running window. For example, if
\code{n = 10}, one time step is one day, and the alignment is "trailing", then
the percentage change value computed on November 10 is 100 * (B - A) / A,
where A is the sum of the variable in between November 6 and 10, and A is
the sum in between November 1 and 5. Note that \code{n} must be even; default is
14.}

\item{align}{String specifying the alignment of the sliding window relative
to the reference time point; either "trailing" or "centered". The default
is "trailing". If the alignment is "centered" and \code{n} is even, then one
more observation will be used before the reference time than after the
reference time.}

\item{new_col_name}{String indicating the name of the new column that will
contain the derivative values. Default is "slide_value"; note that setting
\code{new_col_name} equal to an existing column name will overwrite this column.}

\item{time_step}{Optional function used to define the meaning of one time
step, which if specified, overrides the default choice based on the
metadata. Read the documentation for \code{\link[=epi_slide]{epi_slide()}} for more details.}

\item{na_rm}{Should missing values be removed before the computation? Default
is \code{TRUE}.}
}
\value{
An \code{epi_tibble} object given by appending a new column to \code{x}, named
according to the \code{new_col_name} argument, containing the percentage change
values.
}
\description{
Computes the percentage change of a variable in an \code{epi_tibble} object. See
the \href{https://cmu-delphi.github.io/epitools/articles/pct-change.html}{percentage change vignette} for
examples.
}
\seealso{
\code{\link[=epi_slide]{epi_slide()}}
}
