% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_change.R
\name{pct_change}
\alias{pct_change}
\title{Compute percentage change}
\usage{
pct_change(var, na_rm = FALSE)
}
\arguments{
\item{var}{The variable whose percentage change value is to be computed. If
the length of \code{var} is odd, then the first value is ignored for the sake of
the computation.}

\item{na_rm}{Should missing values be removed before the computation? Default
is \code{FALSE}.}
}
\description{
Computes the percentage change over a sequence of values, defined as 100 * (B
\itemize{
\item A) / A, where A is the sum of the first half of the values and B is the sum
of the second half. This is a convenience function to be used inside of a
call to \code{epi_slide()}, see the \href{https://cmu-delphi.github.io/epiprocess/articles/pct_change.html}{percentage change vignette}
for examples.
}
}
