% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap for an epi_df}
\usage{
plot_heatmap(x, ..., .max_keys = 60)
}
\arguments{
\item{x}{An \code{epi_df} object.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}

\item{.max_keys}{The maximum number of key combinations to display.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot a heatmap for an epi_df
}
\examples{
# Use it on an `epi_df`
plot_heatmap(cases_deaths_subset, case_rate_7d_av)

# Plotting multiple variables
plot_heatmap(cases_deaths_subset, case_rate_7d_av, death_rate_7d_av)
}
