% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_df.R
\name{print.epi_df}
\alias{print.epi_df}
\alias{summary.epi_df}
\alias{group_by.epi_df}
\alias{rowwise.epi_df}
\alias{ungroup.epi_df}
\alias{group_modify.epi_df}
\alias{unnest.epi_df}
\title{Base S3 methods for an \code{epi_df} object}
\usage{
\method{print}{epi_df}(x, ...)

\method{summary}{epi_df}(object, ...)

\method{group_by}{epi_df}(.data, ...)

\method{rowwise}{epi_df}(.data, ...)

\method{ungroup}{epi_df}(x, ...)

\method{group_modify}{epi_df}(.data, .f, ..., .keep = FALSE)

\method{unnest}{epi_df}(data, ...)
}
\arguments{
\item{x}{an \code{epi_df}}

\item{...}{Additional arguments, for compatibility with \code{summary()}.
Currently unused.}

\item{object}{an \code{epi_df}}

\item{.data}{an \code{epi_df}}

\item{.f}{function or formula; see \code{\link[dplyr:group_map]{dplyr::group_modify}}}

\item{.keep}{Boolean; see \code{\link[dplyr:group_map]{dplyr::group_modify}}}

\item{data}{an \code{epi_df}}
}
\description{
Print and summary functions for an \code{epi_df} object.

Prints a variety of summary statistics about the \code{epi_df} object, such as
the time range included and geographic coverage.
}
