% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.epi_df}
\alias{print.epi_df}
\alias{summary.epi_df}
\alias{arrange.epi_df}
\alias{filter.epi_df}
\alias{group_by.epi_df}
\alias{group_modify.epi_df}
\alias{mutate.epi_df}
\alias{relocate.epi_df}
\alias{rename.epi_df}
\alias{slice.epi_df}
\alias{ungroup.epi_df}
\alias{unnest.epi_df}
\title{S3 methods for an \code{epi_df} object}
\usage{
\method{print}{epi_df}(x, ...)

\method{summary}{epi_df}(x, ...)

\method{arrange}{epi_df}(x, ...)

\method{filter}{epi_df}(x, ...)

\method{group_by}{epi_df}(x, ...)

\method{group_modify}{epi_df}(x, ...)

\method{mutate}{epi_df}(x, ...)

\method{relocate}{epi_df}(x, ...)

\method{rename}{epi_df}(x, ...)

\method{slice}{epi_df}(x, ...)

\method{ungroup}{epi_df}(x, ...)

\method{unnest}{epi_df}(x, ...)
}
\arguments{
\item{x}{The \code{epi_df} object.}

\item{...}{Additional arguments, for compatibility with \code{summary()}.
Currently unused.}

\item{object}{The \code{epi_df} object.}
}
\description{
Print, summary, and \code{dplyr} verbs (that preserve class and attributes) for an
\code{epi_df} object.

Prints a variety of summary statistics about the \code{epi_df} object, such as
the time range included and geographic coverage.

\code{dplyr} verbs for \code{epi_df} objexts, preserving class and attributes.
}
