% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_df.R
\name{as_tibble.epi_df}
\alias{as_tibble.epi_df}
\alias{print.epi_df}
\alias{summary.epi_df}
\alias{group_by.epi_df}
\alias{ungroup.epi_df}
\alias{group_modify.epi_df}
\alias{unnest.epi_df}
\title{Convert to tibble, dropping metadata}
\usage{
\method{as_tibble}{epi_df}(x, ...)

\method{print}{epi_df}(x, ...)

\method{summary}{epi_df}(object, ...)

\method{group_by}{epi_df}(.data, ...)

\method{ungroup}{epi_df}(x, ...)

\method{group_modify}{epi_df}(.data, .f, ..., .keep = FALSE)

\method{unnest}{epi_df}(data, ...)
}
\arguments{
\item{x}{The \code{epi_df} object.}

\item{...}{Additional arguments, for compatibility with \code{summary()}.
Currently unused.}

\item{object}{The \code{epi_df} object.}

\item{data}{The \code{epi_df} object.}
}
\description{
Print and summary functions for an \code{epi_df} object.

Prints a variety of summary statistics about the \code{epi_df} object, such as
the time range included and geographic coverage.
}
