% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_df.R
\name{print.epi_df}
\alias{print.epi_df}
\title{Print \code{epi_df} object}
\usage{
\method{print}{epi_df}(x, ...)
}
\arguments{
\item{x}{The \code{epi_df} object.}

\item{...}{Additional arguments passed to \code{print.tibble()} to print the
data.}
}
\value{
The \code{epi_df} object, unchanged.
}
\description{
Prints a brief summary of the \code{epi_df} object, then prints the underlying
tibble.
}
