% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tsibble}
\alias{arrange}
\alias{filter}
\alias{group_by}
\alias{ungroup}
\alias{group_modify}
\alias{mutate}
\alias{relocate}
\alias{rename}
\alias{slice}
\alias{left_join}
\alias{right_join}
\alias{inner_join}
\alias{full_join}
\alias{anti_join}
\alias{semi_join}
\alias{cross_join}
\alias{unnest}
\alias{complete}
\alias{drop_na}
\alias{pivot_wider}
\alias{pivot_longer}
\alias{full_seq}
\alias{autoplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:filter-joins]{anti_join}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{cross_join}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr:mutate-joins]{full_join}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr:group_map]{group_modify}}, \code{\link[dplyr:mutate-joins]{inner_join}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{relocate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:mutate-joins]{right_join}}, \code{\link[dplyr:filter-joins]{semi_join}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr:group_by]{ungroup}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{tidyr}{\code{\link[tidyr]{complete}}, \code{\link[tidyr]{drop_na}}, \code{\link[tidyr]{full_seq}}, \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{pivot_wider}}, \code{\link[tidyr]{unnest}}}

  \item{tsibble}{\code{\link[tsibble:as-tsibble]{as_tsibble}}}
}}

