% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tsibble}
\alias{arrange}
\alias{filter}
\alias{group_by}
\alias{ungroup}
\alias{group_modify}
\alias{mutate}
\alias{relocate}
\alias{rename}
\alias{slice}
\alias{left_join}
\alias{right_join}
\alias{inner_join}
\alias{full_join}
\alias{anti_join}
\alias{semi_join}
\alias{cross_join}
\alias{unnest}
\alias{complete}
\alias{drop_na}
\alias{pivot_wider}
\alias{pivot_longer}
\alias{full_seq}
\alias{autoplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:anti_join]{anti_join()}}, \code{\link[dplyr:arrange]{arrange()}}, \code{\link[dplyr:cross_join]{cross_join()}}, \code{\link[dplyr:filter]{filter()}}, \code{\link[dplyr:full_join]{full_join()}}, \code{\link[dplyr:group_by]{group_by()}}, \code{\link[dplyr:group_modify]{group_modify()}}, \code{\link[dplyr:inner_join]{inner_join()}}, \code{\link[dplyr:left_join]{left_join()}}, \code{\link[dplyr:mutate]{mutate()}}, \code{\link[dplyr:relocate]{relocate()}}, \code{\link[dplyr:rename]{rename()}}, \code{\link[dplyr:right_join]{right_join()}}, \code{\link[dplyr:semi_join]{semi_join()}}, \code{\link[dplyr:slice]{slice()}}, \code{\link[dplyr:ungroup]{ungroup()}}}

  \item{ggplot2}{\code{\link[ggplot2:autoplot]{autoplot()}}}

  \item{tidyr}{\code{\link[tidyr:complete]{complete()}}, \code{\link[tidyr:drop_na]{drop_na()}}, \code{\link[tidyr:full_seq]{full_seq()}}, \code{\link[tidyr:pivot_longer]{pivot_longer()}}, \code{\link[tidyr:pivot_wider]{pivot_wider()}}, \code{\link[tidyr:unnest]{unnest()}}}

  \item{tsibble}{\code{\link[tsibble:as_tsibble]{as_tsibble()}}}
}}

