% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{removed_by_compactify}
\alias{removed_by_compactify}
\title{get the entries that \code{compactify} would remove}
\usage{
removed_by_compactify(history_df, ukey_names, abs_tol)
}
\description{
get the entries that \code{compactify} would remove
}
\keyword{internal}
