% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{session_tz}
\alias{session_tz}
\title{An attempt at a more "accurate"/relevant \code{\link[=Sys.timezone]{Sys.timezone()}}}
\usage{
session_tz()
}
\description{
The documentation for \code{\link[=Sys.timezone]{Sys.timezone()}} states that it tries to
detect the "initial" value of the "TZ" environment variable, but
\code{print} seems to use the current value, so the latter seems more
relevant.  It also mentions invalid time zones maybe being treated
as UTC, and maybe being warned about, and \code{print} on one platform
seems to follow as-UTC, no warning, plus sometimes displaying and
sometimes not displaying the invalid time zone; we don't want to
encounter these situations, so try to abort before they happen.
Also, adds a gate for session timezones of "UTC" used by some CI
and cloud computing services, requiring users to confirm that "UTC"
is actually relevant to them.
}
\keyword{internal}
