% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epi_df.R
\name{sum_groups_epi_df}
\alias{sum_groups_epi_df}
\title{Aggregate an \code{epi_df} object}
\usage{
sum_groups_epi_df(.x, sum_cols = "value", group_cols = character())
}
\arguments{
\item{.x}{an \code{epi_df}}

\item{sum_cols}{character vector of the columns to aggregate}

\item{group_cols}{character vector of column names to group by. "time_value" is
included by default.}
}
\value{
an \code{epi_df} object
}
\description{
Aggregates an \code{epi_df} object by the specified group columns, summing the
\code{value} column, and returning an \code{epi_df}. If aggregating over \code{geo_value},
the resulting \code{epi_df} will have \code{geo_value} set to \code{"total"}.
}
