% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblish.R
\name{tblish_cast_cols}
\alias{tblish_cast_cols}
\title{Cast (via \code{\link{vec_cast_patched}}) (a subset of) columns of \code{x} to required ptypes}
\usage{
tblish_cast_cols(x, cols_to)
}
\arguments{
\item{x}{\code{tbl}-ish object; see \code{\link{obj_is_tblish}}}

\item{cols_to}{named list of \code{\link[vctrs:vec_ptype]{vctrs::vec_ptype}}s; can be list subclass,
such as \code{vec_ptype} of a tibble}
}
\value{
\code{x}, with columns
}
\description{
Cast (via \code{\link{vec_cast_patched}}) (a subset of) columns of \code{x} to required ptypes
}
\keyword{internal}
