% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_sensible_int}
\alias{test_sensible_int}
\title{Is \code{x} an "int" with a sensible class? TRUE/FALSE}
\usage{
test_sensible_int(
  x,
  na.ok = FALSE,
  lower = -Inf,
  upper = Inf,
  tol = sqrt(.Machine$double.eps),
  null.ok = FALSE
)
}
\arguments{
\item{x}{object}
}
\value{
Boolean
}
\description{
Like \code{\link[checkmate:checkInt]{checkmate::test_int}} but disallowing some non-sensible classes that
\code{test_int} accepts, such as \code{difftime}s. We rely on \code{\link{is.numeric}} to
determine class appropriateness; note that \code{is.numeric} is NOT simply
checking for the class to be "numeric" (or else we'd fail on integer class).
}
\keyword{internal}
