% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline-roxygen.R
\docType{data}
\name{tidyselect_arg_roxygen}
\alias{tidyselect_arg_roxygen}
\title{Description of a single arg that tidyselects value variables}
\format{
An object of class \code{character} of length 1.
}
\usage{
tidyselect_arg_roxygen
}
\description{
Not meant for when describing tidyselect \code{...}.
}
\keyword{internal}
