% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_column_names}
\alias{time_column_names}
\title{potential time_value columns}
\usage{
time_column_names()
}
\description{
the full list of potential substitutions for the \code{time_value} column name:
time_value, date, time, datetime, dateTime, date_time, target_date, week, epiweek, month, mon, year, yearmon, yearmonth, yearMon, yearMonth, dates, time_values, target_dates, time_Value, Time_Value, Date, Time, Datetime, DateTime, Date_Time, Target_Date, Week, Epiweek, Month, Mon, Year, Yearmon, Yearmonth, YearMon, YearMonth, Dates, Time_Values, Target_Dates, Time_Value
}
