% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{time_type_unit_abbrs}
\alias{time_type_unit_abbrs}
\title{Helper data for \code{\link{time_type_unit_abbr}}}
\usage{
time_type_unit_abbrs
}
\description{
Helper data for \code{\link{time_type_unit_abbr}}
}
\keyword{internal}
