% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\docType{data}
\name{time_type_unit_pluralizer}
\alias{time_type_unit_pluralizer}
\title{Helper data for \code{\link{format_time_delta}}}
\format{
An object of class \code{character} of length 4.
}
\usage{
time_type_unit_pluralizer
}
\description{
Should not be altered on the basis of untrusted user input, as it is used as
a cli format string and may run code.
}
\keyword{internal}
