% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{to_integerish_time_delta}
\alias{to_integerish_time_delta}
\title{Adjust time_type to a dayl if there are non-integer differences}
\usage{
to_integerish_time_delta(time_delta)
}
\description{
Adjust time_type to a dayl if there are non-integer differences
}
\keyword{internal}
