% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{vec_position_lag}
\alias{vec_position_lag}
\title{Lag entries in a vctrs-style vector by their position in the vector}
\usage{
vec_position_lag(x, n)
}
\description{
This is like \code{\link[dplyr:lead-lag]{dplyr::lag}}, though it has different performance
characteristics. It's unclear whether we want to use this.
}
\examples{

vec_position_lag(1:5, 3)
dplyr::lag(1:5, 3)

# Try on a more exotic class:
vec_position_lag(epipredict::dist_quantiles(list(1:6, 11:16), 0:5 / 5), 1)
dplyr::lag(epipredict::dist_quantiles(list(1:6, 11:16), 0:5 / 5), 1)

}
\keyword{internal}
