% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{version_column_names}
\alias{version_column_names}
\title{potential version columns}
\usage{
version_column_names()
}
\description{
the full list of potential substitutions for the \code{version} column name:
version, issue, issues, release, version_recorded, report_date, recorded_date, issue_date, release_date, as_of, revision, report_time, update_date, update_time, publish_date, publish_time, published_date, published_time, upload_date, upload_time, as_of_date, as_of_time, asof, Version, Issue, Issues, Release, Version_Recorded, Report_Date, Recorded_Date, Issue_Date, Release_Date, As_Of, Revision, Report_Time, Update_Date, Update_Time, Publish_Date, Publish_Time, Published_Date, Published_Time, Upload_Date, Upload_Time, As_Of_Date, As_Of_Time, Asof
}
\keyword{internal}
