% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{version_column_names}
\alias{version_column_names}
\title{potential version columns}
\usage{
version_column_names()
}
\description{
the full list of potential substitutions for the \code{version} column name:
version, issue, release, Version, Issue, Release
}
