% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-utils.R
\name{versions_standardize}
\alias{versions_standardize}
\title{Standardize a \code{versions}/\code{.versions} argument into a vector of versions}
\usage{
versions_standardize(
  versions,
  archive,
  versions_arg = caller_arg(versions),
  call = caller_env()
)
}
\arguments{
\item{versions}{Either
(a) a vector containing the set of desired versions to
include,
(b) a description of the desired spacing, such as \code{"week"},
\code{"2 weeks"}, \code{"month"}, or another string accepted by
\code{\link{seq}}'s / \code{\link{seq.Date}}'s \code{by} parameter, or
(c) \code{NULL},
to include all versions in the archive.

In case (a), we accept vectors that can be automatically converted
to match the \link[vctrs:vec_ptype]{ptype} of versions in the archive;
we try both character-to-Date and \code{\link[vctrs:vec_cast]{vctrs::vec_cast}} conversions.}

\item{archive}{the \code{epi_archive} to select versions from.}
}
\value{
a vector with the same ptype as \code{archive$DT$version}
}
\description{
Standardize a \code{versions}/\code{.versions} argument into a vector of versions
}
\keyword{internal}
